Compute velocity for x, y, and its amplited

Compute velocity for x, y, and its amplited

compute_velocity_table(x, y, trial, sample_rate, velocity_time_window)

Arguments

x

vector with x coordinates in degrees of visual angle

y

vector with y coordinates in degrees of visual angle

trial

vector with trial index

sample_rate

sample rate in Hz

velocity_time_window

Time window for velocity computation in milliseconds

Value

data.frame with columns x, y, and amp

Examples

compute_velocity_table(rnorm(1000), rnorm(1000), rep(1, 1000), 250, 20)
#>                x           y        amp
#> 1             NA          NA         NA
#> 2             NA          NA         NA
#> 3             NA          NA         NA
#> 4     27.8558612 -142.298531 144.999382
#> 5     42.7667368  -86.163735  96.193466
#> 6     33.1835479  -31.065157  45.455383
#> 7     34.1546186   25.958213  42.899496
#> 8     87.2692251  -19.611179  89.445604
#> 9     43.2328181   18.185317  46.901837
#> 10   -29.8508079   79.645872  85.056074
#> 11   -85.4152149   80.681510 117.495808
#> 12   -65.0712095   37.575862  75.141252
#> 13    -5.7210279   14.485977  15.574777
#> 14   109.4637383   24.095268 112.084307
#> 15   121.8072968   10.589404 122.266729
#> 16   -39.3451659   39.652262  55.860039
#> 17  -110.3397243  127.502047 168.616804
#> 18   -79.3246348   84.626922 115.991869
#> 19    77.2803521   33.333828  84.162919
#> 20   132.5838195  -39.587056 138.367641
#> 21    66.1741459 -100.258434 120.128145
#> 22    33.6223325  -91.335126  97.327111
#> 23   -90.1130674  -72.665961 115.761422
#> 24  -148.1465303   78.577548 167.695633
#> 25   -65.0048253  102.275450 121.185374
#> 26   -28.2006418   17.296366  33.082329
#> 27    39.9702560   64.792377  76.129321
#> 28     4.2344119  129.710364 129.779463
#> 29    42.9322462   88.838148  98.668102
#> 30    94.8498768  -25.347401  98.178357
#> 31    71.9475714 -180.088654 193.928791
#> 32    92.4416989 -127.853015 157.771548
#> 33   -78.4366080  -96.783694 124.576823
#> 34  -102.6167134  -26.654810 106.022020
#> 35    -0.5497340   46.801441  46.804669
#> 36    97.6386319   14.931567  98.773752
#> 37    54.4280336   35.290562  64.867824
#> 38  -111.8632484   42.003958 119.489409
#> 39   -82.3713143   70.489239 108.414788
#> 40   -63.0788401   67.003720  92.024120
#> 41   -28.8626881   15.675803  32.844871
#> 42    80.1199782    8.794334  80.601186
#> 43   108.1672359   54.158134 120.967989
#> 44    40.1369793    4.243548  40.360684
#> 45   -99.0089291  -72.151077 122.509371
#> 46  -145.5331276 -145.569149 205.840395
#> 47   -76.9779689  -21.274350  79.863669
#> 48    25.8455451   40.407169  47.965941
#> 49   125.0282615   56.540644 137.218478
#> 50    76.0771722   81.012667 111.134101
#> 51    15.4912867    7.030346  17.011929
#> 52   -21.0183505  -81.626897  84.289509
#> 53  -109.2781481  -83.441920 137.492791
#> 54  -101.0860043  -12.488769 101.854551
#> 55   -20.4150874    8.384117  22.069645
#> 56    79.7540510  -67.462724 104.460173
#> 57     6.7241882  -85.672852  85.936327
#> 58   -80.6629217   79.045155 112.936458
#> 59     4.3639600   57.516905  57.682220
#> 60    96.0976078  -31.400806 101.097779
#> 61   113.8976013  -50.607634 124.634651
#> 62    17.9908820   -2.906774  18.224192
#> 63   -17.1147188  135.366373 136.444012
#> 64   -28.0422315  106.972460 110.586952
#> 65   -44.9869338  -18.588171  48.675911
#> 66  -107.2196477  -78.162152 132.685247
#> 67   -74.5146190 -145.393771 163.376183
#> 68   147.2138250  -71.716198 163.753239
#> 69   228.6165623   72.280224 239.770647
#> 70    58.1813470  107.569604 122.295907
#> 71  -151.1826787   12.319529 151.683793
#> 72  -180.2802079  -90.914746 201.907019
#> 73   -40.4324046  -33.265772  52.358294
#> 74     3.1695507  112.716678 112.761233
#> 75    17.2365210  112.988020 114.295189
#> 76   114.9887404  -45.747303 123.754702
#> 77   113.7630170 -136.828792 177.944211
#> 78    49.1482499 -111.053764 121.443357
#> 79    -6.1757703   52.774249  53.134372
#> 80  -100.4368660   93.103298 136.951773
#> 81  -153.9930437   25.848213 156.147327
#> 82   -74.6229621   -8.611104  75.118158
#> 83    69.0517331  -68.010180  96.920206
#> 84   151.9875828  -81.261774 172.347617
#> 85    74.2502554  -94.904164 120.498551
#> 86   -47.1370094  -16.839549  50.054651
#> 87   -46.3701631  157.351891 164.042097
#> 88   -80.7679675  171.839841 189.874684
#> 89   -79.8132429   47.185610  92.718043
#> 90     2.1593286  -76.158061  76.188667
#> 91    85.9081047  -72.905110 112.673677
#> 92    67.1231014  101.804272 121.941053
#> 93   -77.7933645  100.480077 127.074991
#> 94   -16.9624141   29.322657  33.875385
#> 95    45.5900186  -16.402085  48.450781
#> 96   124.2484332  -61.967257 138.843848
#> 97    39.6781613    8.950376  40.675124
#> 98    15.2087664    2.425932  15.401030
#> 99    20.3629799   86.457759  88.823393
#> 100 -109.4262365   15.186038 110.474961
#> 101  -70.3525997  -81.969119 108.020483
#> 102   -9.7365872 -160.168255 160.463925
#> 103   22.2099989 -184.973652 186.302271
#> 104  -52.0966595  -17.497587  54.956596
#> 105 -130.2319404   65.645837 145.841469
#> 106 -102.5352124  126.518389 162.850768
#> 107   -7.2346528   84.623236  84.931927
#> 108  102.5223058   69.956208 124.115649
#> 109  136.2525202   55.064073 146.958502
#> 110   59.0423360  -89.689117 107.378467
#> 111  -40.7207090 -147.527672 153.044406
#> 112  -29.3153548  -90.772920  95.389271
#> 113   87.1672668   42.365425  96.917293
#> 114   86.6679380   28.640228  91.277567
#> 115    7.4756903  -24.608802  25.719236
#> 116  -51.9043607   55.100599  75.697679
#> 117  -55.2738527   58.921381  80.789405
#> 118   -2.9202614  -19.286335  19.506170
#> 119  -33.0748140  -86.736702  92.828868
#> 120  -54.2834585  -43.040777  69.276276
#> 121  -15.8713461   -8.885887  18.189519
#> 122   13.9334737  -25.322578  28.902849
#> 123   80.5693495   -4.205423  80.679029
#> 124  121.6099647   48.577321 130.953196
#> 125   87.9528595   60.184479 106.573341
#> 126  -10.4718394   36.054733  37.544683
#> 127  -54.2085894   -1.971943  54.244444
#> 128   -7.6555411  -28.138821  29.161628
#> 129  -25.2245169  -47.601764  53.872109
#> 130 -124.6840771  -79.899966 148.088229
#> 131  -65.3277127 -120.435644 137.012607
#> 132   61.3695580  -69.372023  92.621273
#> 133   79.2174293   77.291904 110.677186
#> 134  -58.4256407   87.711933 105.389462
#> 135  -40.8063976   19.521271  45.235408
#> 136   31.6442436   83.758507  89.536840
#> 137   22.5821779  199.855153 201.126918
#> 138   46.0348391  139.977394 147.352900
#> 139   21.3294908 -173.112346 174.421419
#> 140   -2.8584213 -166.359040 166.383595
#> 141 -152.0441459   75.245227 169.644530
#> 142 -115.2572450  115.726523 163.330526
#> 143   99.7436937  -83.540615 130.107028
#> 144  192.2052679 -161.042477 250.753952
#> 145  122.2357207  -13.269452 122.953852
#> 146  -79.2831233   48.366300  92.871484
#> 147 -124.6779527   39.690914 130.843267
#> 148 -155.0671813   -7.024803 155.226217
#> 149  -61.7520185   80.316137 101.311370
#> 150   57.3052414   66.800534  88.012511
#> 151   96.2899133  -27.550195 100.153685
#> 152   55.8116732 -126.334827 138.113835
#> 153  -62.1316821 -208.792121 217.840528
#> 154  -54.5155275  -14.335947  56.368982
#> 155  -38.3339780  110.188570 116.666254
#> 156   -6.3531770   83.347350  83.589136
#> 157   71.7132515  -37.426627  80.892168
#> 158   89.7507083  -88.552938 126.082562
#> 159  -60.3276896    7.259744  60.762933
#> 160 -110.3370710   89.527374 142.089479
#> 161  -75.7089437   91.300833 118.607277
#> 162   92.6967713   29.450318  97.262596
#> 163  111.8663496  -13.954319 112.733328
#> 164   94.8441912 -113.752172 148.104615
#> 165   52.9594511  -41.923019  67.544378
#> 166   15.9401589  105.406531 106.604997
#> 167   23.2041747  182.047429 183.520299
#> 168  -52.4645934   17.029887  55.159320
#> 169  -66.2072380 -104.315151 123.551807
#> 170  -42.0869770  -57.410572  71.184882
#> 171   67.8260933   68.325023  96.274024
#> 172   68.3094925   77.669078 103.434387
#> 173   32.0294241  -42.259975  53.026310
#> 174   57.4589679  -35.591351  67.589032
#> 175   46.6591377   49.634605  68.122457
#> 176  -21.0302160   48.264136  52.646907
#> 177 -107.9078097    6.861579 108.125745
#> 178 -103.0019179   43.257981 111.716821
#> 179 -156.3450989   76.968139 174.263835
#> 180 -203.6171024    2.295395 203.630040
#> 181  -19.8475709 -209.507203 210.445228
#> 182  113.5222173 -218.959423 246.638445
#> 183  148.6013210 -102.022678 180.252543
#> 184   36.3872882    5.877083  36.858850
#> 185   35.5441380   92.510496  99.103873
#> 186   44.0944068  102.863130 111.915773
#> 187  -20.6225160   54.442243  58.217231
#> 188  -44.0520236   77.755965  89.367617
#> 189  -62.2184983   35.602627  71.684647
#> 190   10.6662592  -34.691938  36.294623
#> 191   37.2414846  -91.201305  98.511960
#> 192   36.8505109  -76.504563  84.917067
#> 193  -77.7581082   61.046664  98.858578
#> 194 -193.2036043   96.054948 215.764190
#> 195  -43.3585672   55.400193  70.350172
#> 196   83.8441936  -15.478517  85.260972
#> 197  173.4338451   10.976375 173.780837
#> 198   86.1416946   48.132338  98.676814
#> 199  -48.8263982   74.381109  88.975089
#> 200  -29.6575181  -18.903279  35.169622
#> 201  -52.0137124 -127.909092 138.080274
#> 202   27.9834518 -125.915165 128.987218
#> 203   73.4003166  -63.969470  97.363748
#> 204   35.7661466   81.791755  89.269863
#> 205    1.3502060    7.589081   7.708256
#> 206  -78.0444966  -64.738594 101.400341
#> 207  -78.2389653  -23.476397  81.685231
#> 208  -19.7633011   60.211826  63.372329
#> 209  109.5560449   43.585734 117.907774
#> 210  122.8360322  -84.866538 149.301775
#> 211   70.5852045  -91.037229 115.195695
#> 212   75.5104432  -14.322847  76.856821
#> 213  -36.4078298   97.324758 103.911686
#> 214 -174.3098080  103.480370 202.711855
#> 215 -212.0710623    4.309424 212.114843
#> 216  -11.4176230  -87.932084  88.670251
#> 217  176.1531444 -107.182853 206.199161
#> 218   79.5443545   24.046655  83.099615
#> 219  -73.0961183  139.046113 157.088714
#> 220 -108.9353662  113.419320 157.260472
#> 221  -58.4847881    9.706285  59.284757
#> 222   33.9476551  -35.892321  49.403461
#> 223  -21.7023320    7.099472  22.834047
#> 224   -5.1893977   12.482070  13.517837
#> 225  139.0889941  -13.986744 139.790476
#> 226  132.7725624  -31.560679 136.472084
#> 227   28.8950325  -11.998442  31.287146
#> 228  -38.8656083   27.810633  47.790865
#> 229  -63.9821425  -15.797381  65.903504
#> 230  -11.6835563  -85.543877  86.338058
#> 231  -27.8082494  -84.384532  88.848455
#> 232   29.3906418   39.360162  49.122624
#> 233   44.0502478   62.671570  76.603851
#> 234  -74.7830488   32.187234  81.415738
#> 235  -56.2941416   52.688498  77.104528
#> 236   40.3143975    6.810729  40.885654
#> 237   93.8888253  -33.550974  99.703457
#> 238   -5.5853739  -78.156128  78.355451
#> 239  -49.4023213   69.428915  85.211288
#> 240   42.0755967   99.124304 107.684649
#> 241  147.1305573  -24.273100 149.119362
#> 242   48.8966559  -28.486830  56.589596
#> 243 -121.5275996   -7.180643 121.739554
#> 244 -108.7623032   78.175760 133.942854
#> 245  -68.0399142    3.391061  68.124366
#> 246  -99.8064867  -94.895776 137.719073
#> 247  -69.7387357  -54.714269  88.640524
#> 248   50.7693633   24.394872  56.326175
#> 249  160.2630501  102.139342 190.043917
#> 250  -13.4803759   26.781990  29.983254
#> 251 -135.7750625  -35.645589 140.376193
#> 252   29.8614773  -22.708747  37.515264
#> 253  112.1406168  -11.089503 112.687599
#> 254   64.1551353    2.234295  64.194030
#> 255 -116.0637314  -66.913768 133.971049
#> 256  -16.2913573  -14.515670  21.820014
#> 257   84.6569417  -32.028393  90.513070
#> 258   -1.1206085  -77.017158  77.025310
#> 259  -45.0802745  -37.107110  58.388088
#> 260  -95.4551882   99.181056 137.653823
#> 261  -38.9827059  187.250957 191.265711
#> 262   12.4841921  -29.299187  31.848036
#> 263  116.4303175 -123.655114 169.842886
#> 264  164.1682728  -74.036517 180.090610
#> 265   66.9241175  117.775076 135.461456
#> 266  -34.6502743  156.773450 160.557018
#> 267  -76.9184285   47.398625  90.349733
#> 268   12.4718796  -13.269470  18.210618
#> 269   54.0414625 -123.607031 134.904329
#> 270  -36.2373949 -103.902218 110.040083
#> 271 -146.8298676 -124.806071 192.705904
#> 272 -171.2842740    3.402428 171.318064
#> 273   -9.1404871  140.287793 140.585253
#> 274   87.5028504  100.793539 133.476913
#> 275   68.2058816  -14.769597  69.786698
#> 276   88.9749925  -82.061925 121.040113
#> 277   25.8620279  -18.588083  31.849039
#> 278   45.6240439   11.757140  47.114581
#> 279   78.5248818   32.874238  85.128565
#> 280    7.5621881   19.995142  21.377380
#> 281 -116.9870937  -37.993134 123.001863
#> 282 -160.4191981    5.738684 160.521810
#> 283   -3.1647341   76.881271  76.946380
#> 284   52.7825399   33.852934  62.705802
#> 285   42.3195582   16.871907  45.558822
#> 286   -0.6388842  -42.554494  42.559290
#> 287   54.9752106    3.489653  55.085855
#> 288   89.3572575   -5.888778  89.551087
#> 289   27.7243127  -62.537717  68.407628
#> 290  -59.4951766  -59.599175  84.212456
#> 291 -155.4279067  -76.370682 173.177121
#> 292  -37.9964524   33.929555  50.940604
#> 293   21.1777893   66.408813  69.703868
#> 294  -18.4689995    4.058779  18.909723
#> 295  -57.7308583  -73.444854  93.418406
#> 296  -11.7874628 -129.413146 129.948861
#> 297   77.4020812  -33.014737  84.149005
#> 298   18.9738752   98.874604 100.678673
#> 299  -19.2673121  101.337466 103.152854
#> 300  -17.3597512  -12.220464  21.229712
#> 301   54.0573643 -154.285267 163.481321
#> 302   98.1269042  -34.015594 103.855428
#> 303   -1.4527063   91.890966  91.902449
#> 304  -81.9992805  100.874600 129.998335
#> 305  -61.7671428   59.349962  85.659780
#> 306   37.1171980  -53.723470  65.298527
#> 307   62.1844392   48.036860  78.577633
#> 308  -79.2931066  111.892600 137.139894
#> 309 -103.5288058   10.504952 104.060404
#> 310   50.6084159  -77.162485  92.278171
#> 311  130.2852874  -38.156554 135.757794
#> 312   87.3181436   67.600667 110.427843
#> 313  -89.3603533  -21.547315  91.921486
#> 314  -74.3585057 -179.918458 194.678810
#> 315   23.1307459 -108.994387 111.421756
#> 316   30.3218507   58.322071  65.733390
#> 317    0.7640500  172.069594 172.071290
#> 318   -0.9988576  181.291147 181.293899
#> 319   62.0877631   48.242505  78.627155
#> 320   39.6929932 -146.177828 151.471090
#> 321  -44.2922251 -203.796201 208.553813
#> 322  -43.2008848   16.443607  46.224546
#> 323    9.3679212   76.870316  77.439030
#> 324   61.0342659  -25.532049  66.159407
#> 325   68.6413713 -111.382387 130.834529
#> 326   32.8157664  114.971111 119.562665
#> 327    9.8747736  201.856646 202.098037
#> 328  -21.7033942    3.341335  21.959095
#> 329  -63.6937807  -75.197720  98.547424
#> 330 -103.9931656  -89.629977 137.288423
#> 331 -143.2378396  -66.635234 157.978901
#> 332  -53.6821424 -161.160588 169.866146
#> 333   43.8657247  -23.265124  49.653477
#>  [ reached 'max' / getOption("max.print") -- omitted 667 rows ]