Extract saccades using an algorithm proposed by Nyström and Holmqvist (2010) doi:10.3758/BRM.42.1.188 .

Extract saccades using an algorithm proposed by Nyström and Holmqvist (2010) doi:10.3758/BRM.42.1.188 .

Usage

method_nh(x, y, vel, acc, sample_rate, trial, options)

Arguments

x

Gaze x coordinate, _arbitrary units_ as threshold velocity is computed in units of standard deviation.

y

Gaze y coordinate, _arbitrary units_ as threshold velocity is computed in units of standard deviation.

vel

Velocity data.frame with columns x, y, amp.

acc

Acceleration data.frame with columns x, y, amp.

sample_rate

Sample rate in Hz.

trial

Trial id, so that trial borders are respected when computing velocity and saccades.

options

Named list with method options. See details for further information.

Value

logical vector marking samples that belong to saccades

Details

Method options, please refer to Nyström and Holmqvist (2010) for details on parameters and the rationale for default values.

nh_sg_filter_order Order of Savitzky-Golay filter. Defaults to 2.

nh_max_velocity Maximal physiologically plausible velocity in °/s. Defaults to 1000.

nh_max_acceleration Maximal physiologically plausible acceleration in °/s². Defaults to 100000.

nh_initial_velocity_threshold Initial velocity threshold in °/s. Defaults to 100.

See also

Examples

# Do not run this function directly, use extract_saccades() instead