Compute velocity for x, y, and its amplited

Compute velocity for x, y, and its amplited

compute_velocity_table(x, y, trial, sample_rate, velocity_time_window)

Arguments

x

vector with x coordinates in degrees of visual angle

y

vector with y coordinates in degrees of visual angle

trial

vector with trial index

sample_rate

sample rate in Hz

velocity_time_window

Time window for velocity computation in milliseconds

Value

data.frame with columns x, y, and amp

Examples

compute_velocity_table(rnorm(1000), rnorm(1000), rep(1, 1000), 250, 20)
#>                 x             y         amp
#> 1              NA            NA          NA
#> 2              NA            NA          NA
#> 3              NA            NA          NA
#> 4   -5.153071e+01 -146.85472936 155.6333052
#> 5    8.487753e+01   -1.37691892  84.8886962
#> 6    9.805078e+01  204.23489929 226.5520906
#> 7    6.608607e+01  252.40085428 260.9091039
#> 8    1.222015e+02   86.47018238 149.7006749
#> 9    1.247553e+02  -93.43353899 155.8643785
#> 10   8.152366e+01 -129.57609542 153.0884436
#> 11   2.101932e+01 -176.80099346 178.0460699
#> 12  -9.272220e+01  -62.89799597 112.0426893
#> 13  -1.274307e+02    6.20893150 127.5818457
#> 14  -5.403967e+01   22.87576477  58.6820811
#> 15   3.415402e+01  -46.49583779  57.6919424
#> 16   6.449446e+01  -49.96852037  81.5866938
#> 17   9.112795e+00  147.07838618 147.3604245
#> 18  -1.073870e+02  159.78158087 192.5152576
#> 19  -1.111527e+02   41.67597658 118.7089323
#> 20  -3.659109e+01  -43.83395818  57.0992464
#> 21   5.005116e+01  -27.31506699  57.0195727
#> 22   8.932713e+01   62.95657694 109.2834210
#> 23  -4.340368e+01   21.02642637  48.2285160
#> 24   8.409623e+01 -105.50737403 134.9221344
#> 25   5.954657e+01  -41.78937893  72.7471397
#> 26  -9.701300e+00   45.14191167  46.1725830
#> 27  -9.239486e+01  106.19229712 140.7608430
#> 28  -9.651622e+01   37.36648472 103.4970278
#> 29   5.777011e+01   12.07818461  59.0192208
#> 30   1.070853e+02   49.50665686 117.9752783
#> 31   8.195251e+01  -13.49105394  83.0555404
#> 32  -1.946016e+01 -121.57497477 123.1225910
#> 33  -1.599159e+01 -108.10045039 109.2768886
#> 34  -4.824830e+01  -16.34758633  50.9425395
#> 35  -4.782433e+01   -5.29512239  48.1165713
#> 36  -5.617267e+01   51.27570393  76.0563382
#> 37  -6.946964e+00   73.80813806  74.1343480
#> 38  -1.364455e+01   10.71102109  17.3464629
#> 39   6.078930e+00  -94.83068215  95.0253212
#> 40   9.062723e+01  -62.66311655 110.1814876
#> 41   4.091511e+01   52.28021752  66.3872556
#> 42  -6.674733e+01  107.32842946 126.3906557
#> 43  -8.141355e+01   32.06197994  87.4993539
#> 44   2.008528e+01  -79.89061584  82.3767506
#> 45  -3.834433e+01  -72.66680445  82.1629594
#> 46  -8.319344e+01   10.60503829  83.8666487
#> 47  -7.532723e+01  101.24297102 126.1916423
#> 48   1.407791e+01   48.13390473  50.1503759
#> 49   5.601810e+01  -38.48592373  67.9646522
#> 50   2.805570e+00  -43.08334496  43.1745971
#> 51   2.995745e+01   48.67844217  57.1580218
#> 52  -2.304741e+01   73.99520836  77.5014457
#> 53   6.893314e+01   53.04863877  86.9823860
#> 54   9.607843e+01  -52.52886780 109.5004430
#> 55  -5.422548e+01 -122.04955194 133.5533437
#> 56   1.448207e+01  -87.97386534  89.1579004
#> 57   3.402657e+01  -40.20936657  52.6744799
#> 58   1.096563e+02   40.03765490 116.7369633
#> 59  -2.274179e+01   60.27659657  64.4240427
#> 60  -1.510843e+02   -9.53308459 151.3847553
#> 61   4.668702e+01  -40.64630940  61.9015405
#> 62   1.187806e+02  -57.23437611 131.8506855
#> 63   1.273581e+02  -20.03199238 128.9238796
#> 64  -1.490979e+01  -14.56453594  20.8429237
#> 65  -1.549712e+02  -18.89388808 156.1186834
#> 66  -1.768234e+02   81.98034417 194.9032757
#> 67  -1.031722e+02   65.27868139 122.0893250
#> 68   3.334228e+01   42.88513870  54.3216582
#> 69   7.851124e+01   -9.11794162  79.0389211
#> 70   1.055619e+02   28.94026780 109.4571286
#> 71   9.387849e+01   87.38230114 128.2530217
#> 72   7.223290e+01   60.18606791  94.0210357
#> 73  -6.109650e+01   30.95283656  68.4898569
#> 74  -1.005260e+02    0.85492785 100.5296021
#> 75  -8.995053e+00  -72.79556972  73.3492055
#> 76   6.374756e+00  -72.96304147  73.2409922
#> 77   3.624943e+01  -36.20265619  51.2313746
#> 78  -7.950678e+00   17.23140771  18.9772150
#> 79  -5.005629e+01  -49.98850576  70.7423731
#> 80  -9.800983e+01  -82.90327932 128.3700911
#> 81  -7.278775e+01   -3.77972806  72.8858224
#> 82   5.188293e+01   15.46083992  54.1375681
#> 83   1.824490e+02   38.52627340 186.4722979
#> 84   1.715587e+02   18.19126407 172.5205059
#> 85   7.125846e+01   40.62784995  82.0267618
#> 86  -7.702252e+01   14.64259295  78.4020072
#> 87  -6.817301e+01  -41.91116189  80.0256518
#> 88   2.210820e+00  -41.17547996  41.2347896
#> 89  -3.256423e+01   19.39978345  37.9048936
#> 90  -1.089287e+02   81.34749508 135.9517440
#> 91  -5.959419e+01   67.86123670  90.3139789
#> 92   1.168333e+02  -44.55313197 125.0399562
#> 93   9.482612e+01  -70.40849359 118.1073643
#> 94  -4.958293e+01  -84.94410318  98.3563304
#> 95  -1.636919e+02  -64.47471372 175.9319158
#> 96  -1.472748e+02   52.59557804 156.3846923
#> 97  -6.334647e-01  197.30521984 197.3062367
#> 98   1.170891e+02  150.96105508 191.0473775
#> 99   1.454318e+02  -59.29066859 157.0535043
#> 100 -4.150574e+00  -66.32508126  66.4548242
#> 101 -1.051934e+02   -7.73009645 105.4770686
#> 102 -6.228651e+01    0.99140576  62.2944010
#> 103  2.161841e+01 -107.12237906 109.2820185
#> 104  9.701773e+01  -16.92151567  98.4823760
#> 105  8.702695e+01   17.07659720  88.6865236
#> 106  1.783122e+01  -74.05462543  76.1711233
#> 107 -1.010801e+02  -49.40817790 112.5093917
#> 108 -6.348741e+01    0.07271555  63.4874555
#> 109  1.319565e+01   69.33825482  70.5827084
#> 110  3.013923e+01   17.47076993  34.8367756
#> 111  6.021873e+01  -14.47733464  61.9345507
#> 112  5.587023e-01   -0.64901456   0.8563692
#> 113 -7.901623e+01   23.34017289  82.3913100
#> 114 -5.144162e+01  -31.45146240  60.2945658
#> 115 -3.519686e+01  -45.19437481  57.2830735
#> 116  2.935567e+01  -80.44227087  85.6312707
#> 117  1.987711e+01  -48.44794059  52.3669994
#> 118  4.402052e+01  113.08153046 121.3475953
#> 119  4.880032e+01  160.26636637 167.5314280
#> 120 -5.028006e+00  144.72509245 144.8124070
#> 121  3.620947e+00  -58.27650500  58.3888884
#> 122 -3.190139e+01 -111.63919774 116.1077478
#> 123  1.467918e+01  -36.72443306  39.5494919
#> 124  8.314123e+01  -30.14262103  88.4366540
#> 125  1.078351e+02   16.60369260 109.1058961
#> 126 -1.896659e+01   50.48031287  53.9258133
#> 127 -1.268465e+02   67.30161974 143.5950546
#> 128 -3.347322e+01  -35.19522330  48.5711844
#> 129  3.746583e+01  -70.87957617  80.1723316
#> 130  2.145578e+01   46.16023037  50.9030181
#> 131  1.888094e+01   24.12353241  30.6338836
#> 132  6.922038e-01   14.00890929  14.0260003
#> 133 -7.356790e+01  -58.54569896  94.0203932
#> 134 -1.678892e+02  -44.42806565 173.6681512
#> 135 -5.642920e+01   49.32606994  74.9487514
#> 136  2.255937e+02  -61.18366194 233.7433174
#> 137  2.250156e+02  -71.54792638 236.1168076
#> 138 -1.936384e+01  -24.15777561  30.9605622
#> 139 -9.248930e+01   98.40491769 135.0473923
#> 140 -3.554587e+01  119.05382399 124.2470210
#> 141 -1.924474e+01  -76.63509180  79.0145380
#> 142 -1.286356e+02  -63.67358044 143.5320237
#> 143 -1.533471e+02   50.34159374 161.3989562
#> 144  1.184463e+02   75.95171685 140.7060606
#> 145  1.645803e+02   91.96764333 188.5330657
#> 146  1.133163e+02   36.85767653 119.1598614
#> 147 -2.766502e+01    1.56064714  27.7090021
#> 148 -6.976712e+01  -57.43985339  90.3702843
#> 149  3.503710e+01 -105.73291102 111.3869225
#> 150  2.038960e+01  -31.53948093  37.5562889
#> 151  6.780199e+01   -0.57910739  67.8044624
#> 152  3.822633e+01  -32.90908396  50.4406601
#> 153 -4.898032e+01   30.41503372  57.6554050
#> 154 -8.536426e+01   74.53184887 113.3227847
#> 155 -4.888603e+01  100.21810389 111.5056616
#> 156 -6.454114e+01   55.79444258  85.3145822
#> 157 -5.723892e+01  -63.39456908  85.4117385
#> 158 -4.195252e+01  -18.57458014  45.8805967
#> 159  1.435076e+02  -19.53780309 144.8314399
#> 160  1.591693e+02  -56.59876625 168.9327443
#> 161  2.199203e+01  -67.12411430  70.6349505
#> 162 -1.103769e+02   -9.32422029 110.7699920
#> 163 -1.688650e+02   37.60471343 173.0014210
#> 164 -2.643405e+01  -25.11364218  36.4616759
#> 165  2.435379e+01  -66.45301979  70.7750733
#> 166  1.215545e+02  -73.31488630 141.9527387
#> 167  3.669638e+01  -44.66487276  57.8063594
#> 168 -7.210113e+01   60.29957860  93.9926205
#> 169 -1.372323e+02  192.58968682 236.4814913
#> 170 -1.204093e+02   77.07586810 142.9653383
#> 171  9.359106e+00 -153.41777816 153.7029848
#> 172  3.298148e+01 -186.55504759 189.4480502
#> 173  5.122442e+01  -20.10906029  55.0301360
#> 174  9.002179e+01   20.03835896  92.2250408
#> 175  5.829534e+01  -15.31523207  60.2735697
#> 176 -1.054488e+01    1.23662068  10.6171464
#> 177 -3.100508e+01   -7.27099068  31.8462296
#> 178  9.956009e+01  118.72036149 154.9410735
#> 179  2.060557e+02   57.36194939 213.8909471
#> 180  7.975793e+01   -4.03866062  79.8601161
#> 181 -1.170119e+02  -26.01849393 119.8697451
#> 182 -1.912807e+02   16.17403952 191.9633004
#> 183 -6.444355e+01   87.51034233 108.6785671
#> 184  8.653688e+01  -64.99608410 108.2271771
#> 185  5.116956e+00  -91.71766154  91.8602889
#> 186 -1.254776e+02   33.05338123 129.7580592
#> 187 -1.049601e+02  103.04058659 147.0849704
#> 188  3.502245e+01   50.72530477  61.6411275
#> 189  1.055293e+02  -42.22287431 113.6626632
#> 190 -2.482619e+00  -67.23364311  67.2794632
#> 191 -3.483083e+01  -13.61776230  37.3982679
#> 192  1.821007e+01   32.21585388  37.0063233
#> 193  3.464246e+01   51.53546340  62.0967315
#> 194 -3.186464e+01   42.01200510  52.7291560
#> 195 -1.794636e+02  -30.06448165 181.9644186
#> 196 -1.593865e+02  -48.21898476 166.5206030
#> 197  2.761442e+01  -71.57253218  76.7149488
#> 198  2.635380e+02  -39.65898065 266.5054151
#> 199  2.881034e+02   14.19813492 288.4530222
#> 200  5.019955e+01   66.60214891  83.4016855
#> 201 -7.273057e+01   70.93473141 101.5946452
#> 202 -1.004180e+02  -29.32881699 104.6133427
#> 203 -2.204693e+01  -20.52348943  30.1211039
#> 204  1.441338e+01   -3.57059124  14.8490641
#> 205 -1.012334e+02  -20.37175515 103.2628380
#> 206 -1.182996e+02   17.21389407 119.5454767
#> 207 -8.365958e+01   45.63139393  95.2950690
#> 208  2.742892e+01  122.73074043 125.7584209
#> 209  1.240781e+02    7.14421052 124.2835710
#> 210  4.715395e+01  -61.51911272  77.5119103
#> 211  2.561330e+01  -23.66682810  34.8734860
#> 212 -4.179780e+01   42.78135826  59.8105391
#> 213 -2.270687e+01   55.86772464  60.3059252
#> 214 -2.101737e+00 -104.54270471 104.5638293
#> 215 -2.719460e+01  -75.71197676  80.4478059
#> 216  8.053049e+01   -2.80099792  80.5791851
#> 217  1.053937e+02  105.30687314 148.9878460
#> 218  1.500008e+02   50.09974649 158.1462210
#> 219  1.822706e+01  -89.08302380  90.9286033
#> 220 -8.488076e+00  -91.66109776  92.0532687
#> 221  3.302559e+01  -69.77361146  77.1948623
#> 222 -7.488284e+01  -15.04867231  76.3799903
#> 223 -1.130794e+02  -21.05792678 115.0234153
#> 224 -2.052563e+02    3.09496332 205.2796144
#> 225 -1.104430e+02   27.69213158 113.8618247
#> 226 -8.597505e+01    3.67445188  86.0535380
#> 227 -2.417098e+01   31.71406081  39.8750263
#> 228  1.445359e+02   53.30724023 154.0529140
#> 229  1.483983e+02   89.93342970 173.5225653
#> 230  1.308666e+02  -58.33619476 143.2800862
#> 231 -2.973761e+01  -83.83947354  88.9571971
#> 232 -1.223413e+02   48.76843053 131.7032621
#> 233 -3.389566e+01   44.73639658  56.1271831
#> 234  3.377512e+01  -52.48990702  62.4175377
#> 235  7.271048e+01 -189.48310256 202.9548233
#> 236  2.379312e+01  -64.30253873  68.5633228
#> 237 -1.868356e+01  169.22561537 170.2538814
#> 238 -2.335776e+01  238.70926906 239.8493280
#> 239 -8.079137e+01  164.11913607 182.9271321
#> 240 -3.731357e+01  -77.88668541  86.3634090
#> 241  1.251409e+02 -147.80746006 193.6679592
#> 242  1.021196e+02  -41.47696428 110.2213355
#> 243 -2.698110e+01   12.15907533  29.5943065
#> 244 -1.234042e+02   16.38477589 124.4872092
#> 245 -5.382981e+01  -20.48993516  57.5976193
#> 246  1.012833e+02  -41.77061537 109.5586494
#> 247  4.842461e+01  -32.06878873  58.0805448
#> 248 -1.742841e+01  -95.30225274  96.8827585
#> 249 -8.149982e+01  -94.58641480 124.8551569
#> 250 -6.632381e+01  -67.74909962  94.8092218
#> 251  3.683412e+01  -29.61040541  47.2602202
#> 252  4.199304e+01   68.40671386  80.2676411
#> 253  2.727039e+01   34.27840810  43.8027774
#> 254 -9.555834e+00   14.10153019  17.0342925
#> 255 -2.380421e+01    9.49651540  25.6285847
#> 256 -2.843262e+01   89.76427684  94.1596489
#> 257  8.575153e+01  209.77804680 226.6277862
#> 258  1.478337e+02  189.79473347 240.5760558
#> 259  8.680435e+01   21.76346543  89.4910255
#> 260 -8.821181e+01  -64.10870642 109.0470051
#> 261 -1.427784e+02  -57.94606372 154.0890135
#> 262 -8.519512e-03  -53.19037296  53.1903736
#> 263  6.419740e+01 -113.73835100 130.6052008
#> 264 -7.138068e+00 -137.76697805 137.9517751
#> 265 -3.239246e+01  -23.01752397  39.7376124
#> 266 -5.164282e+00   55.60318559  55.8424933
#> 267  6.206439e+01   56.78895655  84.1247539
#> 268  3.891064e+01  -16.58195524  42.2965655
#> 269 -9.004090e+01  -28.22493001  94.3610600
#> 270 -8.868495e+01   48.34769821 101.0075225
#> 271 -5.283417e+01  146.43849639 155.6781394
#> 272  1.892042e+01  128.70376108 130.0870493
#> 273 -2.402888e+00   -5.90307049   6.3733908
#> 274  5.690220e+01  -74.48324182  93.7316062
#> 275  1.663724e+02 -110.17345852 199.5444281
#> 276  1.544926e+02  -58.95401478 165.3588472
#> 277 -2.089440e+01  -62.72313059  66.1117763
#> 278 -2.093339e+02  -58.52132319 217.3601455
#> 279 -1.525067e+02   30.41197207 155.5094376
#> 280  1.850440e+01   37.57418498  41.8835546
#> 281  6.090103e+01  102.18798622 118.9593216
#> 282  1.037976e+01    2.49256530  10.6748466
#> 283 -4.232479e+01  -15.23113990  44.9819506
#> 284 -4.459223e+01   23.14333820  50.2402319
#> 285 -1.685595e+01   18.69394027  25.1711442
#> 286  4.266394e+01    4.92929635  42.9477574
#> 287  7.977816e+01 -100.84873328 128.5885748
#> 288  8.537569e+00  -84.23930731  84.6708390
#> 289 -9.023723e+01   78.66727746 119.7134034
#> 290 -1.025733e+01   68.10194847  68.8700819
#> 291  9.220393e+01   26.42563638  95.9159975
#> 292  4.641726e+01  -52.37732553  69.9853273
#> 293 -6.474438e+01  -51.05545229  82.4529802
#> 294 -8.593109e+01   93.41758029 126.9290996
#> 295  2.118615e+01   -0.64489205  21.1959606
#> 296  5.564068e+01  -19.43428133  58.9370569
#> 297  2.162547e+01  -47.19038294  51.9094720
#> 298  3.161700e+01  -42.09707010  52.6478669
#> 299  1.186384e+01  -29.03094029  31.3615399
#> 300  5.054454e+01  -79.67283353  94.3531169
#> 301  7.156604e+01   37.39919559  80.7489768
#> 302  4.577264e+01  106.56594514 115.9803241
#> 303 -9.472303e-01   95.81907053  95.8237524
#> 304 -4.942233e+01   10.65030578  50.5568567
#> 305 -2.860129e+01  -12.96726910  31.4035624
#> 306 -5.616771e+01  -50.97751699  75.8519518
#> 307 -8.502674e+01 -104.48565729 134.7100577
#> 308 -2.168787e+01  -84.80352002  87.5328561
#> 309 -3.825061e+01  -26.96986619  46.8025955
#> 310 -3.218571e+01  103.14777513 108.0526887
#> 311  4.700593e+01   19.12793243  50.7487421
#> 312  8.355778e+01  -64.55146544 105.5878483
#> 313  7.811912e+01  -85.77143448 116.0143787
#> 314 -6.149562e+01   25.16684697  66.4460765
#> 315 -4.180466e+01  180.54367015 185.3203885
#> 316 -4.278678e+01   99.86860267 108.6482673
#> 317 -4.383257e+01   -6.67092761  44.3372968
#> 318  5.828953e+00  -77.92044561  78.1381631
#> 319  9.140245e+01  -39.08186134  99.4072383
#> 320  8.412935e+01   43.13753832  94.5441399
#> 321  3.317044e+01   59.53581555  68.1527067
#> 322 -2.447425e+01   19.75737176  31.4538193
#> 323 -4.501046e+01    5.55961803  45.3525138
#> 324 -7.867325e+01   25.27091736  82.6323140
#> 325 -4.125516e+01   42.22327852  59.0321386
#> 326  1.097569e+02   -3.56627422 109.8148448
#> 327  1.008352e+02  -27.85319647 104.6113272
#> 328 -2.131074e+01    5.12724063  21.9188580
#> 329 -1.294647e+02   -6.42162594 129.6238802
#> 330 -4.178653e+01  -53.03966985  67.5227424
#> 331  3.361126e+01  -82.05837458  88.6752133
#> 332  3.629019e+01   17.55322530  40.3124520
#> 333  6.361058e+01   43.00682063  76.7847115
#>  [ reached 'max' / getOption("max.print") -- omitted 667 rows ]