Compute velocity for x, y, and its amplited

Compute velocity for x, y, and its amplited

compute_velocity_table(x, y, trial, sample_rate, velocity_time_window)

Arguments

x

vector with x coordinates in degrees of visual angle

y

vector with y coordinates in degrees of visual angle

trial

vector with trial index

sample_rate

sample rate in Hz

velocity_time_window

Time window for velocity computation in milliseconds

Value

data.frame with columns x, y, and amp

Examples

compute_velocity_table(rnorm(1000), rnorm(1000), rep(1, 1000), 250, 20)
#>                x            y        amp
#> 1             NA           NA         NA
#> 2             NA           NA         NA
#> 3             NA           NA         NA
#> 4     62.1284853   33.8026510  70.728834
#> 5     32.0980087  113.7720839 118.213236
#> 6    -92.1530694   18.5804798  94.007566
#> 7    -59.5874337 -110.6544510 125.678438
#> 8     81.9707509 -143.5694046 165.322043
#> 9    158.8966352   37.1228427 163.175507
#> 10    50.7921986  192.2850498 198.880335
#> 11   -94.8669379  167.1706934 192.212842
#> 12   -91.8961412  -43.2446651 101.562797
#> 13   -48.3993336  -82.9010290  95.995188
#> 14    32.9411643   -7.7864795  33.848923
#> 15     7.1519197  -28.5087044  29.392111
#> 16     6.1870804  -21.2258038  22.109154
#> 17    -3.9980084  -54.6437708  54.789833
#> 18    14.3517418  -78.6942004  79.992185
#> 19    65.8866819  -21.2204148  69.219657
#> 20   -14.2159709   53.4407435  55.299249
#> 21    30.4054034  144.7304389 147.889785
#> 22    61.7121138   97.5150584 115.401783
#> 23    -4.6947345  -51.2556186  51.470176
#> 24  -139.5326039 -158.2798527 211.002036
#> 25  -207.0312306 -145.0819826 252.805680
#> 26   -54.1172239  -18.8627423  57.310357
#> 27    33.5602691   81.1809646  87.844412
#> 28     8.8356986   75.4901077  76.005434
#> 29    35.3834856  -40.3000815  53.629168
#> 30    96.8225842  -73.5917501 121.615618
#> 31   -10.6198401   49.0817660  50.217534
#> 32   -44.2256005   70.1578940  82.933912
#> 33   -58.8343326   30.0450513  66.061969
#> 34   107.0948853    5.2076443 107.221425
#> 35   104.7567011   29.0223280 108.702631
#> 36   -25.5014298   93.0393209  96.470919
#> 37   -84.6610326  -83.8226768 119.137448
#> 38  -178.8418784 -205.9035654 272.728245
#> 39   -78.1953095  -90.5910703 119.671419
#> 40   -39.8646854   26.5260236  47.883432
#> 41    88.4379300  135.7473047 162.014191
#> 42    80.1904094   98.8412288 127.279575
#> 43    86.6814741    0.5036535  86.682937
#> 44   135.5050944  -84.6210209 159.757153
#> 45    48.0275167 -170.4569440 177.093795
#> 46   -66.6319396  -23.9482430  70.804899
#> 47  -127.9927593   88.3410786 155.519428
#> 48   -56.3802112   75.2495961  94.027815
#> 49    49.4862396   24.5149066  55.225615
#> 50    62.5513339   36.7356767  72.540880
#> 51    76.0633471  131.7278452 152.111334
#> 52    -5.8565552   46.9820366  47.345655
#> 53  -127.7252417  -24.2607690 130.008932
#> 54  -131.6997988  -64.2517451 146.537107
#> 55    10.6581342    7.1105795  12.812344
#> 56   104.7008042   14.8103496 105.743108
#> 57    46.2758189 -135.6222948 143.299889
#> 58   -28.8727321 -151.2352391 153.966659
#> 59   -25.0484800  -85.0717432  88.682737
#> 60    58.8731517  103.3862126 118.973766
#> 61    59.7922307  205.1497479 213.685587
#> 62    25.6681666  217.3273377 218.837900
#> 63   -47.0020200  122.2325175 130.957925
#> 64    22.0750198  -67.3500884  70.875531
#> 65   -29.4482124 -104.5113171 108.580904
#> 66   -56.7849049  -50.2545051  75.829023
#> 67   -75.6204512    6.3521228  75.886772
#> 68    36.8920261  -39.4279117  53.996128
#> 69    67.0066855  -47.2192692  81.972894
#> 70    10.2805023    3.7727586  10.950910
#> 71   -25.6307267  -24.8872209  35.725452
#> 72   -33.4797738  -58.2228746  67.162477
#> 73   105.7982074  -46.2006336 115.445915
#> 74   102.5633173  138.0388374 171.970796
#> 75     8.3388712  145.3332408 145.572276
#> 76  -115.6916280  -48.8218550 125.571200
#> 77   -20.3196960 -102.9227138 104.909366
#> 78   140.8098935 -102.6751616 174.268801
#> 79    53.9563874    0.4245419  53.958058
#> 80   -62.3087685  -38.4628147  73.224113
#> 81    -5.0577135  -51.8143660  52.060628
#> 82    99.1006892   74.2881948 123.853472
#> 83   -28.1107036  113.7780973 117.199262
#> 84  -194.4911942  121.2652394 229.198785
#> 85  -198.0918001   46.3674075 203.446056
#> 86   -20.1391235   20.7718155  28.931862
#> 87   178.6779404    9.1196323 178.910520
#> 88   159.4269831  -94.7066459 185.435465
#> 89   104.8975050 -167.0703766 197.271380
#> 90   -47.9504812 -122.4511770 131.504903
#> 91  -131.0499979   39.2729879 136.808149
#> 92  -148.3563950   45.4076628 155.149849
#> 93  -152.1337457  -53.7210824 161.340111
#> 94     3.8185277   35.3479265  35.553580
#> 95   118.3950215   47.1847462 127.451094
#> 96   224.2176211   36.1962373 227.120473
#> 97   136.4726182  -32.5797514 140.307575
#> 98     2.8184623  -38.6289587  38.731643
#> 99   -57.5459706   53.7329487  78.732258
#> 100  -95.6891166  -22.8621508  98.382341
#> 101  -48.3325689   -7.3954329  48.895088
#> 102  -34.9543793   39.0346658  52.397650
#> 103    3.9361616   47.0003417  47.164876
#> 104   17.3104306   87.2775936  88.977690
#> 105   19.0103472   31.7841771  37.035486
#> 106   57.9948497  -57.1211933  81.401679
#> 107    6.3185818  -62.3468090  62.666172
#> 108  -96.2775793  -72.7512916 120.673621
#> 109  -64.5666642  -48.5709845  80.796006
#> 110    5.1480764  -80.4629592  80.627480
#> 111   75.6222785  -35.9322101  83.724863
#> 112   63.8084968   86.0196671 107.102322
#> 113   83.0691357   55.3326354  99.810730
#> 114   54.8500189  -11.7412885  56.092624
#> 115  -94.3359575  -53.6690745 108.534061
#> 116 -126.9571510    7.5284827 127.180172
#> 117 -101.7238215  122.5778611 159.289259
#> 118   -7.4067960   75.0664184  75.430947
#> 119   11.7742455  -99.8481542 100.539976
#> 120  -67.9506446 -175.4352174 188.135073
#> 121  -66.1156586  -36.6468680  75.592812
#> 122  -67.8197202  117.2286330 135.432887
#> 123   23.7837955   81.7895209  85.177430
#> 124   65.5081720  -82.2904459 105.180978
#> 125   32.7332621    0.2160056  32.733975
#> 126  101.0153044  159.3189541 188.644165
#> 127   40.6996241   96.9923804 105.185461
#> 128  -14.0589976  -24.8650775  28.564445
#> 129  -91.7235249 -127.7297582 157.251697
#> 130 -165.0369211  -61.7282777 176.203194
#> 131  -55.6091920    9.5918774  56.430367
#> 132   56.5653609   38.9276136  68.665851
#> 133  143.2750216    6.4695997 143.421015
#> 134   93.5720725 -141.6903272 169.799534
#> 135  103.3435074 -191.9313212 217.985120
#> 136   84.3949254   18.3431719  86.365360
#> 137   34.7560106  242.7589425 245.234346
#> 138  -20.7367459  246.5058070 247.376485
#> 139   15.9940325   36.7852145  40.111857
#> 140   17.4719583  -89.6456167  91.332393
#> 141  -83.5156092 -108.3112658 136.770564
#> 142  -87.2316954  -51.3679108 101.232559
#> 143 -101.2526713    0.2514156 101.252983
#> 144  -98.3488177  -16.6931105  99.755450
#> 145  -14.2695074  -61.7082964  63.336661
#> 146   39.1827547  -18.1963420  43.201795
#> 147   60.6956371   64.9723887  88.912157
#> 148   65.3125658   82.0454209 104.867451
#> 149   97.5537674  -45.0945332 107.472110
#> 150   94.9359090  -64.8636657 114.978789
#> 151  -44.2394993   41.8017741  60.864781
#> 152  -25.7730894   35.0604770  43.514241
#> 153   -7.2687099   44.2223329  44.815721
#> 154  -32.9577934  -55.6189978  64.650515
#> 155  -66.0686201  -21.0938484  69.354257
#> 156   18.3358543  -24.7997261  30.842016
#> 157   78.7552542  -59.4830738  98.694610
#> 158   20.1707591  -34.4544007  39.924494
#> 159  -94.5547077  -28.3122857  98.702474
#> 160 -105.6263348   82.9635344 134.312586
#> 161   46.0883823  203.5718792 208.723858
#> 162  130.5974531  176.3304780 219.426827
#> 163   68.3058070    9.0492915  68.902634
#> 164  -46.5187816 -190.6365789 196.230228
#> 165 -128.0604142 -219.5681369 254.184257
#> 166  -61.5381455  -57.3672166  84.130499
#> 167  -70.0843572   86.4478968 111.288166
#> 168 -122.2099972   60.4343112 136.336310
#> 169  -76.6866392  -21.1289648  79.544162
#> 170   37.5336359   12.2969304  39.496687
#> 171  221.3433583   80.1841237 235.419574
#> 172  173.4891814   52.4094110 181.232564
#> 173   50.3406380  -77.0782072  92.061012
#> 174  -65.6772072 -110.5701011 128.604987
#> 175  -55.0913892  -14.0677743  56.859154
#> 176   25.7914335   74.5272582  78.863872
#> 177   56.0069655   53.6674954  77.569197
#> 178   62.7910669  -18.0289836  65.328113
#> 179  -18.5444644  -61.0763961  63.829643
#> 180 -129.4154111  -19.9841167 130.949279
#> 181 -202.4145153  -15.1062693 202.977426
#> 182 -157.2465368   53.0661525 165.959302
#> 183  -22.7017888   96.0559775  98.702189
#> 184  106.9753483   17.2611584 108.358999
#> 185  229.0993161   -6.1389499 229.181551
#> 186  251.9144144  -20.3048723 252.731399
#> 187   73.1825580  -44.5443859  85.673153
#> 188 -147.4205079 -111.0952457 184.594040
#> 189 -330.1620166   41.5737830 332.769194
#> 190 -159.7256515  204.8696931 259.776586
#> 191  165.7186540  188.1511218 250.725980
#> 192  201.1223891  -26.6603963 202.881720
#> 193  -24.1532224 -159.4982306 161.316657
#> 194 -169.6058930  -74.8741627 185.397678
#> 195  -80.7216071  -50.4766893  95.204380
#> 196   30.6898662   42.5297559  52.446621
#> 197   52.0428971   46.4329956  69.745869
#> 198   51.3151897   28.1560179  58.532128
#> 199  132.4190494    0.8380872 132.421701
#> 200   48.9425031  -92.0083587 104.215674
#> 201   29.5547426 -148.2000789 151.118319
#> 202  -32.5107643  -71.4430754  78.492438
#> 203  -99.1616879   74.6726026 124.133146
#> 204  -49.0115997  117.4165256 127.235126
#> 205  -38.2968621   57.4981224  69.084613
#> 206    3.5182878  -54.4167424  54.530360
#> 207   20.4175940  -59.0962391  62.523944
#> 208   98.2188896  -38.6314111 105.543054
#> 209  155.0084878   20.1211155 156.308959
#> 210  149.4888151  142.5711453 206.575500
#> 211   40.5431055   94.7384196 103.049074
#> 212  -51.9048146  -27.3368364  58.663553
#> 213 -122.9597386  -66.7192173 139.894786
#> 214 -152.2590660  -64.7105133 165.439638
#> 215  -93.8997999   63.2749857 113.229397
#> 216   20.6456756   32.9210123  38.859194
#> 217  128.3377119  -26.4757548 131.040200
#> 218   32.0427629  -27.3091606  42.101412
#> 219  -94.2707164   -4.3400862  94.370569
#> 220  -78.0715589   -1.0952440  78.079241
#> 221   61.7464902 -105.8129939 122.511300
#> 222  111.2906223 -138.1473916 177.398716
#> 223   75.1130715   19.8134562  77.682344
#> 224  -16.2785573  125.4165960 126.468628
#> 225  -55.3620236   37.0088566  66.592861
#> 226  -31.8440763  -34.4556648  46.917353
#> 227 -110.7527427 -168.9372254 202.004842
#> 228  -81.1465329  -73.1381585 109.242620
#> 229  -85.5242578   13.0425791  86.513048
#> 230   84.6788281   74.0015518 112.457697
#> 231  161.0798000  102.7682540 191.070709
#> 232  -15.7267500  -14.8584572  21.635721
#> 233 -155.0487034  -28.7732845 157.695917
#> 234 -106.4342792  -19.7042065 108.242836
#> 235   68.6005204  -37.8107532  78.330610
#> 236  153.6847522   -1.0693863 153.688473
#> 237  -14.7505730   62.2964757  64.018984
#> 238 -103.5913624   85.4391279 134.279615
#> 239  -54.2378713   34.3612997  64.206274
#> 240   43.7962390  -80.3949884  91.550340
#> 241  187.5370291  -13.5247808 188.024086
#> 242  123.7305511   86.1413063 150.763304
#> 243   85.3224234  132.1914842 157.335643
#> 244  -79.1310189   84.0154957 115.413698
#> 245 -130.7304918    7.4197862 130.940883
#> 246  -90.1099274  -50.5912889 103.340590
#> 247   36.1477139  -42.9940265  56.170664
#> 248    4.2067432  -68.1767141  68.306376
#> 249  -28.6668977  -90.1402046  94.588834
#> 250  -18.0846166   50.8022906  53.925190
#> 251    7.4749441  126.9945051 127.214304
#> 252   15.2191926   76.0266813  77.535025
#> 253  -19.6125791 -106.9827407 108.765620
#> 254   62.1499888   31.2559681  69.566922
#> 255   19.7817595  160.9673825 162.178347
#> 256   28.9943315  -24.6707925  38.069926
#> 257  -12.0590337 -161.9097701 162.358227
#> 258  -28.7215281 -159.4737153 162.039477
#> 259   14.9674842   33.1459669  36.368678
#> 260   -9.1934112   77.4835693  78.027061
#> 261  -46.5055321   -0.7535993  46.511638
#> 262  -44.5592449  -42.9229983  61.870107
#> 263   36.8053899 -151.1112014 155.528878
#> 264   98.2187824  -91.5637767 134.279017
#> 265  -16.5541596   78.1050769  79.840110
#> 266 -105.5432989   81.3181941 133.236769
#> 267  -75.5323549  -38.0124361  84.558157
#> 268  -38.5073143  -45.7022174  59.762078
#> 269   32.6070147   81.1839116  87.487399
#> 270   17.0683610  179.4154856 180.225540
#> 271   -7.4244281   47.9946782  48.565536
#> 272   22.5343018  -81.3889258  84.450885
#> 273   28.5134992  -46.3891405  54.451556
#> 274  105.8674111  -55.8552430 119.698442
#> 275  130.2525941  -49.7606665 139.434078
#> 276   95.2493184   -9.2953463  95.701808
#> 277   -5.8688854    0.5455414   5.894186
#> 278 -101.7403359  -25.3620726 104.853854
#> 279  -50.1793520  -60.2872494  78.438000
#> 280   36.8329358   19.6391542  41.741604
#> 281   -3.0500265   90.9655719  91.016690
#> 282  -53.3674831   96.4673590 110.245361
#> 283  -17.1383995   27.9612579  32.795681
#> 284   34.0944185 -156.6716995 160.338551
#> 285   -8.1056889 -190.5020687 190.674436
#> 286  -78.3200087  -89.6207881 119.020626
#> 287   -0.2492797  119.9421439 119.942403
#> 288  162.3155546  173.4588598 237.559077
#> 289   92.3540900   97.5776300 134.352789
#> 290  -93.0474104  -12.2079882  93.844848
#> 291 -147.4181126 -116.9540870 188.176402
#> 292  -38.3275966  -63.7286095  74.366258
#> 293  174.0973951   -5.0706472 174.171222
#> 294  121.6145343   73.3689432 142.032027
#> 295   -2.2784237   63.8433356  63.883979
#> 296  -53.4499366   50.9557186  73.847011
#> 297   -1.3198171  -34.3820366  34.407359
#> 298  -31.9101233  -24.2901004  40.103179
#> 299 -148.5451104   45.2217904 155.276077
#> 300 -133.7399429   19.5000693 135.154079
#> 301   36.5512726  -33.1773506  49.363267
#> 302  146.9381272   -1.5943082 146.946776
#> 303   51.9801104  125.9828657 136.285048
#> 304    2.3435697  112.3573275 112.381766
#> 305  -50.5369655  -40.6330667  64.846210
#> 306  -85.0357202 -116.3896066 144.144421
#> 307  -28.3795583  -58.8247992  65.312758
#> 308   31.2770789  -28.0132212  41.988049
#> 309   -0.7520794  -46.4738727  46.479958
#> 310  -60.5261920   28.5460968  66.920098
#> 311  -59.3755634  113.7554602 128.318986
#> 312  -51.3939062  114.8430833 125.818390
#> 313   19.9111574    2.0143545  20.012791
#> 314  -37.7419338 -130.6086353 135.952452
#> 315  -24.6629167 -107.3722504 110.168324
#> 316   77.5631979  -96.8528396 124.082723
#> 317  139.3021686   56.1908754 150.208218
#> 318  223.8831373  192.9409903 295.550139
#> 319   80.7749496  133.1013380 155.693798
#> 320    4.1157608   16.2741903  16.786565
#> 321  -57.9857868  -82.5480518 100.878800
#> 322 -141.6620982  -95.7499556 170.985976
#> 323  -11.3762068  -19.8984455  22.920869
#> 324    3.5929192   28.5051431  28.730685
#> 325   97.9375760   46.9004963 108.588330
#> 326   15.9641115   24.0744788  28.886561
#> 327  -42.1759534  -51.6003733  66.643901
#> 328   -8.9980666    5.6015396  10.599172
#> 329   -3.2901348   58.0302971  58.123492
#> 330  -66.3103769  -43.5539712  79.334825
#> 331 -197.2886077  -71.3817618 209.805030
#> 332 -157.7088598  -26.5908706 159.934858
#> 333  104.3577973   43.2382589 112.960599
#>  [ reached 'max' / getOption("max.print") -- omitted 667 rows ]