Compute velocity for x, y, and its amplited

Compute velocity for x, y, and its amplited

compute_velocity_table(x, y, trial, sample_rate, velocity_time_window)

Arguments

x

vector with x coordinates in degrees of visual angle

y

vector with y coordinates in degrees of visual angle

trial

vector with trial index

sample_rate

sample rate in Hz

velocity_time_window

Time window for velocity computation in milliseconds

Value

data.frame with columns x, y, and amp

Examples

compute_velocity_table(rnorm(1000), rnorm(1000), rep(1, 1000), 250, 20)
#>                x            y        amp
#> 1             NA           NA         NA
#> 2    139.7225343  120.8610775 184.742487
#> 3    -13.7440946  197.7031412 198.180302
#> 4    -35.9776961   97.4408082 103.870620
#> 5    -41.4653342  -32.7649014  52.848015
#> 6     20.8256851  -64.5109906  67.789211
#> 7     11.3208529   32.8318350  34.728822
#> 8     21.8024997   -3.6694659  22.109138
#> 9     73.4512253    1.4958526  73.466455
#> 10    27.3605781  -41.9977403  50.123961
#> 11    23.5499315  -52.3477639  57.401112
#> 12    58.1036453   -2.3283791  58.150279
#> 13    64.5451242    4.0453324  64.671770
#> 14   -52.7139053    8.7627290  53.437264
#> 15  -130.5932789    5.0480986 130.690810
#> 16  -115.4874255   24.4373553 118.044609
#> 17   -59.7470795  -10.5629413  60.673629
#> 18  -134.1418414   -8.6412249 134.419881
#> 19   -47.0726933   -2.6471000  47.147063
#> 20    96.3868093   10.8761166  96.998489
#> 21   225.6734948   84.2702878 240.894184
#> 22   194.2567411  126.0173111 231.551386
#> 23    -3.8308178  129.5514592 129.608085
#> 24  -107.0018898   -4.6748447 107.103962
#> 25  -119.1576212  -60.2240110 133.512060
#> 26     9.6454663  -51.6802416  52.572639
#> 27    15.6617685   15.3729989  21.945844
#> 28    46.5311901   23.5580353  52.154891
#> 29    14.1576738  -42.7866972  45.068184
#> 30   -40.3656186 -109.3855093 116.595767
#> 31   -43.2496415 -117.4947212 125.202001
#> 32    14.8819088  -38.7908515  41.547580
#> 33    74.5458994   10.5710227  75.291684
#> 34    14.5938010   30.0686580  33.423094
#> 35   -37.3279573   10.4838733  38.772258
#> 36   -13.1669552  -10.9415277  17.119747
#> 37    30.6235524    7.8971876  31.625425
#> 38    58.0277090   73.2553744  93.453544
#> 39   114.1746531   31.3484691 118.400076
#> 40     3.0560860    5.8531910   6.602992
#> 41   -84.8989479  -15.1607554  86.241984
#> 42  -100.5526375    9.7042421 101.019826
#> 43   -71.7627450   27.7752402  76.950345
#> 44    38.8761482   63.2153148  74.212741
#> 45   109.3648067   63.9690801 126.699267
#> 46   117.8884496    5.8208393 118.032066
#> 47   -76.7980311  -72.5889398 105.674461
#> 48  -229.4460714 -103.0693244 251.532871
#> 49  -205.2682422  -95.5444765 226.415102
#> 50   -58.7716396  -78.2187674  97.838035
#> 51    97.6558518  127.5588462 160.648450
#> 52    72.8458209  217.3271955 229.210871
#> 53    78.9967470   27.6255431  83.687853
#> 54    58.8125631 -208.9803399 217.098365
#> 55     5.5951712 -257.9210179 257.981700
#> 56   -60.0771303  -21.8826853  63.938357
#> 57  -117.4500376  109.4261597 160.525997
#> 58    38.2988599  120.2241669 126.177070
#> 59   123.5764314   76.1430952 145.151319
#> 60   124.4630505  -50.8734977 134.458781
#> 61    76.8737339   -7.6352866  77.251981
#> 62    18.0570208   31.4950021  36.304148
#> 63   -25.0727756   58.6212421  63.758090
#> 64   -27.3204443   -8.0262652  28.475035
#> 65    27.4239478  -63.9102553  69.545623
#> 66    -0.6634179  -50.9680306  50.972348
#> 67   -94.1767287   -5.3866698  94.330655
#> 68  -102.2148521   67.8655052 122.693124
#> 69     4.2439432   58.1278250  58.282545
#> 70   -11.3703558  -47.0535453  48.407862
#> 71    -5.6120498  -63.0627766  63.311996
#> 72    70.8470984   63.7893820  95.333082
#> 73   160.7450928   96.4444967 187.458064
#> 74   105.0193003   54.4921900 118.315055
#> 75   -50.6239032   -4.4138512  50.815959
#> 76  -102.9989412  -53.3005839 115.972989
#> 77  -116.2197083  -35.5156856 121.525242
#> 78   -80.5070741  -10.2723389  81.159780
#> 79     2.6665299   15.2745270  15.505533
#> 80    25.0495993   37.0523305  44.725358
#> 81    32.0670155  -26.9313378  41.875893
#> 82   -40.7233682   -3.7803795  40.898459
#> 83   -83.6160275    4.9699523  83.763599
#> 84   -27.6519207   45.2889697  53.063354
#> 85    43.0773493   60.5999425  74.350595
#> 86   139.4429989    4.0232173 139.501026
#> 87    66.2536715  -71.7429413  97.655510
#> 88    41.7739646  -65.8246571  77.961206
#> 89   -18.3992583  -42.5257033  46.335388
#> 90  -138.3459159  -70.2648455 155.166817
#> 91   -84.0597642  -47.3396327  96.473234
#> 92    30.7321024  -23.6397259  38.772397
#> 93   175.5244121   48.2604599 182.038159
#> 94    94.7485856   22.5475801  97.394496
#> 95   -86.4949716   35.2540535  93.403578
#> 96  -100.9916754   47.9974395 111.817140
#> 97     0.9059687  -42.8332186  42.842799
#> 98   -14.0884555 -102.9817751 103.940996
#> 99   -44.4377707  -15.6704000  47.119814
#> 100  -63.0317772   34.2567262  71.739307
#> 101   35.1325270   31.5699252  47.232982
#> 102   89.2955852   55.7322795 105.260574
#> 103   52.7721937   39.7406937  66.062298
#> 104    5.2496139   29.4418581  29.906211
#> 105  -90.5558062 -164.0321995 187.368398
#> 106  -57.6121405  -83.3286089 101.305557
#> 107   26.4293897  123.8542486 126.642756
#> 108   96.8955548  139.3096733 169.693646
#> 109   -3.1712570   83.4563069  83.516538
#> 110  -81.5552863 -185.4861561 202.623737
#> 111  -26.1796413 -193.2351572 195.000512
#> 112   80.2497211   24.4233929  83.883967
#> 113   47.9433880  135.6934993 143.914190
#> 114  -14.9006892  156.2964207 157.005101
#> 115  -13.6314145  -20.1167127  24.300156
#> 116   78.5790774  -67.6383420 103.680358
#> 117   40.5603743  -21.2141224  45.773168
#> 118  -91.0431139 -176.5876596 198.675741
#> 119  -54.2433754 -187.2376187 194.936579
#> 120   20.0213999    8.5624872  21.775506
#> 121   36.1891460  199.9413552 203.190058
#> 122  -64.5167226  180.7827338 191.950005
#> 123 -111.3758574   18.7406066 112.941542
#> 124  -27.2618555 -135.8755815 138.583485
#> 125   11.3601920  -92.6885398  93.382115
#> 126   -0.5044839   22.3412686  22.346964
#> 127   26.6528078  118.0777639 121.048463
#> 128   71.0859342   91.3920891 115.783090
#> 129   15.9361237  -37.4328193  40.683854
#> 130  -35.6046852   29.9511723  46.527049
#> 131   46.1256897   93.3313958 104.107294
#> 132  165.2748769   37.1004067 169.387795
#> 133   96.6027271  -99.1351000 138.419128
#> 134  -81.5918041 -138.4147890 160.673197
#> 135 -155.9590584  -42.1230541 161.547453
#> 136 -156.6751486    7.2266968 156.841727
#> 137  -56.2438910    7.2134050  56.704572
#> 138   48.2260233    2.7158968  48.302437
#> 139   88.6327120    5.1506093  88.782242
#> 140   48.1570160   33.3851051  58.597470
#> 141    1.6076194   63.5372842  63.557619
#> 142  -60.4197974   22.1234608  64.342827
#> 143  -40.3989227   49.0288681  63.528756
#> 144   11.0323428   13.1734596  17.182917
#> 145   41.8299731  -71.6909959  83.002082
#> 146   12.5859205  -85.9375291  86.854270
#> 147  -90.0117151  -48.8522843 102.414133
#> 148  -49.9594555   88.5225328 101.647361
#> 149  -12.9066792   83.3959572  84.388791
#> 150   48.2633010  -85.7412311  98.391590
#> 151   83.0324484 -175.4343296 194.091709
#> 152   -7.8357985 -121.4604268 121.712920
#> 153  -10.9389125   64.6093681  65.528851
#> 154   65.9654333  187.3916371 198.663193
#> 155  105.6571693  108.8178148 151.673182
#> 156  111.1114612    6.6416356 111.309784
#> 157   -6.1610980   38.5479775  39.037235
#> 158  -47.1825596   85.7386681  97.863748
#> 159  -76.0881646   67.0465771 101.413275
#> 160  -44.1512710  -18.5158847  47.876641
#> 161   34.8566196  -16.3228709  38.489220
#> 162   22.8150088  -80.5032346  83.673744
#> 163  -30.6434554 -130.1148721 133.674610
#> 164   -4.9705058  -82.2813206  82.431315
#> 165   52.9058670   -3.5010638  53.021583
#> 166   64.1728833  107.7289771 125.394144
#> 167  108.8835786   18.6069457 110.461994
#> 168  -14.2718292   37.3830569  40.014723
#> 169  -84.0695692    4.7250315  84.202247
#> 170 -111.8206471    4.8907697 111.927551
#> 171  -56.5287218  -29.2763832  63.660058
#> 172   17.8607192  -20.6551538  27.306422
#> 173   38.9598952   37.2592605  53.908496
#> 174   28.3140314   -4.6483803  28.693062
#> 175  -12.9715178  -93.4151800  94.311485
#> 176 -128.0526905  -97.9756903 161.235007
#> 177   15.8157534   64.1720786  66.092312
#> 178  165.8522274  118.7057187 203.955900
#> 179   49.2417013   29.9425147  57.630715
#> 180  -87.6651597   -1.8876341  87.685480
#> 181 -107.2585573   71.7810336 129.061671
#> 182   35.0725176   61.3335664  70.653293
#> 183  -26.0729862 -106.3046645 109.455390
#> 184  -46.0279381 -281.8113938 285.545500
#> 185  -43.4112694 -207.3278806 211.823956
#> 186  -11.6548264   66.6381562  67.649677
#> 187   55.8064078  250.2993920 256.445200
#> 188   -7.8917878  156.5804567 156.779207
#> 189  -35.1986087  -76.9490593  84.617373
#> 190  -35.5173172 -111.5782549 117.094777
#> 191  105.2433839  -28.5390961 109.044256
#> 192    9.9255181  -21.5638576  23.738489
#> 193  -95.6474336  -37.6676119 102.797279
#> 194  -24.9079042  -25.2692201  35.481505
#> 195  158.5463073   87.3732261 181.027656
#> 196   59.8972832  113.9603843 128.742587
#> 197 -173.7617637   83.9570555 192.981703
#> 198  -92.0177293   50.7657921 105.092474
#> 199   -9.0303172   31.7078652  32.968703
#> 200   67.8376053    8.0911603  68.318428
#> 201   34.8058663   -7.1124291  35.525132
#> 202   35.2595545   44.5456402  56.811533
#> 203   59.2372909  -80.4427557  99.900418
#> 204   41.9700589 -140.5129321 146.647093
#> 205   25.1375211  -40.9399168  48.041355
#> 206   57.7545360  112.9538308 126.862738
#> 207  -25.3401570  112.8439893 115.654181
#> 208  -62.9959647 -123.7583150 138.869047
#> 209  -80.9160712 -111.4575158 137.732307
#> 210  -30.0500937    4.6729066  30.411251
#> 211   25.9049706   44.0064168  51.064980
#> 212  -40.8593760   18.9890810  45.056340
#> 213  -58.8163629   19.2210052  61.877392
#> 214   -6.8914256   33.6179258  34.317003
#> 215   42.3881110   51.8611814  66.980102
#> 216   44.3565494    2.0391596  44.403397
#> 217   49.9746249  -21.6858112  54.476945
#> 218   71.1820119   28.6477752  76.730527
#> 219   59.0655948    0.1390923  59.065759
#> 220    5.9152680  -54.9826915  55.299971
#> 221  -33.2415970 -161.7452886 165.125837
#> 222  -15.8918946 -128.7022719 129.679710
#> 223   32.0696108   12.4596083  34.404967
#> 224   28.6286198   22.6953052  36.533201
#> 225  -11.1602792   37.7367756  39.352459
#> 226  -73.7379286   54.6324397  91.771377
#> 227  -35.0478934   93.4249949  99.782686
#> 228  -15.7709493   23.5720896  28.361351
#> 229  -77.1058543    9.3213462  77.667241
#> 230  -90.7740982   27.1979568  94.761098
#> 231   40.7135061   39.5207551  56.740459
#> 232  157.2429524   -7.8372390 157.438141
#> 233   78.2686117  -87.9977636 117.769189
#> 234  -81.2175118  -75.8194470 111.107483
#> 235 -117.0422524  -58.3714490 130.790347
#> 236    0.7846896  108.3769391 108.379780
#> 237  113.6193663  158.8813152 195.326989
#> 238  148.2615756   31.7878696 151.631011
#> 239   96.0425229  -89.6629485 131.391060
#> 240  -18.1314261 -119.0072586 120.380547
#> 241 -139.5523834    2.4459337 139.573817
#> 242 -105.3693285   88.0960616 137.344863
#> 243   37.2714347   23.3237766  43.967697
#> 244  164.8829863  -37.8273524 169.166509
#> 245   61.5010604 -137.2921356 150.437731
#> 246 -120.5225676  -92.2525914 151.776908
#> 247 -182.3649589   44.9488805 187.822736
#> 248 -120.2163686  117.4911449 168.095641
#> 249   21.3910613  104.9994718 107.156272
#> 250   18.0623890   25.1190933  30.938952
#> 251   66.5092611  -37.1143094  76.163993
#> 252  106.2318673  -26.3298872 109.446209
#> 253   96.8803854  -25.3604325 100.144698
#> 254    9.3812302    5.6923205  10.973149
#> 255  -48.8803233  -43.3991647  65.366456
#> 256  -25.9232171 -104.1303467 107.308631
#> 257  -57.8167582  -34.1470338  67.147580
#> 258  -43.7090294  107.9911481 116.501362
#> 259  -24.7307815  224.1168003 225.477165
#> 260   53.0308778   84.5883978  99.837223
#> 261    9.7430422 -116.5460437 116.952585
#> 262  -37.5102258 -125.5760778 131.058645
#> 263  -66.6607861  -14.2610569  68.169188
#> 264  -42.9251439   50.2630418  66.097968
#> 265   59.8078564  -54.6164707  80.993448
#> 266   76.9857857  -96.3672888 123.342878
#> 267   28.8507240   41.3171174  50.393139
#> 268  -27.5909302   99.5465210 103.299416
#> 269   19.3308615    9.6922839  21.624583
#> 270  -38.2252953  -89.9474414  97.732878
#> 271 -111.6561681 -112.4469727 158.465837
#> 272  -29.2970511   11.8092844  31.587599
#> 273   67.9687156   84.4070805 108.371129
#> 274   57.4555094   84.3197696 102.034108
#> 275  -49.0245517   89.1608091 101.749971
#> 276  -75.5629384   28.3353277  80.700982
#> 277   71.7557215   -8.6299410  72.272813
#> 278   39.0710703  -53.5850272  66.316692
#> 279  -25.6188485  -34.0684798  42.626127
#> 280  -12.2277152    1.3428603  12.301231
#> 281   -0.0784622   58.6201692  58.620222
#> 282  -12.1864849  -31.4750743  33.751899
#> 283  -97.4761339  -54.1299428 111.497298
#> 284  -14.7171938  -36.4704849  39.328006
#> 285   90.5261429   19.9527304  92.698943
#> 286  118.7953045   54.7204410 130.792397
#> 287   48.0270183  -31.4505992  57.408490
#> 288   52.6797071  -42.4127426  67.631297
#> 289   73.2057436  -70.2708159 101.474472
#> 290   79.2328758  -67.5045345 104.089917
#> 291   -2.7803148  -25.7124973  25.862379
#> 292  -82.7515317   49.5299941  96.441880
#> 293  -98.2386958   14.2936718  99.273110
#> 294  -56.0903395  -15.3399059  58.150141
#> 295   27.7412407   28.5297878  39.793533
#> 296  -49.0193959   29.2445295  57.080151
#> 297 -179.0177807  -43.3300829 184.187030
#> 298 -215.0665106  -69.0777743 225.887899
#> 299    8.8496791   96.7055451  97.109625
#> 300  136.3445409  150.9295788 203.395112
#> 301   77.0713283   59.5355525  97.388252
#> 302   64.4816602  -28.2877160  70.413631
#> 303   68.2032500    5.7628353  68.446282
#> 304   39.4729149   89.4010655  97.727486
#> 305  -49.2572134   84.3416544  97.671837
#> 306   30.3249332   24.2497174  38.828474
#> 307   86.5584406 -136.6870093 161.789067
#> 308  -33.1742338 -141.1704409 145.015941
#> 309  -39.6793432   -3.3380858  39.819506
#> 310  -16.8562418   36.5239205  40.225982
#> 311  -37.0362149   -7.6390883  37.815829
#> 312 -128.6073316  -35.9019410 133.524511
#> 313 -108.7297092   59.3856816 123.890310
#> 314  144.6700680    6.3465591 144.809210
#> 315  173.7243518  -45.7923359 179.658255
#> 316   54.7036424   57.7553406  79.549782
#> 317 -129.4679233   79.2357716 151.790153
#> 318 -101.2727212   57.2060093 116.312904
#> 319   84.9405023  -93.4322667 126.271443
#> 320  126.8629818  -90.6937114 155.947316
#> 321   58.7093080 -116.1506762 130.145159
#> 322 -140.0536187  -12.3987234 140.601367
#> 323  -48.0443930  159.1351217 166.229512
#> 324   77.2584024  141.5151246 161.230863
#> 325  121.0627943   68.9031277 139.297671
#> 326   16.7456044 -108.0519051 109.341801
#> 327 -133.3260823  -84.3397272 157.762587
#> 328  -89.3595310  -75.5015456 116.985508
#> 329  -59.5088795  -16.6107725  61.783691
#> 330   49.0385041  102.0293724 113.202331
#> 331  145.4151856   77.9397332 164.985388
#> 332   92.7487405  -28.3677849  96.990000
#> 333  -17.1049981 -135.0339838 136.113033
#>  [ reached 'max' / getOption("max.print") -- omitted 667 rows ]