Differentiate x and y and compute change amplitude via an algorithm proposed by Engbert & Kliegl (2003)

Differentiate x and y and compute change amplitude via an algorithm proposed by Engbert & Kliegl (2003)

diff_ek(x, y, trial, sample_rate, options = NULL)

Arguments

x

vector with x coordinates in degrees of visual angle

y

vector with y coordinates in degrees of visual angle

trial

vector with trial index

sample_rate

sample rate in Hz

options

List with method specific options, see Details.

Value

data.frame with columns x, y, and amp

Details

Method options, please refer to Engbert & Kliegl (2003) for details on parameters and the rationale for default values.

ek_velocity_time_window Time window for velocity computation in milliseconds. Defaults to 20 ms.

See also

Examples

diff_ek(rnorm(1000), rnorm(1000), rep(1, 1000), 250, list("ek_velocity_time_window" = 20))
#>                 x            y        amp
#> 1              NA           NA         NA
#> 2    206.22459412   12.8718166 206.625910
#> 3    111.03486633  -91.0278244 143.578572
#> 4     47.41121674   -3.0268154  47.507737
#> 5    -91.99497986  122.7524872 153.398988
#> 6   -139.25993347  162.4580841 213.976536
#> 7   -180.35113525   -4.1619587 180.399152
#> 8   -163.79914856  -84.1547470 184.152607
#> 9    -46.75622559  -39.0584946  60.923810
#> 10    83.24595642   12.7736254  84.220275
#> 11   172.61544800   24.8539791 174.395565
#> 12   173.88792419  -43.2366524 179.182639
#> 13   153.88623047  -75.6642456 171.481923
#> 14    81.17166138  -77.3337860 112.113126
#> 15   -31.37047768   76.1741867  82.380905
#> 16  -183.87228394  228.7881622 293.518381
#> 17  -163.62644958  143.3123322 217.513309
#> 18   -79.23162842   -1.8327330  79.252822
#> 19   -45.08788681  -97.2566299 107.199672
#> 20    79.08464050 -107.0110779 133.062959
#> 21    44.35026169 -117.5792389 125.665521
#> 22   107.56414032  -57.3549271 121.900090
#> 23    90.88690186   38.9308357  98.873853
#> 24    44.77977371   35.0908470  56.891086
#> 25    46.01854706  -16.5171680  48.892980
#> 26   -39.23682022  -16.4070644  42.529047
#> 27   -31.06540108    7.5029635  31.958623
#> 28   -73.45840454    4.0668755  73.570896
#> 29  -112.84643555  -40.1205521 119.766342
#> 30  -132.64837646  -75.3122406 152.536964
#> 31   -62.05609512  -80.6861954 101.790083
#> 32     9.75903702    6.0885935  11.502599
#> 33    -7.53868818  142.6734009 142.872430
#> 34     3.57710361  144.6006775 144.644916
#> 35    55.58576965  200.7726440 208.325304
#> 36   138.15878296  130.6191406 190.129454
#> 37    70.01396942   27.9207554  75.375888
#> 38   -55.10711288 -141.9419861 152.263986
#> 39  -102.05544281 -137.5588531 171.282665
#> 40   -45.18598938   42.6710548  62.149759
#> 41    72.30063629   44.4216232  84.856718
#> 42    24.55402946  -65.1403580  69.614414
#> 43   -38.68341827 -124.1573944 130.044090
#> 44   -61.46985245  -37.2184677  71.859287
#> 45   -35.00431824   23.4184036  42.115602
#> 46    66.71572113   32.5829468  74.247127
#> 47   137.41920471  -66.7570114 152.776099
#> 48   135.06347656 -121.3813400 181.591774
#> 49   -54.76870728 -104.5615845 118.037012
#> 50  -100.89691162   15.1142406 102.022679
#> 51    31.29792213  115.9978790 120.146027
#> 52    76.60443878  105.4715805 130.355262
#> 53    43.20733643   57.7126198  72.094524
#> 54  -130.58453369   11.8927870 131.124974
#> 55   -65.76210785  -53.7088470  84.907568
#> 56     0.88652915  -92.7180252  92.722263
#> 57    21.38345146  -34.9982796  41.013797
#> 58    72.83087158   76.4375839 105.579544
#> 59   106.37687683   13.8394794 107.273348
#> 60    94.54701233 -142.5741882 171.074652
#> 61  -143.23213196 -177.6997833 228.238158
#> 62  -236.08116150  -32.3917236 238.292968
#> 63  -127.61998749  195.9162598 233.816257
#> 64    99.59565735  201.7259369 224.972551
#> 65   107.92274475  119.1421509 160.755003
#> 66    36.59952927    5.4880052  37.008698
#> 67    46.04984665   36.4191170  58.710650
#> 68    53.64509201   23.7699738  58.675442
#> 69    49.03499222  -95.3389587 107.209829
#> 70   -48.35859299 -155.1441040 162.506143
#> 71  -109.96031952 -122.0771866 164.298848
#> 72   -23.15781975   72.4147720  76.027520
#> 73     8.52486324  110.2020264 110.531262
#> 74    -0.06820434   71.8186417  71.818674
#> 75     1.52737057   17.3009186  17.368208
#> 76   -11.68612099   32.1676445  34.224593
#> 77    50.91797638  127.4309540 137.227141
#> 78   -64.83138275   40.4552994  76.418188
#> 79  -123.45771790  -95.1897202 155.893845
#> 80   -49.19681931 -138.7257385 147.190888
#> 81    37.42621613  -45.9337196  59.250555
#> 82    50.54133224  -35.9834938  62.042228
#> 83   -57.57168961  -71.9158478  92.121597
#> 84   -62.77228928  -45.1496506  77.323032
#> 85    62.05154037   84.1085968 104.521049
#> 86   144.84600830  160.0536041 215.864593
#> 87   165.56858826   18.1191158 166.557077
#> 88    36.64540100  -66.5879059  76.005491
#> 89  -157.88536072  -86.9460373 180.242616
#> 90  -167.25914001  -26.5959625 169.360459
#> 91   -11.84866714  -21.4499416  24.504916
#> 92   182.96803284   -6.0683565 183.068637
#> 93   116.13191986   80.8315201 141.493312
#> 94   -26.87939072   44.9960022  52.413184
#> 95   -61.41704178  -87.8527985 107.192198
#> 96    20.47767067  -39.9902916  44.928370
#> 97    76.22109985    3.5124483  76.301988
#> 98    -2.02812839   32.5509567  32.614078
#> 99   -50.29724503   47.2852402  69.034099
#> 100  -59.20650864  127.9579315 140.991641
#> 101 -112.99702454  149.9573517 187.764573
#> 102 -100.41769409  -85.7119827 132.023700
#> 103   20.76787186 -142.5003204 144.005715
#> 104  168.93592834 -103.4808655 198.110165
#> 105   81.41693115  -85.2240448 117.863712
#> 106  -58.14007187  -39.6679726  70.383350
#> 107 -150.81306458  100.0896912 181.004217
#> 108 -121.31398010  206.5267181 239.521120
#> 109  -11.48051739  115.1706696 115.741459
#> 110   91.88909149  -31.4050560  97.107583
#> 111  193.27105713  -85.6652985 211.405404
#> 112   55.38768387  -45.5022583  71.681595
#> 113  -85.10713959   -8.2362528  85.504743
#> 114  -92.53221893   60.8257942 110.733865
#> 115    7.21958780  -14.4609270  16.162947
#> 116   82.52313232 -144.2127075 166.154664
#> 117   12.57460022  -87.4785309  88.377678
#> 118  -42.50807953   60.7599411  74.153269
#> 119   -0.37385887   79.8515701  79.852445
#> 120    0.97950083   -2.1137364   2.329657
#> 121    2.09187770  -92.8070602  92.830633
#> 122  -49.30237579  -27.7271252  56.564280
#> 123  -28.54035950   58.1372108  64.764862
#> 124   27.70608330  107.5832672 111.093593
#> 125   36.08214951  121.9697418 127.194888
#> 126   15.69571209  -56.5240288  58.662775
#> 127   -6.46999359 -141.8335266 141.981020
#> 128   -4.38169861  -62.8470230  62.999584
#> 129   42.80669403   73.1379471  84.744158
#> 130   38.70165253   98.7698898 106.081615
#> 131   91.21661377   -0.1748058  91.216781
#> 132   29.97960854  -64.9935913  71.574743
#> 133  -43.66521072  -55.2034378  70.385156
#> 134  -26.82619286   23.1311398  35.421664
#> 135  -91.49355316   36.9113541  98.658595
#> 136  -90.43454742  -15.0335875  91.675603
#> 137 -133.64212036  -21.8506775 135.416648
#> 138   20.79239273   44.1575279  48.807898
#> 139  183.90997314  132.5169525 226.679556
#> 140   99.01336670   24.0488243 101.892064
#> 141   25.23727608  -85.2405624  88.898108
#> 142  -99.27073669 -132.0217743 165.179987
#> 143  -99.68641663  -39.0171165 107.050068
#> 144  -60.07272339   52.5697250  79.826738
#> 145   38.98060226   15.6154804  41.992030
#> 146   91.35984039    6.4054437  91.584115
#> 147   45.40045547   17.9908886  48.835166
#> 148   50.36641693   -9.8506651  51.320674
#> 149   55.45009232  -93.1275940 108.385707
#> 150  -13.23538780 -122.3663940 123.080095
#> 151 -102.98641205  -21.7449493 105.257037
#> 152   -0.17676751   37.1940765  37.194497
#> 153   63.39302444   63.7656898  89.915175
#> 154   19.34531593   60.0799065  63.117639
#> 155  -33.84683609   26.0340042  42.701027
#> 156  -53.93556595  -61.8897324  82.093753
#> 157  -36.89813995  -48.2131386  60.712268
#> 158  -36.42470169   48.4574776  60.620838
#> 159   -7.72526121   -1.1165564   7.805534
#> 160  -27.12013626  -54.3755646  60.763507
#> 161  -90.87815094  -44.9111824 101.369880
#> 162   80.71877289  130.1827850 153.176623
#> 163   87.32203674  136.3785858 161.939053
#> 164   31.74551392    6.5698323  32.418210
#> 165   28.57450676  -71.5023651  77.000589
#> 166   44.34243774  -78.9313583  90.534033
#> 167   38.96333694  -11.3125029  40.572335
#> 168 -114.32891083  -38.7304840 120.711020
#> 169  -19.55830956  -34.2559280  39.446116
#> 170   72.99813843   17.5013466  75.066806
#> 171    9.81861496   53.6565666  54.547524
#> 172  -77.96497345   49.6568489  92.435598
#> 173  -49.51640320    2.7350693  49.591882
#> 174   52.68787384   -9.6668482  53.567341
#> 175   68.56730652  -43.4163513  81.156978
#> 176    3.39898968    3.3886831   4.799615
#> 177    3.69998837   19.3461132  19.696751
#> 178    8.47192860   38.9041328  39.815890
#> 179   45.09163284  -52.1033020  68.905801
#> 180   12.43366623  -91.8507233  92.688464
#> 181  -31.04553604   59.8003578  67.378840
#> 182   39.42374039  186.3242188 190.449326
#> 183  -46.67513657  130.1882172 138.302351
#> 184  -94.44726562 -137.8864441 167.131557
#> 185 -106.64472198 -146.7493744 181.406934
#> 186   -2.65941358  110.0873337 110.119451
#> 187   67.39428711  235.8058319 245.247590
#> 188   -1.86810708   20.5470848  20.631833
#> 189   28.40270805 -153.2002106 155.810842
#> 190  162.94355774 -148.8690796 220.709324
#> 191  191.40499878  -51.3340645 198.169270
#> 192   61.56593323  -43.6228561  75.454077
#> 193  -60.29186630   11.2593050  61.334176
#> 194  -58.50002289  157.4417114 167.958760
#> 195  -34.77737808   67.0980225  75.575199
#> 196 -162.77545166  -43.2605743 168.426022
#> 197  -59.93019485  -77.6690063  98.102512
#> 198   53.54013824  -33.8026543  63.317974
#> 199   67.59449768   65.7775650  94.317041
#> 200    4.71892929   78.3208466  78.462878
#> 201  -48.27988434   70.0046158  85.038776
#> 202   68.40806580  -10.2429771  69.170673
#> 203   16.13672066 -117.1094666 118.215993
#> 204  -37.19120026 -113.3090515 119.256558
#> 205  -65.11726379   39.8979530  76.368218
#> 206  -31.55344009  115.6251984 119.853269
#> 207   36.55683136    1.0420301  36.571680
#> 208    5.86375904 -125.2803574 125.417509
#> 209  -43.54943848  -73.8892288  85.768128
#> 210  -96.05365753  174.0012207 198.752937
#> 211  -18.51300812  121.8210831 123.219754
#> 212  -25.63027954 -193.0499268 194.743897
#> 213    3.78554058 -279.3912964 279.416941
#> 214  136.31671143  -81.3893356 158.765455
#> 215  107.70669556   96.8229828 144.828941
#> 216    6.89560890   54.9337502  55.364847
#> 217 -114.96916199   -6.1467576 115.133361
#> 218  -20.91731071   55.2800407  59.105133
#> 219   -3.09588194   50.4362297  50.531156
#> 220 -100.03049469  -28.2770138 103.950418
#> 221 -141.33593750    6.3775945 141.479755
#> 222  -40.55461502   48.4033852  63.147165
#> 223  135.32658386  -25.7591076 137.756364
#> 224  149.92132568 -114.4378204 188.606518
#> 225   -2.22024727   -2.2707458   3.175812
#> 226 -105.98354340  116.3223877 157.363939
#> 227    8.00001335   75.5558090  75.978158
#> 228  179.79273987   45.5171204 185.464923
#> 229  144.63328552   31.2983360 147.980989
#> 230  -60.67902756   81.6010590 101.689120
#> 231  -78.43732452   23.0335064  81.749350
#> 232  -10.17731762  -14.2252083  17.490979
#> 233   45.43110657  -16.2796936  48.259858
#> 234  -40.85428238  -70.1054230  81.140882
#> 235  -97.15569305  -27.7973957 101.054064
#> 236  -96.31404877   32.8478050 101.761359
#> 237  -73.33924866  102.0692749 125.685251
#> 238  -34.55025864   -9.9741564  35.961148
#> 239   53.57174683 -147.7633972 157.174914
#> 240  173.04051208 -143.9352264 225.078582
#> 241  130.39840698  -46.8980751 138.575517
#> 242   68.79364014   65.7462616  95.158477
#> 243  -41.37591171   53.6094589  67.719570
#> 244  -48.39719772  -24.3288345  54.168080
#> 245  -72.36205292   13.6375103  73.635918
#> 246  -96.02235413   93.3670654 133.931704
#> 247  -78.13285065   59.5743713  98.253998
#> 248  -41.02029419  -40.4577942  57.615082
#> 249   57.33001709 -131.8069916 143.735221
#> 250   37.21800232  -80.5320816  88.716379
#> 251   42.28664398   -6.9696994  42.857169
#> 252   -2.51151395   74.3646240  74.407023
#> 253  -11.76111221  102.2346420 102.908920
#> 254   46.85015869   84.0394821  96.216277
#> 255   31.90343285   -7.4103236  32.752739
#> 256   17.51188087 -107.5296478 108.946276
#> 257  -78.68994141  -68.8936234 104.586989
#> 258    7.85785246    0.1350987   7.859014
#> 259  127.37351227   64.1697540 142.624573
#> 260   19.90662193   75.4045410  77.987938
#> 261  -87.14292145   30.4195213  92.299708
#> 262  -87.21203613   -5.6875877  87.397299
#> 263  -60.40447235 -143.6787567 155.859826
#> 264  -45.30232239 -208.0883179 212.962552
#> 265  -58.06842041  -85.7092514 103.527857
#> 266   30.50443077   33.4557686  45.274814
#> 267   13.63391018  134.5781708 135.267023
#> 268   -9.56841183  107.0674973 107.494202
#> 269   -1.21612036   21.4157791  21.450281
#> 270  -73.26106262  -33.5918198  80.595246
#> 271  -23.02773857  -97.6143036 100.293714
#> 272    7.33815670   33.6323433  34.423583
#> 273  135.89028931  134.7815704 191.395513
#> 274  101.23017120  106.6968002 147.077377
#> 275    8.09900761  -51.4778976  52.111111
#> 276    5.62443638 -140.8063812 140.918669
#> 277   -7.81111622   18.5921631  20.166360
#> 278    1.40035963  172.4285431 172.434229
#> 279   -8.81647778  178.8246460 179.041851
#> 280 -100.35132599  -21.9192429 102.717291
#> 281  -48.06796646 -130.8698883 139.418281
#> 282   40.65674210 -131.8518372 137.977816
#> 283  144.58630371  -37.0517998 149.258283
#> 284  139.60763550    2.6352682 139.632505
#> 285  -31.57807541 -106.6791992 111.254781
#> 286  -97.73182678  -87.2106552 130.985527
#> 287  -60.85244370  -35.5067787  70.453894
#> 288   76.71804047   76.2443848 108.161287
#> 289   67.97248840   84.8131943 108.690097
#> 290  -75.96234894   17.1220837  77.868121
#> 291 -167.88365173   48.4632339 174.738678
#> 292   -2.53571653   42.8083992  42.883434
#> 293  133.18492126   54.5275421 143.914822
#> 294   68.69570160  -21.0088806  71.836429
#> 295  -76.46965790    2.1112866  76.498798
#> 296  -95.64031219  -26.6963596  99.296349
#> 297   64.54970551  -72.8247147  97.314457
#> 298  100.37521362  -46.0035362 110.415166
#> 299   61.89774323   54.5755081  82.521614
#> 300    8.05028343  199.5858917 199.748180
#> 301    7.48270798   91.7674103  92.071975
#> 302    2.79831100  -55.6801224  55.750395
#> 303  -18.35310364 -140.8944855 142.084807
#> 304   -1.78361762  -82.1664352  82.185792
#> 305  -49.07864380   -6.1809936  49.466332
#> 306  -43.55530548   -5.0886917  43.851561
#> 307  -41.41989899   57.2729568  70.680971
#> 308  -55.06145477   50.4189873  74.658141
#> 309  -46.12218857  102.8448944 112.713480
#> 310  -30.85380363   40.5159149  50.926384
#> 311   16.83067322  -40.4095535  43.774463
#> 312  -42.18483353  -35.3935280  55.065979
#> 313  -72.05078125  -41.7970161  83.296492
#> 314   10.37494946  104.6968460 105.209644
#> 315   88.90493774   47.2119751 100.663094
#> 316   91.81536102  -64.0797577 111.965512
#> 317   93.47413635  -49.4689484 105.757227
#> 318   67.01581573  -33.0783577  74.734847
#> 319   -5.50015783   -6.9726911   8.880887
#> 320  -79.67799377  -61.7384453 100.797908
#> 321 -122.58490753   11.6691914 123.139066
#> 322   -9.73796177   59.3835526  60.176692
#> 323   58.19223404    7.5770779  58.683458
#> 324    1.25832844  -71.8815689  71.892582
#> 325   19.70372772  -66.8931808  69.734744
#> 326   81.20130157  -66.0564957 104.676225
#> 327   93.22215271  -67.7555771 115.244037
#> 328   -7.93833017   59.1944618  59.724379
#> 329  -67.21492767  118.1341171 135.917314
#> 330  -31.33059692   48.0713158  57.379942
#> 331  -74.81652832  -46.9415321  88.323385
#> 332 -120.76703644 -105.2729568 160.209465
#> 333   28.23939705 -170.3919220 172.716156
#>  [ reached 'max' / getOption("max.print") -- omitted 667 rows ]