Differentiate x and y and compute change amplitude via an algorithm proposed by Engbert & Kliegl (2003)

Differentiate x and y and compute change amplitude via an algorithm proposed by Engbert & Kliegl (2003)

Usage

diff_ek(x, y, trial, sample_rate, options = NULL)

Arguments

x

vector with x coordinates in degrees of visual angle

y

vector with y coordinates in degrees of visual angle

trial

vector with trial index

sample_rate

sample rate in Hz

options

List with method specific options, see Details.

Value

data.frame with columns x, y, and amp

Details

Method options, please refer to Engbert & Kliegl (2003) for details on parameters and the rationale for default values.

ek_velocity_time_window Time window for velocity computation in milliseconds. Defaults to 20 ms.

See also

Examples

diff_ek(rnorm(1000), rnorm(1000), rep(1, 1000), 250, list("ek_velocity_time_window" = 20))
#>                  x            y        amp
#> 1               NA           NA         NA
#> 2    -8.077625e+00  -22.8139839  24.201775
#> 3     1.794729e+02   -9.5910778 179.728993
#> 4     8.556277e+01  -22.6826839  88.518318
#> 5    -1.250885e+02  -31.3869953 128.966154
#> 6    -1.289804e+02   30.4282360 132.520984
#> 7    -7.633570e-01   61.8767281  61.881437
#> 8     9.173954e+00   20.4329453  22.397917
#> 9    -1.072052e+02  -27.0159683 110.556837
#> 10   -1.529640e+02  -54.3395920 162.329256
#> 11   -4.512189e+01   56.0185509  71.930961
#> 12    4.856470e+01   64.5880737  80.809340
#> 13    2.971875e+01   49.0585365  57.358035
#> 14   -3.395208e+00   16.1309185  16.484355
#> 15   -1.585288e+01   53.4598579  55.760830
#> 16    1.628215e+01   24.3025589  29.252739
#> 17    2.025349e+01  -90.9500732  93.177894
#> 18    4.296729e+01 -102.3871689 111.037472
#> 19    7.227572e+01  -50.5920944  88.223237
#> 20    1.205680e+02  -39.1483192 126.764467
#> 21   -1.410433e+01  -61.4128685  63.011686
#> 22   -6.441120e+01   -2.9839394  64.480282
#> 23   -7.939085e+01   73.5607376 108.231643
#> 24   -8.991124e+01  153.8264313 178.175761
#> 25   -3.586937e+01   74.5485992  82.729110
#> 26   -2.684113e+01    4.8637047  27.278234
#> 27    1.132715e+01  -29.6977615  31.784608
#> 28    1.397943e+01   -4.6691718  14.738581
#> 29   -4.701713e+01   14.6554928  49.248287
#> 30    1.641826e+01  -51.8658142  54.402408
#> 31    6.592610e+01 -100.3466415 120.065396
#> 32    1.289184e+02  -44.4935608 136.380474
#> 33    6.760725e+01   52.5322189  85.617602
#> 34   -1.048286e+02   12.4974060 105.570962
#> 35   -7.611858e+01  -96.1474380 122.631022
#> 36    3.149728e+01 -134.9082031 138.536283
#> 37    1.497010e+02   12.9708099 150.261834
#> 38    4.571788e+01  175.6718903 181.523379
#> 39   -8.511722e+00  213.5549011 213.724461
#> 40    3.940387e+01  123.8294525 129.947676
#> 41   -3.713738e+01  -92.2396088  99.435057
#> 42   -1.049199e+02  -99.1587143 144.362839
#> 43   -9.571910e+01  -10.5544291  96.299233
#> 44    4.652169e+01   36.4348221  59.091146
#> 45    6.842945e+01  -98.5041428 119.940218
#> 46   -7.930933e+01 -126.0471115 148.922273
#> 47   -7.451449e+01   48.9754753  89.168414
#> 48    5.398159e+01   99.2824326 113.008907
#> 49    1.347377e+02  100.4916992 168.085757
#> 50    1.332608e+01  -37.4317169  39.733082
#> 51   -1.226151e+02  -80.4006882 146.624487
#> 52   -1.127652e+02  -35.4936752 118.219256
#> 53    4.415364e+01  -21.7472935  49.218781
#> 54    4.128328e+01   69.1218643  80.511748
#> 55   -4.045891e+01   39.2636147  56.378672
#> 56   -3.606009e+01  -14.9660883  39.042458
#> 57    4.198073e+01  -44.8748283  61.450238
#> 58    1.493961e+02   21.8554478 150.986233
#> 59    7.267809e+01  109.5365143 131.454757
#> 60    5.147539e-01   45.4452934  45.448209
#> 61   -7.006047e+01  -97.3463974 119.936611
#> 62   -1.128565e+02 -134.9128723 175.892204
#> 63   -6.160200e+01   44.8312607  76.188244
#> 64    1.616345e+01   99.0262146 100.336675
#> 65    3.750221e+01  110.5354385 116.724029
#> 66    4.304031e+01  -44.8281174  62.145217
#> 67    1.173400e+01  -84.2923508  85.105153
#> 68    5.076120e+01  -22.1076241  55.366478
#> 69    5.063875e+01   -5.4714909  50.933486
#> 70    4.501285e+01   38.3567963  59.138822
#> 71    2.104716e+01  -71.5144424  74.547291
#> 72   -1.058250e+02  -66.2787933 124.867131
#> 73   -1.206789e+02  -37.0577660 126.240560
#> 74   -5.288466e+01   26.0822029  58.966676
#> 75    6.366418e+01   23.7159519  67.938014
#> 76    4.714768e+01   69.2931137  83.811930
#> 77    1.496102e+01  207.8789062 208.416582
#> 78    4.364905e+01   97.2532043 106.599367
#> 79   -1.983776e+01 -114.5480347 116.253124
#> 80   -7.247018e+01 -171.5417633 186.221653
#> 81   -9.204671e+01   -8.2258844  92.413543
#> 82   -4.281021e+01   45.5285416  62.494498
#> 83    1.054889e+02  -56.2650948 119.556092
#> 84    9.018996e+01  -26.2573357  93.934424
#> 85    5.319594e+01  -35.7327423  64.083045
#> 86   -7.499035e+01   22.3376274  78.246546
#> 87   -7.330949e+01  -41.2665215  84.126142
#> 88    5.298899e+01  -31.2002888  61.492201
#> 89    3.854849e+01  -16.6996899  42.010307
#> 90   -1.015154e+02   11.6652880 102.183398
#> 91   -9.284415e+01  127.7671967 157.938259
#> 92    6.771463e+01   19.5086651  70.468852
#> 93    2.043765e+02   55.3860550 211.748402
#> 94    2.998005e+01   35.2081604  46.243035
#> 95   -2.067643e+02   34.2042809 209.574394
#> 96   -9.895097e+01 -101.9363022 142.064439
#> 97    7.682458e+01 -154.4358063 172.488939
#> 98    1.410725e+02  -57.4793205 152.332928
#> 99   -2.930311e+01   53.1370773  60.681307
#> 100  -6.844438e+01   71.0664368  98.666468
#> 101  -3.031963e+01    1.2329903  30.344694
#> 102   3.375787e+01   10.9122696  35.477756
#> 103   8.017630e+00  -18.3298759  20.006667
#> 104  -1.326785e+01   12.8647146  18.480712
#> 105  -1.771230e+01   37.4762154  41.451081
#> 106  -6.135316e+01  -18.4857788  64.077565
#> 107   6.325386e+01  -88.4467468 108.737656
#> 108   8.332809e+01 -111.2977982 139.035139
#> 109  -1.685714e+01  -74.1827850  76.073969
#> 110  -2.090152e+01   55.5858765  59.385714
#> 111   9.366139e+00   26.8130417  28.401827
#> 112   7.821201e+01   96.8812332 124.511410
#> 113   4.710155e+01  120.0781631 128.985741
#> 114  -1.531557e+00   57.6075249  57.627880
#> 115  -4.187147e+01  -40.0741081  57.958211
#> 116  -1.146347e+02  -77.5616379 138.408525
#> 117  -6.469894e+01   48.9238243  81.114074
#> 118   9.970206e+01   76.1709290 125.469162
#> 119   1.309002e+02   -5.6527419 131.022234
#> 120  -6.040194e+01  -70.3644714  92.733773
#> 121  -1.131476e+02   10.7274265 113.654967
#> 122  -8.878702e+00   27.4634132  28.862959
#> 123   8.133452e+01  -39.9244652  90.605004
#> 124   5.899996e+01   11.3928108  60.089863
#> 125  -3.361998e+01   83.9127960  90.397238
#> 126  -5.059217e+01  110.2900009 121.340231
#> 127  -2.508564e+01   56.8592377  62.147102
#> 128  -2.487826e+00  -63.0402908  63.089362
#> 129   4.351749e+01 -145.5614777 151.927336
#> 130   4.189474e+01 -148.7973938 154.582772
#> 131  -2.884379e+01  -74.8874207  80.250171
#> 132  -1.070396e+02   89.9151688 139.793506
#> 133  -1.258450e+02  194.3829041 231.563529
#> 134  -1.278184e+01  192.7965393 193.219774
#> 135   9.836974e+01   97.3181915 138.374264
#> 136   1.364814e+02  -94.3932343 165.943492
#> 137   2.209558e+01 -233.1728516 234.217406
#> 138  -9.073870e+00 -254.2043762 254.366271
#> 139  -4.138307e+01 -129.2496338 135.713030
#> 140   2.818983e+01   44.7428589  52.882792
#> 141   6.594559e+01   57.7651787  87.667761
#> 142  -3.352991e+01   39.6782913  51.948260
#> 143  -4.238742e+01   73.2514038  84.631327
#> 144  -8.529970e+01   95.9854202 128.410433
#> 145   4.772546e+01   14.8349895  49.977960
#> 146   2.747295e+00 -100.5764465 100.613961
#> 147  -3.447561e+01  -55.2988892  65.165439
#> 148  -3.210084e+00   40.0213547  40.149888
#> 149  -4.224370e+01   32.9949646  53.602220
#> 150   2.807216e+01  -97.9632416 101.906049
#> 151   4.400857e+01 -121.5666428 129.287287
#> 152   1.693123e+02   29.6579838 171.890186
#> 153   2.346097e+02  218.4077148 320.536475
#> 154   9.278928e+01  171.3156738 194.830468
#> 155  -7.755863e+01   32.6223602  84.140120
#> 156  -1.974349e+02 -126.0838470 234.259878
#> 157  -8.071347e+01  -86.6867599 118.445172
#> 158  -4.027995e+01   -6.2043691  40.754978
#> 159  -7.908800e+01   54.1212997  95.833326
#> 160  -1.561984e+02   39.4276428 161.097771
#> 161  -4.633281e+01  -11.6716471  47.780295
#> 162   1.034257e+02   75.5335007 128.070983
#> 163   1.670699e+02   26.0484695 169.088333
#> 164   1.163660e+02  -55.6104736 128.971226
#> 165   8.085287e+01  -46.4878044  93.264688
#> 166  -1.100148e+01   -2.5265281  11.287868
#> 167  -1.022998e+02   14.2975025 103.294111
#> 168  -5.060721e+01  -49.2762222  70.634524
#> 169   3.569227e+01 -102.2464752 108.297184
#> 170   1.236662e+02   40.0077934 129.976745
#> 171  -7.159995e+01  113.0796127 133.841518
#> 172  -1.080998e+02  131.6261444 170.326183
#> 173  -3.722802e+01   33.4070053  50.019528
#> 174   5.157543e+01  -62.6539879  81.151385
#> 175  -1.403031e+00  -50.4063034  50.425826
#> 176  -1.263638e+02  -41.3953400 132.971395
#> 177  -7.036749e+01  -84.3695374 109.862655
#> 178  -1.809713e+01 -124.0052261 125.318802
#> 179   7.422413e+01  -65.6109085  99.065698
#> 180   9.378937e+01   35.2936516 100.210216
#> 181  -2.116138e+01  139.8562622 141.448146
#> 182  -1.044318e+02  142.5800476 176.734465
#> 183  -1.712544e+02   83.3757782 190.472014
#> 184  -2.798316e+01  -15.5932751  32.034470
#> 185   1.052997e+02 -139.9693451 175.155462
#> 186   9.885110e+01 -107.6825485 146.174795
#> 187   1.281787e+02   15.7008944 129.136719
#> 188  -2.807669e+00   88.4223938  88.466958
#> 189  -6.190978e+01  100.8948746 118.374813
#> 190  -5.238468e+01   56.4311142  76.997566
#> 191   6.611857e+01  -11.1010523  67.044004
#> 192   1.884991e+02 -144.4959869 237.509977
#> 193   4.245706e+01 -172.6634827 177.806862
#> 194  -1.103775e+02  -42.5535889 118.296278
#> 195  -9.326322e+01   44.1405220 103.181463
#> 196   5.258543e+00   -1.4257603   5.448401
#> 197   1.460549e+01  -52.3014793  54.302534
#> 198  -6.268286e+01   27.3284092  68.381160
#> 199   3.374431e+01  107.1578445 112.345370
#> 200   8.805130e+01   73.6173019 114.771680
#> 201  -2.479579e+01  -19.0250664  31.253549
#> 202  -1.267954e+02  -60.9442024 140.681435
#> 203  -2.971315e+01  -71.4375458  77.370498
#> 204   1.594727e+02 -110.1318359 193.805442
#> 205   8.693166e+01  -60.6124039 105.976307
#> 206   3.076535e+01   27.2240715  41.081104
#> 207  -7.861144e+01   98.5557556 126.067426
#> 208  -8.528455e+01   67.5134888 108.772819
#> 209  -8.785757e+01   -1.6125542  87.872372
#> 210  -4.584419e+01   46.8432999  65.543762
#> 211   6.001301e+01   85.7164230 104.636833
#> 212  -3.085621e+01   69.4803009  76.023797
#> 213   8.955528e-02  -36.8765030  36.876612
#> 214   7.862588e+01  -72.5018692 106.951155
#> 215   1.222050e+02  -29.9724751 125.826897
#> 216   8.148642e+01   11.0910234  82.237749
#> 217  -3.071452e+01   47.2452431  56.351529
#> 218  -4.027537e+01   -1.8570722  40.318163
#> 219  -5.833248e+01  -33.4093742  67.222501
#> 220  -4.035297e+01  -31.4393406  51.154609
#> 221  -1.280413e+01   18.9716187  22.888167
#> 222  -2.896860e+01  -30.6019516  42.138569
#> 223  -1.072006e+02 -111.6438828 154.778336
#> 224  -9.783620e+01  -37.0469398 104.615480
#> 225   1.674001e+01   51.3757172  54.034176
#> 226   8.120405e+01   59.7067451 100.791829
#> 227   5.001429e+01   85.0271988  98.646102
#> 228  -2.667146e+01  103.1832199 106.574592
#> 229  -6.400441e+01   57.0930252  85.768164
#> 230  -1.634270e+01   -6.4469185  17.568345
#> 231   1.107460e+02  -57.8673592 124.953211
#> 232   1.436642e+02  -51.0996666 152.481360
#> 233   1.477849e+02  -87.0328827 171.508317
#> 234   6.855023e+01  -61.9279976  92.380794
#> 235   2.587239e+01   20.5862293  33.063176
#> 236   4.761752e+01   93.5592804 104.979840
#> 237  -2.355733e+01  104.1942139 106.824070
#> 238  -3.765152e+01   42.7016678  56.930388
#> 239  -2.303913e+01  -33.8898773  40.979573
#> 240  -6.829387e+01  -27.9133511  73.778098
#> 241  -9.210484e+01  -20.9579201  94.459179
#> 242  -1.713764e+02   28.0845604 173.662357
#> 243  -8.864550e+01  102.2628326 135.335552
#> 244   1.555824e+01   61.5392609  63.475502
#> 245   6.487934e+01  -28.0725956  70.692288
#> 246   1.400544e+02 -106.6663361 176.048135
#> 247   1.331491e+02  -11.6260223 133.655683
#> 248   6.194496e+01    7.9099011  62.447933
#> 249  -3.151613e+01    6.2132649  32.122747
#> 250  -1.022372e+02  -29.1133518 106.301601
#> 251  -7.313013e+01  -33.8705788  80.593006
#> 252   1.129096e+01   39.9906693  41.554055
#> 253   1.548966e+01   53.6730919  55.863497
#> 254   2.482620e+01   33.2004242  41.456100
#> 255  -7.210076e+01   21.0612392  75.113884
#> 256  -5.432963e+01    7.7141981  54.874564
#> 257   3.445219e+01   22.5212746  41.160187
#> 258   1.143704e+02  -49.2592812 124.527353
#> 259   8.562348e+01  -17.0272522  87.300103
#> 260  -1.003016e+02   15.4648066 101.486793
#> 261  -1.637963e+02  -42.4199066 169.200073
#> 262  -9.493375e+01 -127.9884872 159.353288
#> 263   2.692176e+00 -161.6381836 161.660602
#> 264   1.115882e+02 -107.5821686 155.002775
#> 265   1.537605e+02  -71.3594284 169.512443
#> 266   6.540392e+01  101.6611481 120.882845
#> 267  -1.002264e+01  162.7700348 163.078317
#> 268  -5.574609e+01   71.3244095  90.525121
#> 269   3.746956e+01  -69.8262482  79.244386
#> 270  -4.291981e+01  -92.4217682 101.901389
#> 271  -6.443288e+01   64.2923584  91.022541
#> 272  -2.508442e+01   78.1816711  82.107258
#> 273   1.325527e+01    7.5322433  15.245883
#> 274   2.518362e+01  -74.3848648  78.532305
#> 275  -1.400455e+01 -126.9221115 127.692403
#> 276  -3.035607e+01   25.4475155  39.611453
#> 277  -6.843682e+01  144.9380951 160.283031
#> 278   3.188746e+01  179.7564392 182.562833
#> 279   1.099324e+02   13.7384176 110.787578
#> 280   9.053770e+01   12.5189409  91.399116
#> 281  -7.245327e+00    0.2799290   7.250732
#> 282   9.498380e+00  -24.9204941  26.669275
#> 283   5.887778e+01   -8.2312317  59.450372
#> 284   1.767286e+01  -42.3476982  45.887444
#> 285  -5.097213e+01 -127.9101257 137.692261
#> 286  -8.834760e+01 -208.5854950 226.524188
#> 287  -5.568506e+01  -19.3087406  58.937710
#> 288   2.085143e+01  116.1339874 117.991039
#> 289   4.038394e+01  115.4386673 122.298603
#> 290   3.174670e+00  -22.3885670  22.612529
#> 291  -4.120549e+01   33.3418732  53.005402
#> 292   2.366675e+01   90.2678375  93.318794
#> 293  -2.760313e+00   28.4063416  28.540140
#> 294   2.762549e+01  -37.6096649  46.665346
#> 295  -1.452012e+01 -172.0338287 172.645510
#> 296  -9.854109e+01  -74.4148712 123.482468
#> 297  -9.313708e+01   41.1816559 101.835384
#> 298  -5.469982e+01  103.0194321 116.640788
#> 299   7.748067e+01   87.3809814 116.784801
#> 300   3.497821e+01   -2.7475717  35.085960
#> 301   5.300639e+00  -14.6169777  15.548402
#> 302  -4.647025e+01  -76.5917740  89.586742
#> 303  -7.653239e+01  -48.6844711  90.704928
#> 304  -8.419600e+01   -6.3071051  84.431901
#> 305  -3.791636e+01   63.8460922  74.256138
#> 306   4.703150e+00    7.3035307   8.686840
#> 307   1.790562e+01 -142.4713287 143.592098
#> 308   8.383813e+01 -167.3107300 187.140888
#> 309   4.776479e+01  -39.7213326  62.122935
#> 310   3.238754e+01   91.1839676  96.765018
#> 311   1.989401e+01  123.8511124 125.438709
#> 312   5.419326e+00  116.7756271 116.901310
#> 313   5.406345e+01   33.9008102  63.813180
#> 314  -7.191940e+00    6.2102180   9.502147
#> 315  -3.009915e+01    7.8354516  31.102297
#> 316   9.461310e+00   63.9007072  64.597343
#> 317  -2.114411e+01   51.6686478  55.827614
#> 318  -1.008060e+02 -117.5235138 154.834198
#> 319  -1.560686e+02 -142.5826874 211.393558
#> 320  -7.332007e+01  -69.5782928 101.079035
#> 321   3.663582e+01   21.0789757  42.267084
#> 322   5.385767e+01  114.5787430 126.605438
#> 323   3.276965e+01   45.7742271  56.295027
#> 324   1.123602e+02  104.5343399 153.467416
#> 325   1.522832e+02   96.9855194 180.544647
#> 326   1.323008e+02   52.3257942 142.272631
#> 327   5.811001e+01 -100.9406662 116.472276
#> 328  -8.926819e+01 -144.5216522 169.868530
#> 329  -1.248693e+02   27.8989029 127.948016
#> 330  -3.502336e+01   46.7339935  58.401216
#> 331   3.637534e+01  -96.2394180 102.884357
#> 332   7.023663e+01 -146.3615112 162.341851
#> 333   3.995766e+01   13.7835321  42.268193
#> 334  -1.518618e+01  145.6266174 146.416297
#> 335  -4.819999e+00   81.8436966  81.985505
#> 336  -2.671523e+01   17.2741756  31.813527
#> 337  -3.198857e+01    3.2040107  32.148624
#> 338   5.066960e+00  -86.4182587  86.566677
#> 339  -5.341912e+01 -102.7575531 115.813284
#> 340  -5.099371e+01  -81.4809265  96.122315
#> 341  -3.642356e+01  111.7023087 117.490773
#> 342   6.451291e+01  113.4454498 130.505883
#> 343   1.494785e+02   53.6888847 158.827924
#> 344   1.392017e+02  -12.5242243 139.763984
#> 345   3.161116e-01  -32.1498566  32.151411
#> 346  -1.475537e+02    8.6515274 147.807080
#> 347  -1.409628e+02  -68.3503723 156.659758
#> 348  -1.928990e+01  -16.2717209  25.236268
#> 349   1.395154e+02   51.3612862 148.669202
#> 350   6.410757e+01   98.1143341 117.201547
#> 351  -4.881116e+01   -1.9904509  48.851724
#> 352  -9.698062e+01  -10.0774469  97.502799
#> 353  -1.200850e+02   43.9358063 127.870070
#> 354  -1.061971e+02   80.1823196 133.067751
#> 355  -1.128463e+02   29.3843098 116.609238
#> 356  -8.144404e+00  -38.0227242  38.885201
#> 357   7.893238e+01  -57.2937088  97.534044
#> 358   1.087476e+02 -108.2058411 153.409736
#> 359   7.003011e+01  -69.3987045  98.592068
#> 360   1.121306e+01  -74.9678116  75.801751
#> 361   4.000248e+01  -75.3296356  85.292159
#> 362   3.450135e+01  -76.7932739  84.187590
#> 363   6.044273e+01  -53.6734695  80.834185
#> 364   2.619139e+01   91.3501282  95.030705
#> 365  -8.745516e+01  163.4685822 185.392510
#> 366  -1.149634e+02   91.9848785 147.233827
#> 367  -6.754928e+01  -99.5811081 120.329972
#> 368   5.604247e+01 -118.5922089 131.167339
#> 369   4.222216e+01  118.6777420 125.964746
#> 370  -2.540946e+01  163.1311798 165.098221
#> 371   7.255486e+01   33.2616768  79.815708
#> 372   4.836124e+01 -147.7240753 155.438771
#> 373   3.982022e+00 -114.6119690 114.681123
#> 374  -5.565761e+00  -41.8823166  42.250516
#> 375   4.322592e+01   92.9548645 102.513837
#> 376   6.464175e+01  107.4561462 125.400872
#> 377  -4.568219e+01  -14.5108671  47.931486
#> 378  -3.975755e+01  -90.2443390  98.613909
#> 379   2.120724e+01 -109.6787643 111.710242
#> 380   5.162133e+01   25.7105598  57.669706
#> 381  -1.467440e+01   75.4960861  76.909019
#> 382  -1.025728e+02    3.3923819 102.628844
#> 383  -7.714974e+01    6.1841321  77.397198
#> 384   1.800171e+01   72.3399582  74.546167
#> 385   4.340702e+01  123.0041656 130.438470
#> 386   5.048168e+01   51.2343636  71.926075
#> 387   8.965429e+01 -108.2156296 140.529406
#> 388   1.494256e+02  -53.0680466 158.569284
#> 389   9.993816e+01   -5.3615279 100.081872
#> 390  -3.393816e+01   48.6641655  59.329583
#> 391  -1.235256e+02  -17.2994785 124.731113
#> 392  -1.728519e+02  -59.3925362 182.771052
#> 393  -7.161314e+01  -52.7384186  88.936962
#> 394  -2.504667e+01   42.7487640  49.545864
#> 395   5.390180e-01  155.1418610 155.142797
#> 396   1.381024e+01   96.5953445  97.577576
#> 397   2.350816e+01   -3.9506543  23.837807
#> 398   5.602363e+01 -128.9496613 140.593962
#> 399  -4.413950e+01 -140.4117737 147.186146
#> 400  -5.569236e+01  -79.6971741  97.227970
#> 401   6.733572e+01   48.4521446  82.956073
#> 402   2.855066e+01  154.4230957 157.040226
#> 403  -5.561331e+01  190.3031006 198.262730
#> 404  -1.431235e+02   15.7224569 143.984459
#> 405   2.637947e+00 -166.4693298 166.490230
#> 406   1.121081e+02 -143.4118347 182.030713
#> 407   3.806921e+01  -67.8335648  77.785971
#> 408  -4.619397e+01  -88.2398376  99.599957
#> 409  -2.623984e+01  -88.3565521  92.170546
#> 410   6.310497e+00   -6.6974440   9.202072
#> 411  -7.437817e+01  178.2868652 193.179500
#> 412  -1.209229e+02   86.5794449 148.722390
#> 413  -2.568823e+00  -40.5221519  40.603493
#> 414   1.554969e+02 -129.8703003 202.597080
#> 415   9.961335e+01 -100.0535583 141.186168
#> 416  -2.410157e+00   51.4843407  51.540724
#> 417  -8.306356e+01   85.4526291 119.170915
#> 418   2.238532e+01   76.6289291  79.831668
#> 419   4.774776e+01  -46.4155273  66.590166
#> 420   5.965284e+00   31.8926792  32.445764
#> 421  -6.973201e+01   93.0637512 116.290219
#> 422  -2.892353e+01   96.5564651 100.795444
#> 423   8.777171e+01   38.7444458  95.942714
#> 424   9.597993e+01  -22.6683445  98.620486
#> 425   5.229927e+01  -54.6816978  75.665724
#> 426  -4.715396e+01  -82.8754120  95.351086
#> 427  -1.175769e+02 -125.8136826 172.201658
#> 428  -1.225200e+02  -44.4826469 130.345097
#> 429   4.505286e+01   54.5894241  70.779697
#> 430   1.667373e+02  138.4808807 216.744707
#> 431   9.865926e+01  106.7095490 145.329201
#> 432  -1.753189e+01  -86.9038391  88.654636
#> 433  -8.949892e+01  -69.4442902 113.280920
#> 434  -8.906823e+01  -89.8712082 126.530564
#> 435   4.673037e+00   18.6335125  19.210546
#> 436   8.829584e+01   43.7699509  98.549295
#> 437   1.288305e+02   49.0968170 137.868794
#> 438  -1.281303e+01   51.3250771  52.900258
#> 439  -9.126456e+01   14.1212873  92.350590
#> 440  -2.348415e+01   60.9182358  65.288104
#> 441  -2.214902e+01   99.3796844 101.817978
#> 442  -1.321943e+01   69.5803528  70.824987
#> 443  -6.065195e+01  -52.8718643  80.461750
#> 444  -9.320169e+00  -41.2065964  42.247475
#> 445   9.460252e+01  -26.9050446  98.354042
#> 446   8.390987e+01  -58.7931900 102.457332
#> 447  -5.000809e+01  -72.4137039  88.003144
#> 448  -1.375764e+02  -31.7044239 141.182281
#> 449  -1.094092e+02    5.6733098 109.556158
#> 450   2.598830e+00  -51.7796822  51.844859
#> 451   5.510059e+01 -107.2537689 120.579625
#> 452   7.853204e+01   63.4845886 100.983042
#> 453   9.362263e+01  163.7278137 188.605390
#> 454   1.577194e+01   99.2257080 100.471366
#> 455  -3.994048e+00    7.2980595   8.319501
#> 456  -1.108959e+01 -103.7707214 104.361591
#> 457   3.593665e-03  -12.0603380  12.060339
#> 458  -4.003386e+01   33.0748672  51.929341
#> 459  -7.769379e+01   81.0646362 112.284459
#> 460  -1.308821e+02   38.1419449 136.326530
#> 461  -4.403879e+01  -57.4190865  72.362744
#> 462   1.081741e+02  -57.5986938 122.552993
#> 463   1.472138e+02  -44.0759659 153.670427
#> 464   1.322524e+02   60.1093254 145.271577
#> 465   1.282252e+01   -2.1086917  12.994753
#> 466   2.347661e+01  -12.7259798  26.703966
#> 467  -3.255332e+01   75.2909241  82.027079
#> 468  -6.522300e+01   41.5913315  77.355533
#> 469  -7.990137e+00  -13.9852571  16.106822
#> 470  -7.003680e+01 -113.4807892 133.353074
#> 471  -7.919068e+01  -39.4877663  88.489817
#> 472  -8.039981e+01   31.7299213  86.434469
#> 473   8.132256e+00    5.7269239   9.946418
#> 474   1.336230e+02  -32.7659264 137.581701
#> 475   1.248791e+02 -109.5916519 166.147854
#> 476   5.264868e+01 -116.8282089 128.143333
#> 477  -2.399907e+01  -54.0794678  59.165396
#> 478  -1.122099e+01  108.7359543 109.313395
#> 479  -4.460261e+00  113.1624603 113.250326
#> 480  -5.027876e+01  113.6764450 124.299186
#> 481  -1.293479e+02   57.6739769 141.623314
#> 482  -1.353655e+02  -11.0526628 135.815987
#> 483  -7.115482e+01  -85.1994781 111.004319
#> 484   1.232291e+01  -81.8600006  82.782328
#> 485   4.996025e+01   13.1552744  51.663216
#> 486   1.703656e+00   66.0351639  66.057137
#> 487   3.279651e+01   60.1419945  68.503069
#> 488   4.022391e+01   -2.3430200  40.292089
#> 489   1.166971e+02  -62.3592529 132.313594
#> 490   1.443342e+02  -36.5985756 148.901992
#> 491   1.164943e+02   17.6003494 117.816366
#> 492  -1.188914e+01   89.8060379  90.589603
#> 493  -1.959701e+02   71.4246292 208.580318
#> 494  -8.738054e+01  -21.0610237  89.882842
#> 495   6.500185e+01 -129.4949493 144.893695
#> 496   9.960120e+01 -143.2783813 174.496684
#> 497   1.783777e+01  -43.4370766  46.957063
#> 498  -8.914232e+01   60.3296661 107.638384
#> 499  -4.682864e+01   46.5178032  66.006272
#> 500  -7.845441e+01  -18.5459976  80.616673
#> 501  -7.101760e+01   57.4026184  91.315717
#> 502  -1.591548e+01   65.1804123  67.095371
#> 503  -6.531631e+01   19.5605278  68.182360
#> 504   1.489086e+01 -162.5624847 163.243068
#> 505   7.302256e+00 -211.4725342 211.598572
#> 506   7.570940e+01  -87.5229034 115.724554
#> 507   1.314570e+02   84.1609497 156.089745
#> 508   6.527323e+01  209.2078094 219.154061
#> 509   8.095759e+01  143.9719238 165.172776
#> 510   3.846080e+01  -40.6521988  55.962798
#> 511  -1.717433e+01  -95.4778671  97.010209
#> 512  -1.064716e+02  -16.1177921 107.684681
#> 513  -8.966702e+01   37.2815170  97.108625
#> 514   4.235819e+00   13.2888260  13.947583
#> 515  -4.121533e+00    1.5099337   4.389412
#> 516  -1.017846e+02   71.1576462 124.191465
#> 517  -2.020481e+01   26.0074081  32.933566
#> 518   8.532930e+01  -79.5076981 116.630028
#> 519  -1.386841e+01  -33.1844025  35.965780
#> 520   1.229927e+01    6.5317669  13.926093
#> 521   4.072324e+01  -22.7315922  46.638048
#> 522   3.792951e+01  -95.6342621 102.881289
#> 523  -7.593274e+01  -67.5122223 101.605517
#> 524  -9.616632e+01   12.0340834  96.916358
#> 525   1.016788e+02    8.4491491 102.029238
#> 526   1.116252e+02   68.1394806 130.779096
#> 527  -4.886648e+00  106.9251328 107.036738
#> 528  -1.358870e+02   67.4534607 151.707773
#> 529  -6.238451e+01  -62.5464706  88.339621
#> 530   8.119257e+01   -3.2381110  81.257111
#> 531   1.000994e+02  123.9203796 159.298941
#> 532   4.815971e+01   87.4584274  99.841547
#> 533  -2.861408e+01  -47.1509438  55.154121
#> 534   6.714826e-01 -130.5446167 130.546344
#> 535  -9.312294e+01  -11.9367199  93.884862
#> 536  -4.554549e+01   68.2768402  82.073863
#> 537   6.251682e+01   45.6948395  77.436241
#> 538   1.471359e+01  -77.9379807  79.314681
#> 539  -6.190017e+01  -82.7221909 103.317919
#> 540  -1.794027e+02  -48.0961800 185.737949
#> 541  -1.230676e+02   80.2803268 146.937299
#> 542   9.473618e+00   88.0677795  88.575861
#> 543   1.825889e+02   67.0734406 194.518742
#> 544   3.090326e+02  -24.4715824 309.999972
#> 545   1.547791e+02  -73.5843887 171.380385
#> 546  -6.872481e+01    5.6145868  68.953773
#> 547  -9.534832e+01   15.5828457  96.613287
#> 548   5.806208e+01   39.7352982  70.356938
#> 549   6.574722e+01   -2.2329164  65.785122
#> 550  -1.721023e+01   60.8547211  63.241513
#> 551  -6.618015e+01   22.9441452  70.044596
#> 552   1.453890e+01  -43.0346222  45.424203
#> 553   6.953555e+01  -98.5197296 120.587434
#> 554  -5.867159e+01 -111.2745667 125.795010
#> 555  -5.394102e+01  -24.4037952  59.204552
#> 556  -3.797477e+01    3.0495231  38.097021
#> 557   1.267909e+02   42.0851288 133.593033
#> 558   1.315261e+02    3.3469980 131.568687
#> 559  -4.732447e+01   12.2955017  48.895657
#> 560  -1.082962e+02   86.6619644 138.702459
#> 561  -1.121794e+02   66.1292038 130.220162
#> 562   1.538497e+00   56.0403671  56.061482
#> 563   1.581468e+00  -56.8320885  56.854088
#> 564   5.066179e+00   23.8258820  24.358547
#> 565   7.972853e+01   34.6904526  86.948641
#> 566   1.044064e+02   49.0705299 115.363007
#> 567   7.346870e+01   21.4346600  76.531661
#> 568  -7.036700e+01  -88.0223999 112.691868
#> 569  -1.609009e+02  -39.9627571 165.789384
#> 570  -1.911283e+02  -75.6502762 205.555292
#> 571  -9.909491e+01  -29.3587322 103.352486
#> 572   5.709932e+01  -25.3330956  62.466779
#> 573   1.912512e+02  -41.1436729 195.626705
#> 574   1.894078e+02  -55.3440781 197.327889
#> 575   1.612436e+01 -131.6517029 132.635463
#> 576  -1.122990e+02   -4.0409994 112.371679
#> 577  -1.585791e+02  134.7733307 208.113353
#> 578  -1.371966e+02  193.8281708 237.470574
#> 579  -2.926782e+01  123.7312927 127.145735
#> 580   5.200529e+01    0.1673889  52.005557
#> 581   6.798470e+01  -67.7155685  95.954771
#> 582  -4.762197e+00  -98.7060547  98.820867
#> 583  -5.396697e+01   21.4946842  58.090064
#> 584  -6.627126e+01   58.0996704  88.133149
#> 585   2.907643e+01  -23.6071548  37.453124
#> 586   8.974421e+01 -142.5879517 168.479515
#> 587   5.477478e+01  -98.4748993 112.683551
#> 588   9.146138e+01   52.2877388 105.352701
#> 589   7.651089e+01   81.8603439 112.049238
#> 590   1.558915e+01   26.3725796  30.635512
#> 591  -1.438838e+02  -56.7323151 154.664462
#> 592  -1.433219e+02   56.8294868 154.177667
#> 593  -5.909766e+01  115.2040558 129.477826
#> 594  -2.301436e+01  -21.7269859  31.650006
#> 595   5.815884e+01 -153.2042847 163.871912
#> 596   4.139122e+01 -118.5669861 125.584090
#> 597  -1.147283e+01   71.1542053  72.073205
#> 598  -4.291617e+01   42.1678772  60.165832
#> 599   5.444486e+01  -19.0758247  57.689948
#> 600   1.131547e+02   30.6894264 117.242556
#> 601   6.375075e+01   62.1586418  89.038501
#> 602   9.536366e+01   48.2890701 106.892762
#> 603   5.864040e+00  -35.5610008  36.041251
#> 604  -1.018849e+02   42.2699280 110.305410
#> 605  -8.066453e+01   32.8857002  87.110478
#> 606  -4.185318e+01  -79.6845322  90.007296
#> 607  -1.937538e+00 -125.4168625 125.431828
#> 608  -1.705853e+01 -106.4259796 107.784426
#> 609  -5.978405e+00   93.0667114  93.258534
#> 610   9.762326e+01   97.0671768 137.667490
#> 611   6.873447e+01  -17.8482437  71.013997
#> 612  -2.342872e+01  -31.1336861  38.964232
#> 613  -1.376179e+02   26.0625439 140.064071
#> 614  -1.934231e+02  107.4498978 221.264463
#> 615  -4.926471e+01   43.5380402  65.746271
#> 616   1.369107e+02  -14.3524532 137.660970
#> 617   2.031977e+02  -14.7356606 203.731344
#> 618   1.007429e+02  -53.3075371 113.977327
#> 619  -5.988102e+01  -97.0903702 114.071366
#> 620  -7.380715e+01  -89.4952393 116.003851
#> 621   7.637558e+00  -35.7719688  36.578218
#> 622   4.899500e+01  -11.4478569  50.314647
#> 623   1.478138e+01   51.7776108  53.846172
#> 624  -4.144234e+01  171.4118500 176.350475
#> 625  -5.690809e+01  184.9636688 193.520256
#> 626  -1.031683e+01   65.2214279  66.032353
#> 627   4.940079e+01 -138.1756134 146.741057
#> 628   8.307223e+01 -156.1310425 176.855583
#> 629   4.617537e+01    2.1991475  46.227704
#> 630   7.821050e+00   11.4593840  13.873943
#> 631  -1.164933e+02   -6.4797359 116.673405
#> 632  -1.805671e+02 -122.5505829 218.227209
#> 633  -5.659922e+01  -64.4415436  85.768203
#> 634   1.330817e+02   22.8394165 135.027303
#> 635   2.079458e+02   35.6099319 210.972757
#> 636   8.636446e+01   -5.9371815  86.568293
#> 637   7.081523e+00   -8.4516277  11.026240
#> 638  -9.267265e+01   76.2044144 119.980548
#> 639  -1.302575e+02  115.4233017 174.038941
#> 640  -8.249175e+01  119.1726227 144.937929
#> 641   3.037292e+01   91.4047852  96.318995
#> 642   7.660538e+01    4.1577582  76.718133
#> 643   2.835245e+01  -73.0396042  78.349507
#> 644  -2.062596e+01  -59.6015015  63.069559
#> 645   5.184433e+01    5.3643541  52.121121
#> 646   8.386115e+01   89.0139999 122.295483
#> 647  -2.849623e+01  -34.9407158  45.087566
#> 648  -9.738850e+01  -50.5931282 109.746004
#> 649  -1.288906e+01  -33.3820724  35.783943
#> 650   7.573370e+01  -71.5139160 104.162531
#> 651   1.831840e+00  -57.1774826  57.206819
#> 652  -1.608899e+01  -46.5291939  49.232323
#> 653   1.370951e+01  143.8746796 144.526379
#> 654  -1.040900e+01  198.6758118 198.948298
#> 655  -7.195120e+01    8.2288017  72.420223
#> 656  -1.130377e+01 -105.5770035 106.180407
#> 657   2.791084e+01 -143.0901031 145.786806
#> 658  -5.688698e+01  -40.1209602  69.611925
#> 659  -1.025062e+02   77.4627762 128.483493
#> 660  -1.399415e+01   68.9707794  70.376166
#> 661   2.397180e+01   12.3788252  26.979294
#> 662  -3.241420e+01 -107.7089615 112.480669
#> 663  -1.374351e+01  -93.0749435  94.084160
#> 664   1.389512e+01   28.2938881  31.521712
#> 665  -1.564345e+01   68.2543640  70.024108
#> 666  -1.306960e+01   65.6543503  66.942574
#> 667   6.707559e+01   50.9568214  84.236172
#> 668   1.661511e+02  -53.4731941 174.543857
#> 669   1.706223e+02 -145.4130859 224.180620
#> 670   4.854062e+01  -66.8509903  82.615048
#> 671  -5.547684e+01   34.0188980  65.076608
#> 672  -1.061553e+02   77.3519211 131.347878
#> 673  -2.898342e+01  -23.3662949  37.229318
#> 674   4.494844e+01   -8.1955967  45.689493
#> 675   1.016265e+01   97.1394882  97.669645
#> 676  -5.593533e+01  -28.7196560  62.877501
#> 677  -7.188850e+01 -196.9117126 209.623898
#> 678   2.516729e+01 -121.4559097 124.036004
#> 679   8.740979e+01  125.4350662 152.886975
#> 680   2.981641e+01  271.9767151 273.606198
#> 681  -1.079485e+02  -50.1229973 119.017641
#> 682  -2.471710e+02 -178.7821198 305.051733
#> 683  -1.598908e+02  -21.0718441 161.273334
#> 684   1.066968e+02  165.1162872 196.589909
#> 685   1.565245e+02  147.0077667 214.735207
#> 686   6.545737e+01  -30.4047852  72.174226
#> 687  -6.036499e+01  -53.1237869  80.411870
#> 688  -3.602339e+01  -87.6679001  94.780511
#> 689   4.682958e+01  -84.4228058  96.541286
#> 690  -1.807995e+01  -44.7068672  48.224356
#> 691   4.164009e+01   19.2953682  45.893446
#> 692   6.323687e+01   95.3649902 114.426324
#> 693   3.897704e+01  106.5350037 113.441249
#> 694  -4.493397e+00   40.7663345  41.013225
#> 695   7.892725e-01  -11.7344007  11.760915
#> 696  -2.304741e+01  -43.0691986  48.848120
#> 697  -1.345257e+02  -69.3738403 151.360183
#> 698  -6.399656e+01  -58.8833084  86.964384
#> 699  -1.807578e+01  -84.8317413  86.736142
#> 700   1.856324e+01  -38.1652336  42.440299
#> 701   3.028219e+01    4.1532230  30.565670
#> 702   9.003262e+01   62.3315849 109.503880
#> 703   1.593166e+02  106.4792175 191.623613
#> 704   4.250034e+01  107.9500732 116.015072
#> 705  -1.019917e+02   79.7839203 129.490435
#> 706  -7.048432e+01   42.2256203  82.164728
#> 707   3.213329e+01  -25.4388294  40.983928
#> 708   5.367003e+01 -173.1994934 181.324397
#> 709  -8.511082e+00 -150.4446564 150.685212
#> 710  -5.568969e+01  -10.4473763  56.661183
#> 711   3.441684e+01   67.7825699  76.019706
#> 712  -4.480290e+00   49.4484138  49.650968
#> 713  -6.607039e+01   76.8891068 101.376679
#> 714  -1.754611e+01   38.8173485  42.598739
#> 715   6.842686e+01   10.1194649  69.171088
#> 716   7.996288e+00  -57.7444572  58.295480
#> 717  -9.970280e+01   18.5438766 101.412638
#> 718  -9.038129e+01   35.0918808  96.954724
#> 719   1.671553e+01  -22.8124657  28.281045
#> 720   8.367184e+01   26.2754116  87.700476
#> 721   9.289603e+01  -50.4395103 105.706272
#> 722   6.074386e+01 -106.5914841 122.684805
#> 723   1.650581e+01  -19.7254009  25.720289
#> 724  -5.677119e+01  108.0029984 122.014816
#> 725  -1.382309e+02   48.5574913 146.511426
#> 726  -5.580257e+01 -121.6001205 133.792812
#> 727   2.897504e+01 -106.8719940 110.730195
#> 728  -7.323146e+00   40.4024925  41.060807
#> 729  -3.312314e+00   57.0634270  57.159480
#> 730   5.440860e+01   69.6712952  88.399008
#> 731   1.544385e+02   36.2473183 158.635166
#> 732   2.298676e+01  -30.8975582  38.510394
#> 733  -6.609422e+01  -90.5692444 112.121511
#> 734   8.903667e+01 -110.1002655 141.596603
#> 735   1.186290e+02   90.7667770 149.370174
#> 736  -8.982115e+01  148.2295990 173.320089
#> 737  -2.877567e+02  -17.7745686 288.305123
#> 738  -1.146538e+02 -197.7063446 228.546033
#> 739   1.482851e+02  -39.1767578 153.373020
#> 740   1.519887e+02  149.1422577 212.941239
#> 741   2.829223e+01  124.4129333 127.589295
#> 742   5.244936e+00   38.8628998  39.215231
#> 743  -2.558686e+01   45.5086517  52.208474
#> 744  -9.124446e+01   72.0129776 116.238637
#> 745  -1.578871e+02 -107.9688568 191.273616
#> 746  -7.707381e+01  -82.2871246 112.745477
#> 747   9.740737e+01   55.1505547 111.936499
#> 748   1.178217e+02  101.6759338 155.627596
#> 749   2.803592e+01  -53.4536781  60.359825
#> 750  -5.482690e+01 -134.1469421 144.918569
#> 751  -7.622809e+00  -60.3966713  60.875817
#> 752   7.034856e+01  107.8648224 128.777872
#> 753   2.759586e+01  131.5903473 134.452783
#> 754   3.496671e+01  -75.9809189  83.640723
#> 755  -1.058985e+01 -170.5518341 170.880289
#> 756   2.953379e+01 -122.4974899 126.007459
#> 757  -2.983251e+01   18.9828491  35.359966
#> 758  -4.916851e+01  141.0785370 149.401123
#> 759  -2.213470e+01  132.3223877 134.160945
#> 760  -3.085068e-01   63.6835823  63.684330
#> 761   6.122696e+01   -0.8133193  61.232357
#> 762  -6.192348e+01  -57.0058136  84.167570
#> 763  -1.126620e+01   50.9877281  52.217581
#> 764   1.527667e+01   49.9719048  52.254836
#> 765   2.550493e+01   53.4106255  59.187804
#> 766  -4.520279e+01    2.0478697  45.249158
#> 767  -9.785250e+01  -82.4732361 127.972445
#> 768  -1.663992e+01 -113.0339966 114.252226
#> 769   3.643344e+01 -202.2257233 205.481480
#> 770   1.236871e+02 -128.7930450 178.566944
#> 771   8.010239e+01    4.2148919  80.213201
#> 772   4.798526e+01  134.2324524 142.551521
#> 773  -8.936810e-01   15.6653185  15.690789
#> 774  -1.153076e+00   26.4517441  26.476864
#> 775  -4.242599e+01   94.5168610 103.602132
#> 776  -9.147618e+01  142.9851837 169.742907
#> 777  -4.691989e+01   32.5496101  57.104757
#> 778  -1.392776e+01 -147.6772461 148.332571
#> 779  -4.539817e+00  -65.0445938  65.202831
#> 780  -2.877871e+01   -2.4838102  28.885698
#> 781   9.235214e+01   84.8222885 125.394333
#> 782   1.112303e+02   15.8196764 112.349642
#> 783  -2.779390e+01   -2.2276504  27.883032
#> 784  -1.040949e+02   29.0322208 108.067628
#> 785  -6.590548e+01   -0.4580450  65.907071
#> 786   3.321906e+01  -57.6259804  66.515107
#> 787   5.852657e+01  -23.6009979  63.105990
#> 788   7.574173e+01   12.4895563  76.764566
#> 789  -3.517686e+01  -34.5211716  49.286129
#> 790  -2.703833e+01  -44.9200211  52.429759
#> 791   6.921561e+01   23.0151958  72.941761
#> 792   1.098723e+02   70.1038818 130.332183
#> 793   5.853128e+01    9.9335060  59.368218
#> 794  -1.414832e+02   60.1315727 153.731265
#> 795  -3.674909e+01  220.2041321 223.249536
#> 796   1.290397e+01  122.1850281 122.864533
#> 797   1.701760e+01  -82.1988220  83.941915
#> 798   8.709734e+01 -224.4754028 240.780299
#> 799   9.434448e+00 -140.9251709 141.240620
#> 800  -9.588444e+00   40.5589447  41.676927
#> 801  -7.124131e+01   35.6578560  79.666850
#> 802  -6.374900e+01   20.7328854  67.035723
#> 803   2.619284e+01  -27.0368481  37.643808
#> 804   4.016644e+01   16.9637394  43.601735
#> 805  -2.614285e+01   46.0783882  52.977982
#> 806  -9.090321e+01  -26.0150051  94.552497
#> 807  -1.601896e+02 -107.3109512 192.811708
#> 808  -1.270974e+02  -95.9983368 159.277854
#> 809  -3.441019e+01  -86.3393860  92.943803
#> 810   9.440387e+01  -56.8129501 110.180769
#> 811   2.236777e+02   75.0647202 235.937367
#> 812   1.178218e+02  172.7035217 209.065730
#> 813   4.611620e-01  108.3363190 108.337300
#> 814  -2.237067e+01  -99.2934189 101.782268
#> 815   7.169881e+01 -163.8683777 178.867448
#> 816  -1.801186e+00  -42.9401703  42.977930
#> 817  -1.464374e+02    2.8505638 146.465150
#> 818  -5.174761e+01   -0.0433282  51.747626
#> 819   6.438335e+01   35.9451904  73.737867
#> 820   4.505501e+01   93.7615204 104.024884
#> 821  -9.631248e+01  206.5853729 227.933348
#> 822  -4.597303e+01  148.3220673 155.283468
#> 823   1.327110e+02   78.2876358 154.081677
#> 824   1.584589e+02  -82.0757446 178.453491
#> 825   2.556601e+01 -187.1475372 188.885737
#> 826  -1.672628e+02  -74.8284225 183.237899
#> 827  -1.418025e+02   84.3077164 164.971903
#> 828   3.749918e+00  184.2775879 184.315738
#> 829   7.201454e+01   69.6851120 100.210324
#> 830  -1.058063e+01  -87.1883392  87.827993
#> 831  -1.395194e+02 -141.6657410 198.833719
#> 832  -1.395691e+02  -92.8726654 167.645090
#> 833  -4.198456e+01  -44.7880974  61.389549
#> 834   5.102768e+01   44.5109367  67.712976
#> 835   8.966049e+01  127.4371262 155.817922
#> 836   3.513811e+01  132.6726074 137.246886
#> 837  -3.874660e+01  -11.5989084  40.445442
#> 838  -6.476112e+01 -195.0466156 205.516874
#> 839  -3.776265e+01 -187.3270111 191.095335
#> 840   7.585118e+01   19.1321926  78.226865
#> 841   9.255133e+01  175.6761780 198.564519
#> 842   3.993585e+01  138.9606934 144.585429
#> 843  -6.397542e+00  -11.9558363  13.559888
#> 844  -2.471494e+01  -73.1423111  77.205090
#> 845   9.229259e+01  -16.6600227  93.784217
#> 846   1.345452e+02  -24.9495525 136.838965
#> 847  -5.062568e+01 -111.1715317 122.155922
#> 848  -2.391955e+02 -122.5939560 268.781972
#> 849  -1.450365e+02  -18.1161976 146.163548
#> 850   7.463573e+01   88.5284348 115.791949
#> 851   1.459736e+02   34.9961052 150.109982
#> 852   7.124388e+00  -66.1801224  66.562493
#> 853   3.178614e+01   -8.5601082  32.918599
#> 854   7.698024e+01   41.7319336  87.564328
#> 855   6.357847e+01    3.3564494  63.667008
#> 856   7.246803e+01  -46.7385979  86.232889
#> 857  -7.735280e+01   -6.6148643  77.635120
#> 858  -1.149004e+02   58.7248154 129.037619
#> 859  -7.110027e+01   17.3411465  73.184446
#> 860   1.354946e+02   -9.4633551 135.824671
#> 861   1.516405e+02  -11.9037333 152.106959
#> 862   4.094693e+01   15.5015869  43.782994
#> 863  -1.118093e+02   -6.5871739 112.003144
#> 864  -2.012688e+02  -26.5016861 203.006111
#> 865  -7.166594e+01  -64.1836700  96.205771
#> 866   1.482851e+02   34.2343483 152.185581
#> 867   1.815857e+02   60.5407753 191.411989
#> 868  -7.215111e+01   12.0937128  73.157640
#> 869  -1.831782e+02   23.3610668 184.661794
#> 870  -7.494941e+01  -72.9858093 104.615211
#> 871   3.972813e+01 -107.0618896 114.195326
#> 872   7.543478e+01 -111.1721497 134.348999
#> 873   1.118740e+02   74.6948318 134.518065
#> 874   6.919546e+01  141.2589569 157.296231
#> 875  -4.945079e+01  -47.2420540  68.390002
#> 876  -1.043300e+02 -171.6940765 200.906957
#> 877   4.076939e-02  -93.9420547  93.942064
#> 878   1.406201e+02    4.8475795 140.703663
#> 879   2.910049e+01   99.4820099 103.650899
#> 880  -5.993083e+01   66.3456726  89.406112
#> 881  -3.767810e+01   14.9770088  40.545658
#> 882  -3.941269e+01   -0.4628718  39.415411
#> 883   7.168900e+00    4.8751774   8.669514
#> 884  -1.715420e+00   -2.5112808   3.041249
#> 885   3.075247e+01  -36.9990692  48.110763
#> 886   6.139113e+01  -40.5647697  73.582411
#> 887  -2.206813e+01   25.6944618  33.870455
#> 888  -7.727836e+01  120.4255600 143.088295
#> 889  -1.081633e+02  114.1258392 157.238661
#> 890   2.383978e+00   32.3820839  32.469720
#> 891   1.268767e+02  -53.8770485 137.842082
#> 892   4.486521e+01  -19.9227982  49.089766
#> 893  -7.352500e+01   37.9188347  82.727044
#> 894  -1.181046e+02    4.5778122 118.193255
#> 895   3.218595e+01  -85.9029083  91.734644
#> 896   7.860751e+01  -57.9162636  97.639303
#> 897  -1.248060e+01   55.6945152  57.075778
#> 898  -5.508042e+01   59.8758888  81.357080
#> 899  -1.688630e+01    3.1799116  17.183100
#> 900   9.476871e+01   -6.4427404  94.987464
#> 901   7.219846e+01   -5.6699190  72.420750
#> 902  -4.797197e+01    0.9572620  47.981523
#> 903  -1.005056e+01   12.8327980  16.300135
#> 904   1.032603e+02   44.9438438 112.617222
#> 905   1.189930e+02    4.5472884 119.079867
#> 906  -4.442927e+01  -47.9363441  65.359416
#> 907  -9.768587e+01   16.7097340  99.104719
#> 908   4.547457e+01   78.3960571  90.630448
#> 909   1.661420e+02    4.6545486 166.207216
#> 910   2.465508e+01  -19.3669052  31.352034
#> 911  -1.708458e+02   34.4733200 174.289099
#> 912  -1.707279e+02   34.6597023 174.210526
#> 913  -5.369510e+01  -18.4078426  56.762776
#> 914   3.545082e+01 -115.1316147 120.465968
#> 915  -3.901313e+01 -142.4158783 147.662813
#> 916  -8.096169e+01  -57.2700882  99.169842
#> 917   4.992603e+01   83.7101059  97.467895
#> 918   2.367888e+02  166.7196503 289.593460
#> 919   2.027451e+02   50.8171463 209.016662
#> 920  -1.138923e+02 -115.4136353 162.147342
#> 921  -2.355452e+02 -145.1713715 276.687992
#> 922  -1.626031e+02  -86.6688385 184.258695
#> 923  -1.586512e+00   62.7080460  62.728112
#> 924   1.428510e+02  123.3558655 188.740756
#> 925   6.932664e+01   26.1728344  74.102631
#> 926   6.529319e+01  -49.7110367  82.063316
#> 927  -5.462029e+01  -93.8123245 108.554722
#> 928  -3.725376e+01   -5.0769711  37.598113
#> 929  -1.350148e+00   59.0089760  59.024420
#> 930  -4.597263e+01   64.5832977  79.274746
#> 931   4.140003e+01   88.9479446  98.110648
#> 932  -1.279776e+01   46.0500908  47.795328
#> 933  -5.362469e+01   69.4693756  87.758772
#> 934  -1.224115e+02  -15.9156370 123.441796
#> 935  -2.448929e+01  -48.6269951  54.445475
#> 936   8.842008e+01  -29.4286976  93.188836
#> 937   6.797895e+01 -110.3776016 129.631604
#> 938   1.924219e+01 -142.5458221 143.838706
#> 939  -3.040794e+01 -123.3772430 127.069221
#> 940  -2.065112e+01  224.0783539 225.027949
#> 941  -6.241170e+01  199.7118378 209.236800
#> 942  -5.811828e+01  -45.8387413  74.019760
#> 943   4.495883e+00 -159.2885895 159.352024
#> 944   8.541526e+01 -146.5348053 169.611957
#> 945   1.726601e+02   99.3712616 199.213819
#> 946   7.795827e+01   87.0147324 116.829170
#> 947  -3.009357e+01   84.2138443  89.429273
#> 948   2.758305e+01   59.7101440  65.773292
#> 949  -1.089125e+01  -18.7152710  21.653651
#> 950  -7.665881e+01   -5.3573184  76.845777
#> 951  -9.490040e+01  -70.1376724 118.005842
#> 952  -2.251971e+01   -8.8868170  24.209768
#> 953   2.249284e+00   54.9077377  54.953789
#> 954  -5.204224e+01   72.7678986  89.462628
#> 955  -1.010029e+02  -22.4002399 103.456990
#> 956  -1.036398e+02  -67.3768387 123.615722
#> 957  -8.451363e+01  -36.5369225  92.073346
#> 958   8.457207e+01   -2.4565170  84.607736
#> 959   1.948396e+02   20.0943623 195.873084
#> 960   1.531706e+02  -21.5317860 154.676623
#> 961   3.455317e+00   14.7096748  15.110054
#> 962  -1.187972e+01   54.4115105  55.693270
#> 963   1.555015e+01   35.3499527  38.618989
#> 964  -9.890112e+00  -77.7236328  78.350350
#> 965  -2.204554e+01  -48.0999413  52.911342
#> 966  -8.646821e+01   12.9761591  87.436444
#> 967  -6.274525e+01   47.0191917  78.407725
#> 968  -2.434960e+01  -89.8457260  93.086828
#> 969   5.716151e+01 -103.6253891 118.345509
#> 970   1.535614e+02  -25.0184078 155.586114
#> 971   1.304081e+02   31.2571526 134.101757
#> 972   5.344363e+01   31.2388992  61.903881
#> 973  -1.836301e+01  -39.9208565  43.941723
#> 974  -1.054439e+02   26.3000278 108.674275
#> 975  -1.505524e+02   84.2556458 172.525512
#> 976  -1.292766e+02  165.8721313 210.299791
#> 977  -5.240113e+01   48.8690186  71.652349
#> 978  -1.015651e+01  -44.2694473  45.419584
#> 979   9.947254e+01  -88.5041809 133.145697
#> 980   2.017315e+02 -122.3038254 235.910658
#> 981   2.048510e+02   -5.3947649 204.921991
#> 982   5.086765e+01   38.4265747  63.750443
#> 983  -8.354934e+01  152.8998108 174.237896
#> 984  -1.361895e+02   47.1110992 144.107692
#> 985  -1.191748e+02  -85.3892593 146.608203
#> 986  -7.551231e+01 -109.9552231 133.387633
#> 987  -2.063241e+01  -71.2107849  74.139545
#> 988   4.936845e+01   78.8717041  93.048316
#> 989   6.689607e+01   17.7613487  69.213799
#> 990   5.192602e+01  -82.1692352  97.201311
#> 991  -8.123584e-01  -86.3327408  86.336563
#> 992   5.271795e+01  -45.1110153  69.384331
#> 993   6.335422e+01   13.2860241  64.732336
#> 994   3.253377e+00   66.4929657  66.572509
#> 995  -4.323817e+01  134.5808868 141.356125
#> 996   4.266310e+01   96.2644119 105.294716
#> 997  -1.582138e+01  -44.9278793  47.632241
#> 998  -2.086428e+02  -68.5518188 219.615952
#> 999  -5.204487e+02   48.5835152 522.711430
#> 1000            NA           NA         NA