Differentiate x and y and compute change amplitude via an algorithm proposed by Engbert & Kliegl (2003)

Differentiate x and y and compute change amplitude via an algorithm proposed by Engbert & Kliegl (2003)

Usage

diff_ek(x, y, trial, sample_rate, options = NULL)

Arguments

x

vector with x coordinates in degrees of visual angle

y

vector with y coordinates in degrees of visual angle

trial

vector with trial index

sample_rate

sample rate in Hz

options

List with method specific options, see Details.

Value

data.frame with columns x, y, and amp

Details

Method options, please refer to Engbert & Kliegl (2003) for details on parameters and the rationale for default values.

ek_velocity_time_window Time window for velocity computation in milliseconds. Defaults to 20 ms.

See also

Examples

diff_ek(rnorm(1000), rnorm(1000), rep(1, 1000), 250, list("ek_velocity_time_window" = 20))
#>                  x            y         amp
#> 1               NA           NA          NA
#> 2     1.942911e+02 -226.6677551 298.5419463
#> 3     1.949800e+01  118.2524490 119.8491291
#> 4    -7.963194e+01  129.7404785 152.2295532
#> 5    -6.510947e+01  -47.6252327  80.6684971
#> 6    -5.977929e+01 -205.3756256 213.8988336
#> 7    -5.187492e+01  -69.9593277  87.0937153
#> 8    -2.907639e+01  109.5439301 113.3371476
#> 9    -1.267512e+01  116.3776016 117.0658138
#> 10    9.098618e+01   37.4146652  98.3785683
#> 11    1.396289e+02  -23.5141392 141.5949776
#> 12    5.426661e+01 -101.2091293 114.8396846
#> 13   -3.639929e+01  -95.1249466 101.8511852
#> 14   -2.880618e+01  -23.4710579  37.1575932
#> 15   -1.514239e+01   31.2953739  34.7662543
#> 16   -7.839505e+01   28.4017620  83.3813166
#> 17   -6.689933e+01  -90.7045517 112.7068590
#> 18    9.840749e+00  -47.2636681  48.2772685
#> 19   -2.545878e+01   72.0447540  76.4107072
#> 20   -4.290066e+01  111.9206161 119.8611311
#> 21    7.926111e+01   32.7559509  85.7629036
#> 22    1.885235e+02  -51.3015289 195.3790512
#> 23    1.011220e+02   35.8017845 107.2726687
#> 24   -4.022456e+01  -11.8369102  41.9300292
#> 25   -6.869274e+01  -79.2752380 104.8964064
#> 26   -4.434865e+00 -134.0217133 134.0950695
#> 27   -1.585817e+01  -19.9862061  25.5133299
#> 28    4.538887e+00  124.4784546 124.5611783
#> 29    1.289823e+01  -11.1488752  17.0488027
#> 30    6.279519e+01   33.7128487  71.2726556
#> 31    1.824497e+02  -43.4951630 187.5625445
#> 32    8.172909e+01  -50.2427902  95.9373846
#> 33   -1.097682e+02    2.6740620 109.8008084
#> 34   -1.953963e+02   45.1495056 200.5446946
#> 35   -1.470747e+02  218.2440948 263.1756930
#> 36   -7.383127e+01   83.2152786 111.2467473
#> 37   -1.185706e+02  -93.8872452 151.2408879
#> 38   -2.865316e+01 -106.1932602 109.9909637
#> 39    1.408537e+02   -5.4810824 140.9603324
#> 40    2.363353e+02   80.6486588 249.7169970
#> 41    1.990540e+02  -12.8159819 199.4661948
#> 42    1.609903e+01  -99.7815018 101.0718898
#> 43   -1.024462e+02  -54.5688515 116.0731456
#> 44   -1.144031e+02  108.3774719 157.5872631
#> 45   -9.119329e+01  172.5417328 195.1585658
#> 46   -1.517512e+01  106.0449600 107.1252430
#> 47    2.722318e+01  -56.1411667  62.3933680
#> 48    5.750349e+01 -189.8050232 198.3244772
#> 49    5.705889e+01  -79.9860382  98.2521419
#> 50   -8.775314e+01  -62.8808250 107.9565236
#> 51   -9.186976e+01  -84.1137772 124.5599456
#> 52   -5.637064e+01  -37.8035088  67.8730781
#> 53    5.875884e+01    9.0359278  59.4495510
#> 54    8.494544e+01  105.4831543 135.4342054
#> 55   -5.533033e+01  121.5331573 133.5355882
#> 56   -6.898538e+01   63.1567039  93.5294189
#> 57   -2.253870e+01   -4.6174974  23.0068275
#> 58    1.348419e+01  -90.3716812  91.3721185
#> 59    2.575523e+01   16.2573910  30.4570917
#> 60    2.476979e+01   97.5235596 100.6200139
#> 61    6.427335e+01   24.3350124  68.7259544
#> 62    8.819137e+01  -88.1748810 124.7097712
#> 63   -3.592744e+01 -156.9847717 161.0434701
#> 64   -1.354687e+02 -106.3437424 172.2230229
#> 65   -5.037474e+01   86.2854309  99.9139114
#> 66    1.143280e+02  230.8041229 257.5683324
#> 67    1.871904e+02   52.3932190 194.3844146
#> 68    9.241842e+01 -113.6743011 146.5025969
#> 69   -6.171070e+01  -54.9005470  82.5971013
#> 70   -1.402437e+02   62.8035202 153.6638024
#> 71   -1.560012e+02   10.5912933 156.3603414
#> 72   -7.083449e+00  -44.3217239  44.8841894
#> 73    6.093084e+01   26.1471634  66.3041614
#> 74    6.991808e+01   78.2171249 104.9116577
#> 75    7.143862e+01   27.7299328  76.6317546
#> 76    2.838554e+01  -37.4930573  47.0262515
#> 77   -9.583503e+01   34.9980278 102.0255598
#> 78   -1.266313e+02   33.0189857 130.8653065
#> 79   -4.351649e+01  -30.1654320  52.9493903
#> 80    4.974553e+01 -109.0208969 119.8339435
#> 81    4.770096e+01 -118.9860229 128.1914796
#> 82    7.295406e+01   17.6774540  75.0652159
#> 83    1.184097e+02   51.2121162 129.0098211
#> 84    1.150607e+01   65.0504456  66.0601999
#> 85   -6.536514e+01   42.4476013  77.9384357
#> 86   -3.324064e+01   24.0241966  41.0134410
#> 87    5.079716e+00    5.0822244   7.1855773
#> 88   -1.930568e+01  -22.2444630  29.4537864
#> 89   -5.472947e+01  -67.6806717  87.0401524
#> 90   -3.462893e+01 -124.4966965 129.2230249
#> 91    8.713361e+01  -53.3799782 102.1845749
#> 92    2.970331e+01  -34.8074226  45.7585350
#> 93   -3.259597e+01   73.8979263  80.7675711
#> 94   -8.934219e+01   93.8951035 129.6083254
#> 95   -2.468111e+01  123.7823792 126.2189942
#> 96    5.962849e+01   73.7925873  94.8730855
#> 97   -5.002208e+01 -101.7210388 113.3550994
#> 98   -8.378374e+01  -77.7894821 114.3281162
#> 99    3.563367e+01   71.9093704  80.2540736
#> 100   1.577951e+02  195.0679474 250.8999498
#> 101   7.045202e+01   76.7813873 104.2058942
#> 102  -5.894395e+01 -115.5532303 129.7186875
#> 103  -7.962891e+01 -187.4869232 203.6961225
#> 104  -1.353392e+01  -40.6991272  42.8903937
#> 105   5.349465e+01   43.8218575  69.1522448
#> 106   1.160947e+02   48.6735268 125.8852000
#> 107   5.091508e+01  -30.7907715  59.5014041
#> 108  -1.340537e+02   -5.1128922 134.1511800
#> 109  -1.571455e+02  145.6993256 214.2964995
#> 110  -7.647783e+01   85.6386719 114.8165513
#> 111   3.264860e+01  -42.5894966  53.6637345
#> 112   7.020068e+01 -191.4910583 203.9533288
#> 113   7.197478e+01 -157.7862701 173.4268627
#> 114   4.382860e+01  -14.8266096  46.2685065
#> 115  -5.951248e+00  114.3887711 114.5434778
#> 116   1.650301e+01   90.6433945  92.1334585
#> 117   1.725623e+01  -62.7362976  65.0662767
#> 118  -1.024418e+01 -129.8392181 130.2427186
#> 119   8.186534e+01  -40.2145042  91.2093221
#> 120   6.871758e+01   92.4133224 115.1621828
#> 121   4.186343e+01  100.3609161 108.7421719
#> 122   5.634542e+01   67.3691330  87.8259990
#> 123  -6.500109e+01  -12.4329777  66.1794588
#> 124  -1.344930e+02 -165.6433868 213.3684840
#> 125  -9.980894e+01 -104.3406754 144.3911370
#> 126   1.082243e+02   -8.0957222 108.5266898
#> 127   2.074817e+02   16.0170860 208.0990417
#> 128   6.441428e+01   62.7587051  89.9324971
#> 129  -1.113048e+02  122.2528610 165.3315794
#> 130  -9.017471e+01  222.3900604 239.9767028
#> 131  -1.116627e+02    1.7429309 111.6763139
#> 132   7.315447e-01 -155.7765198 155.7782375
#> 133   6.613989e+01  -81.6272812 105.0594994
#> 134  -2.681840e+01  110.9640503 114.1588678
#> 135  -6.119373e+01  138.1143646 151.0637306
#> 136  -8.068140e+01  -28.7339611  85.6453705
#> 137   2.590256e+01  -76.7496796  81.0028147
#> 138   1.199910e+02  -58.5392227 133.5090598
#> 139   6.932358e+01  -49.9712143  85.4568939
#> 140  -8.255393e+01  -88.9463501 121.3532238
#> 141  -6.184631e+01  -75.8233032  97.8475286
#> 142  -1.439809e+01   40.3651772  42.8561856
#> 143   4.303059e+01   45.4970703  62.6228027
#> 144  -8.826591e+01    3.1603091  88.3224657
#> 145  -7.936852e+01   10.5563049  80.0674512
#> 146   9.698493e+01    7.4727640  97.2723970
#> 147   1.386231e+02   38.4976006 143.8695087
#> 148  -1.357836e+01  -22.4472733  26.2345550
#> 149  -1.170205e+02  -17.6346149 118.3418035
#> 150   1.633232e+01  -31.9161663  35.8522874
#> 151   6.636964e+01   64.8258286  92.7756256
#> 152   2.511304e+01   82.7701263  86.4960036
#> 153  -7.968645e+01  -77.9433823 111.4679412
#> 154   2.548952e+01 -133.9728699 136.3761172
#> 155   4.374937e+01  -42.4744072  60.9760830
#> 156  -1.385094e+02   57.2934837 149.8913263
#> 157  -2.213385e+02   70.8855362 232.4123157
#> 158  -4.079826e+01   22.8981895  46.7848776
#> 159   1.475751e+02   13.4437265 148.1861989
#> 160   9.255518e+01   40.1265373 100.8791334
#> 161  -2.293813e+01   73.5564804  77.0500725
#> 162  -4.381758e+01   71.4247742  83.7942646
#> 163   9.972879e+01 -114.2588043 151.6604958
#> 164   2.003090e+01 -166.3202820 167.5221573
#> 165  -3.443312e+01    6.0453911  34.9597886
#> 166  -3.421206e+01  195.7852020 198.7518813
#> 167   9.386350e+00  147.9928894 148.2902521
#> 168   1.550853e+02    4.5925198 155.1533263
#> 169   1.967295e+01  -56.3053551  59.6432542
#> 170  -8.412504e+01  -40.6825523  93.4456639
#> 171  -1.331386e+02 -107.9018936 171.3730094
#> 172  -1.020738e+02 -169.3147278 197.7031590
#> 173   2.652070e+01  -99.3774948 102.8554045
#> 174   3.504498e+01   15.4172115  38.2863027
#> 175   2.892323e+01   88.9940567  93.5761483
#> 176   5.117219e+01   50.5160141  71.9059141
#> 177   5.194378e+01   44.0857544  68.1300926
#> 178   2.037246e+01   64.8612366  67.9854172
#> 179   3.765120e+00   70.4779892  70.5784889
#> 180  -1.536491e+01   -8.9409828  17.7769966
#> 181   2.992649e+01  -65.1829758  71.7245794
#> 182  -2.571058e+01  -29.3348160  39.0072442
#> 183  -4.668436e+01   40.5049629  61.8068082
#> 184  -5.954927e-01  122.0787048 122.0801572
#> 185   4.115493e+01   35.3866539  54.2765498
#> 186  -5.547592e+00  -87.1313019  87.3077290
#> 187  -1.217166e+02 -204.5113983 237.9912459
#> 188  -1.142257e+02  -32.1749191 118.6706896
#> 189  -2.411865e+01  156.2894897 158.1395399
#> 190   2.181605e+02  150.8402252 265.2296874
#> 191   2.470942e+02   14.8457384 247.5397816
#> 192   9.120540e+01  -98.6938095 134.3833798
#> 193  -5.545272e+01 -104.6513672 118.4352687
#> 194  -1.076363e+02  -83.0902786 135.9763724
#> 195  -2.896713e+01   69.9634628  75.7230518
#> 196  -4.589574e+01   94.7525787 105.2828104
#> 197  -4.383566e+01  -35.0744514  56.1407322
#> 198   6.636177e+01 -111.9366608 130.1295532
#> 199   3.236037e+01  -73.0584030  79.9044670
#> 200  -9.860452e+01   38.8926659 105.9976007
#> 201  -1.086159e+02   11.7313004 109.2476392
#> 202   6.049266e+01  -21.0898762  64.0636037
#> 203   2.098860e+02   50.2307816 215.8129882
#> 204   4.444817e+01  145.2005920 151.8514122
#> 205  -1.336524e+02   30.2048626 137.0229426
#> 206  -9.714169e+01 -100.7403412 139.9468592
#> 207  -5.225045e+01  -35.6728516  63.2665937
#> 208  -7.823660e+00   82.2642136  82.6354070
#> 209   1.568778e+01   95.9793625  97.2529923
#> 210  -5.527181e+00  -87.4887772  87.6631956
#> 211   7.463386e+01  -93.0836792 119.3096143
#> 212   1.882050e+01 -102.5211792 104.2343683
#> 213  -2.502934e+00  -28.9270477  29.0351299
#> 214  -2.525361e+01  106.2044983 109.1656549
#> 215  -6.939866e+01  173.9973450 187.3265862
#> 216  -7.850441e+00   50.0574150  50.6692631
#> 217  -6.916499e+01  -81.3094254 106.7474541
#> 218  -2.934841e+01  -35.4689789  46.0366973
#> 219   3.761943e+01   -5.0757728  37.9603123
#> 220   1.752052e+02   17.2460670 176.0519649
#> 221   1.178753e+02   30.2494488 121.6948056
#> 222   3.409370e+00  117.3880768 117.4375765
#> 223  -1.183274e+02  120.2909851 168.7343697
#> 224  -1.079076e+02   -5.8086767 108.0638510
#> 225   1.190641e+01 -118.7763214 119.3715929
#> 226   1.481847e+01 -135.3277435 136.1366418
#> 227   2.642671e+01  -39.5298615  47.5497745
#> 228  -5.725002e+01   34.0360489  66.6034332
#> 229  -1.100776e+02  -36.5797539 115.9963190
#> 230  -7.004403e+01  -92.8243790 116.2864195
#> 231   3.825877e+01 -109.0801620 115.5950497
#> 232   1.433036e+02  -53.1895905 152.8562944
#> 233   1.559551e+02   18.7836761 157.0822154
#> 234   1.250441e+02   17.3694839 126.2447498
#> 235   8.206137e+00    6.9776201  10.7716230
#> 236  -1.211944e+02  -72.1773148 141.0590313
#> 237  -1.065771e+02  -10.6253948 107.1054291
#> 238  -4.561683e+01    8.5207586  46.4058014
#> 239  -1.617049e+01   -7.0853682  17.6546633
#> 240  -7.070979e+01   32.7864647  77.9411703
#> 241  -1.996091e+01   93.7833481  95.8840672
#> 242   8.508744e+01  138.1031952 162.2108660
#> 243   8.949337e+01   36.8614044  96.7875324
#> 244   1.416337e+01   -1.0920675  14.2054107
#> 245  -2.290032e+01   11.4293299  25.5940240
#> 246  -3.010241e+01  -15.4777164  33.8484068
#> 247  -1.540105e+01  -45.5573082  48.0901300
#> 248  -9.088123e+00  -45.4196243  46.3199337
#> 249  -1.657284e+01    1.3808309  16.6302668
#> 250   4.250486e+01   41.8683701  59.6625779
#> 251   6.578111e+01   15.9771519  67.6935976
#> 252   6.169813e+01   10.6784916  62.6154067
#> 253   1.271885e+01  -29.9988785  32.5837687
#> 254  -3.584288e+01  -49.1306915  60.8155999
#> 255   3.508786e+01  -48.1189766  59.5532834
#> 256  -4.179901e+01  -15.4350185  44.5577919
#> 257  -6.600870e+01  -35.2244263  74.8191710
#> 258  -1.336316e+01  -30.0228786  32.8625534
#> 259   5.506564e+01   77.2885742  94.8986193
#> 260   5.238438e+01  141.3302917 150.7261566
#> 261   6.749939e+00  177.4051819 177.5335468
#> 262   7.986677e+01  -21.8336506  82.7973968
#> 263   4.010643e+01  -76.6423645  86.5018964
#> 264  -1.451648e+02  -81.7384186 166.5953182
#> 265  -1.276283e+02   -3.8823304 127.6873230
#> 266   2.272544e-04    0.3125408   0.3125409
#> 267   1.274518e+02  -59.3659477 140.5997305
#> 268   7.854563e+01  -57.4281158  97.3005894
#> 269  -6.436288e+01 -122.8808517 138.7165622
#> 270  -4.802248e+01  -68.0849533  83.3169840
#> 271  -4.791022e+01  -61.3754196  77.8609726
#> 272  -2.077183e+01   31.5131168  37.7431509
#> 273   1.770262e+00   94.5551605  94.5717305
#> 274  -3.936074e+00   71.8992844  72.0069425
#> 275   9.716781e+01   42.9787369 106.2485522
#> 276   6.777562e+01  -85.8824310 109.4044175
#> 277   3.457418e+00  -63.5964508  63.6903627
#> 278  -8.209726e+01   -2.3639741  82.1312875
#> 279  -1.856687e+02   87.2985916 205.1680336
#> 280  -7.292150e+01  203.9966736 216.6383811
#> 281  -2.815121e+01  154.4947357 157.0385750
#> 282   9.500441e+01   45.1644058 105.1934477
#> 283   8.830381e+00  -87.8747101  88.3172707
#> 284   5.217266e+00   -3.6096184   6.3442266
#> 285   2.829544e+01   55.3577347  62.1700132
#> 286   4.437647e+01  -89.3913803  99.8002514
#> 287   8.570798e+01 -194.6124573 212.6496349
#> 288   1.002338e+01  -60.2333374  61.0616335
#> 289  -1.152123e+00  161.7798462 161.7839486
#> 290  -1.017675e+02  128.3875122 163.8291113
#> 291   1.172324e+01  -74.3841095  75.3022575
#> 292   9.216308e+01 -122.1307678 153.0031289
#> 293   1.472407e+02  -50.6200180 155.6991366
#> 294   7.911345e+01  -87.7687302 118.1621252
#> 295  -4.178585e+01 -116.7276917 123.9814945
#> 296  -3.150957e+01  -43.0625114  53.3594693
#> 297  -5.333926e+01  101.8836670 115.0015594
#> 298  -8.603617e+01   37.5424881  93.8704486
#> 299  -1.039408e+02  -62.4406776 121.2539915
#> 300  -6.666117e+01   44.6897850  80.2551469
#> 301  -4.224271e+01  183.7572174 188.5501556
#> 302   1.309520e+01  158.8568878 159.3957184
#> 303  -5.688392e+00  -22.6753407  23.3779571
#> 304  -3.278784e+01  -60.3316574  68.6655023
#> 305  -5.673268e+01   25.8090858  62.3274065
#> 306  -1.025268e-01   30.0614624  30.0616372
#> 307   1.234797e+02  -69.7581100 141.8218307
#> 308   1.334220e+02  -56.1485291 144.7552786
#> 309  -1.202232e+00    8.7067356   8.7893462
#> 310  -2.954055e+01   47.4016113  55.8529954
#> 311  -3.824674e+01  -31.2592068  49.3958602
#> 312  -8.674804e+01  -56.6382141 103.6007221
#> 313   3.395959e+00  -27.2840385  27.4945685
#> 314   6.141967e+01  -43.9424095  75.5202734
#> 315   1.593155e+02    6.0133448 159.4289521
#> 316   5.116989e+01   66.6616516  84.0365015
#> 317  -3.272198e+01  116.7679749 121.2661876
#> 318  -3.964524e+00  119.2257919 119.2916884
#> 319  -1.138174e+02   85.8927155 142.5901591
#> 320  -8.219290e+01  -12.9042625  83.1997180
#> 321  -6.662283e+01 -111.9623337 130.2849419
#> 322   5.323613e+01 -119.9835663 131.2636343
#> 323   9.113525e+01  -32.1927605  96.6540653
#> 324   5.192378e+01   45.6248436  69.1209485
#> 325  -7.936331e-03  112.5258102 112.5258105
#> 326  -1.134202e+02  110.5576019 158.3891711
#> 327  -8.874147e+01  -30.5081062  93.8391875
#> 328   3.138954e+00 -132.4848175 132.5219978
#> 329   9.797599e+01  -74.5436478 123.1099106
#> 330   3.270847e+01    1.3023924  32.7343924
#> 331   1.306784e+01   43.0633774  45.0024777
#> 332   1.710515e+01   45.8868141  48.9712760
#> 333   4.059669e+01   38.9458084  56.2571535
#> 334   1.420526e+01  -24.4098530  28.2423521
#> 335  -3.963359e+01 -108.2304382 115.2590516
#> 336  -3.521107e+00    4.8439021   5.9884539
#> 337   3.314233e+01  105.5359497 110.6175889
#> 338   8.686295e+01   -4.9078398  87.0014915
#> 339  -3.448338e+01 -186.0246277 189.1937249
#> 340  -7.707224e+01 -100.6412430 126.7627327
#> 341  -1.355426e+01  143.4120026 144.0511032
#> 342   8.016359e+01  156.9747009 176.2590645
#> 343   5.006249e+01  -33.4783516  60.2250179
#> 344  -1.241998e+01 -176.9863586 177.4216080
#> 345   7.437883e+01  -98.5186157 123.4428126
#> 346   6.785950e+01  -50.8880424  84.8204233
#> 347  -2.508695e+01  -37.8430786  45.4032363
#> 348  -8.570562e+01   58.3366432 103.6755381
#> 349  -5.911472e+01  154.6917114 165.6021621
#> 350  -6.976183e+01  169.1079559 182.9322665
#> 351   9.249398e+00  -23.7741241  25.5100048
#> 352   1.452460e+02 -116.4085159 186.1380053
#> 353   1.931009e+02  -56.5560646 201.2126946
#> 354   3.283139e+01   -2.6933157  32.9416817
#> 355  -2.063441e+02  -43.0526199 210.7876537
#> 356  -1.559150e+02 -120.2350159 196.8906788
#> 357  -5.753280e+01   61.8640289  84.4818409
#> 358  -1.613657e+01  153.9046936 154.7483236
#> 359  -5.980803e+01   98.2416534 115.0148809
#> 360  -5.068577e+01   67.5611267  84.4603631
#> 361   6.311614e+01   43.8705482  76.8652875
#> 362   5.495116e+01    8.5418453  55.6110884
#> 363   6.985093e+00 -118.1603088 118.3665920
#> 364  -2.692236e+01 -152.8172302 155.1706132
#> 365   1.619859e+01  -73.5156097  75.2790753
#> 366   5.433282e+01   -2.1751702  54.3763402
#> 367   1.984125e+01   23.5878372  30.8230660
#> 368  -5.080874e+01  108.4720078 119.7819051
#> 369  -6.946506e+01   58.6099205  90.8873862
#> 370  -7.326160e+01    1.2297732  73.2719175
#> 371   2.651793e+01  -26.7561092  37.6708076
#> 372   7.042794e+01  -34.3594742  78.3624161
#> 373   7.950331e+01   35.1068306  86.9095279
#> 374   2.293193e+01   35.1543083  41.9725962
#> 375  -5.165907e+01   68.7761688  86.0163962
#> 376  -9.071980e+00    6.4282293  11.1185856
#> 377  -1.270916e+01  -84.2661362  85.2191556
#> 378   2.926116e+01 -164.5029144 167.0850811
#> 379   4.983919e+00  -37.4601250  37.7902158
#> 380  -4.651476e+01  187.4206390 193.1064950
#> 381  -5.778532e+01  179.8825836 188.9361974
#> 382   1.542132e+01   29.6096458  33.3848504
#> 383   7.705057e+01 -122.0149612 144.3067592
#> 384   7.238574e+01  -91.8547668 116.9486805
#> 385   6.210113e+01  -73.1942215  95.9892941
#> 386   2.918397e+01  -44.2533340  53.0100125
#> 387  -2.170260e+01  -18.1974106  28.3222284
#> 388  -1.220308e+02    0.3765084 122.0313960
#> 389  -1.106378e+02    3.9073048 110.7067838
#> 390  -2.511653e+01   29.6796436  38.8808603
#> 391   3.394117e+01  105.0006180 110.3500454
#> 392  -1.006509e+01   84.1609268  84.7606491
#> 393  -4.995179e+01   23.9694805  55.4050296
#> 394   3.525787e+01  -99.2678909 105.3433982
#> 395   1.195602e+02  -45.7634506 128.0192643
#> 396   1.033956e+02  -13.8976402 104.3254178
#> 397   5.265039e+01  -58.3729553  78.6095767
#> 398   1.908812e+01 -131.7488098 133.1243974
#> 399  -1.788610e+01  -69.1523590  71.4280147
#> 400  -9.821828e+01   32.1528969 103.3471829
#> 401  -8.461048e+01   21.0251331  87.1836554
#> 402   3.937868e+01  -15.4273148  42.2928158
#> 403   9.751562e+01  -31.8655415 102.5900085
#> 404   5.609810e+01   36.3305893  66.8349340
#> 405  -8.520304e+01  129.2657318 154.8198554
#> 406  -1.109298e+02  160.0115509 194.7026427
#> 407  -6.143563e+01  130.0224152 143.8060000
#> 408   3.500393e+01   14.8874846  38.0383033
#> 409  -6.874509e+00 -124.4889755 124.6786425
#> 410  -2.714190e+01  -70.0403442  75.1154598
#> 411   2.250266e+01    8.7868719  24.1573787
#> 412   4.440278e+01  104.5911636 113.6262217
#> 413   4.427003e+01   -6.5450668  44.7512445
#> 414  -9.801324e+01 -138.7784576 169.9001366
#> 415  -9.394212e+01 -132.2634735 162.2305427
#> 416  -1.171216e+01  -21.0369301  24.0775222
#> 417  -4.376369e+00  106.4359665 106.5259009
#> 418   5.545327e+01  124.2706757 136.0818356
#> 419   1.307811e+02   35.9804268 135.6403278
#> 420   1.781600e+02 -119.0809708 214.2924488
#> 421   7.729874e+01 -104.2789536 129.8044530
#> 422  -6.635829e+01  -58.7101555  88.6019482
#> 423  -4.930752e+01  132.9719391 141.8194906
#> 424  -5.290851e+01   89.4577560 103.9326723
#> 425  -6.781438e+01  -31.9304314  74.9556081
#> 426  -5.124580e+01  -78.4392624  93.6955191
#> 427  -3.090382e+01  -43.9933815  53.7630351
#> 428  -1.072401e+02   25.5056839 110.2314316
#> 429  -1.549896e+02   -8.6421881 155.2303194
#> 430   2.102494e+00   58.7304611  58.7680827
#> 431   1.738774e+02   72.7490311 188.4828537
#> 432   1.130454e+02   59.2729836 127.6422450
#> 433  -6.432891e+01   38.6523590  75.0480754
#> 434  -1.102371e+02   -2.3285482 110.2617272
#> 435   1.855590e+00 -136.1803589 136.1930004
#> 436   1.106030e+02 -129.7672119 170.5067710
#> 437   7.839345e+01  -31.3190823  84.4181117
#> 438   7.186459e+01   88.1357498 113.7208385
#> 439   2.035930e+01   86.7908096  89.1467661
#> 440  -3.339573e+01   40.1233139  52.2030200
#> 441  -3.617333e+00  -48.5056496  48.6403448
#> 442  -5.483720e+01 -114.6762695 127.1131988
#> 443  -1.686406e+01  -75.3887482  77.2519254
#> 444   3.071306e+01   44.2412491  53.8570348
#> 445   2.730731e+01   98.2738647 101.9972621
#> 446   3.577577e+00   94.9277649  94.9951557
#> 447  -6.598993e+01   46.5747223  80.7705114
#> 448   4.524833e-01  -27.0771656  27.0809460
#> 449   7.239500e+01   -4.5010939  72.5347950
#> 450   2.116441e+01  -15.0254068  25.9556369
#> 451  -4.312360e+01   12.4840994  44.8942975
#> 452  -7.172475e+01  -84.6511383 110.9515863
#> 453  -2.963223e+01  -62.9290695  69.5567187
#> 454  -6.595142e+01   24.1991234  70.2508922
#> 455  -8.360064e+01   55.3509636 100.2636328
#> 456  -3.409492e+01   20.4326496  39.7486739
#> 457   1.488506e+02  -78.3935318 168.2321233
#> 458   2.236268e+02  -47.3164215 228.5777392
#> 459   6.982192e+01  111.5840988 131.6286897
#> 460  -4.809715e+01  127.4404831 136.2145824
#> 461  -1.229023e+02   44.4120064 130.6805524
#> 462  -5.998629e+01  -41.6654587  73.0367434
#> 463  -3.496968e+01  -30.1375160  46.1643634
#> 464  -5.325091e+01  105.4470749 118.1302044
#> 465  -5.897920e+01   43.3924446  73.2219267
#> 466  -7.445387e+01  -76.9676285 107.0859234
#> 467  -1.777819e+01 -199.1759338 199.9677894
#> 468   9.581384e+01 -161.8568420 188.0902171
#> 469   1.738440e+02   52.7773514 181.6788203
#> 470   1.473720e+02  114.5030060 186.6264679
#> 471  -6.035429e+00  137.6685944 137.8008283
#> 472  -4.972390e+01  -16.5803909  52.4154172
#> 473  -7.884260e+01 -195.9109650 211.1806370
#> 474  -1.002932e+02 -170.8812256 198.1391553
#> 475  -5.350871e+01  -26.1063156  59.5375685
#> 476   1.230916e+01  159.8565369 160.3297474
#> 477   1.026420e+02   67.2130280 122.6905559
#> 478   2.066385e+01  -45.3160934  49.8050497
#> 479  -1.992177e+01  -91.4063950  93.5521564
#> 480   2.079556e+01  -22.4962101  30.6355144
#> 481   2.169246e+01   25.0061054  33.1039009
#> 482   3.022297e+01  102.8507080 107.1993300
#> 483  -4.415858e+01   91.6153183 101.7022442
#> 484  -3.188023e+01    7.2549558  32.6953086
#> 485   1.631081e+00   94.8977661  94.9117824
#> 486  -6.109875e+00    8.7047768  10.6350230
#> 487  -1.372449e+01  -76.4666443  77.6885403
#> 488  -8.918569e+01 -110.8617706 142.2828866
#> 489  -2.519286e+01  -63.3070946  68.1356617
#> 490  -1.440088e+01   55.7765160  57.6055996
#> 491  -5.269914e+00    3.6034782   6.3841245
#> 492  -2.339707e+01   28.1750336  36.6231539
#> 493  -3.446374e+01   96.0406570 102.0370387
#> 494   9.135168e+01  148.3333740 174.2065404
#> 495   1.160980e+02  160.8619080 198.3817046
#> 496   3.174864e+01  -27.4101925  41.9439468
#> 497  -2.075041e+01 -136.8212738 138.3858397
#> 498  -6.963202e+00  -46.3008919  46.8215631
#> 499   2.139938e+01  -10.8675766  24.0007815
#> 500  -4.736445e+01 -109.3037491 119.1247284
#> 501  -6.945049e+01 -168.4066772 182.1652515
#> 502  -5.823256e+01  -18.3606873  61.0585403
#> 503  -1.085355e+02  162.1747437 195.1425387
#> 504  -4.942055e+01  199.3380737 205.3729742
#> 505   1.092197e+02  110.7708130 155.5606313
#> 506   1.824947e+02  -75.7067413 197.5748953
#> 507   5.772770e+01  -93.9677277 110.2833653
#> 508  -1.516147e+01  -57.3817291  59.3509318
#> 509  -5.762488e+00   21.0544262  21.8287685
#> 510   5.874662e+01   37.5718842  69.7338677
#> 511   7.999854e+01   30.7040863  85.6884342
#> 512   6.585980e+01  -47.5435410  81.2274697
#> 513   6.402110e+01 -133.6345520 148.1785919
#> 514  -6.642811e+01   -8.6266470  66.9859134
#> 515  -1.529408e+02   92.0478134 178.5040644
#> 516  -1.786041e+02  103.7606583 206.5567517
#> 517  -7.899236e+01   11.4485283  79.8176811
#> 518   4.554956e+01    8.3886242  46.3155679
#> 519   2.361653e+01  -71.9130554  75.6916646
#> 520  -3.790626e+01 -114.9621506 121.0503219
#> 521  -3.925190e+01 -104.9399033 112.0405937
#> 522  -3.161526e+00   15.6889296  16.0043043
#> 523   9.167590e+01   90.2088394 128.6161186
#> 524   6.213488e+01   -0.7723511  62.1396801
#> 525  -3.598588e+01   -9.7531528  37.2841478
#> 526  -4.031204e+01   10.6316395  41.6904366
#> 527  -1.076365e+01  166.5482330 166.8956864
#> 528   1.005323e+02   75.3296738 125.6236528
#> 529   2.548709e+01  -52.8119583  58.6403842
#> 530  -1.169934e+01  -85.3618698  86.1598706
#> 531   3.360295e+01  -64.7395477  72.9408484
#> 532  -5.699343e+01   82.2800217 100.0912263
#> 533   4.344174e+01   42.0331688  60.4480925
#> 534   3.059410e+01   -3.7041371  30.8175217
#> 535  -1.994761e+01  -73.5811310  76.2370636
#> 536  -6.801264e+01  -85.3530197 109.1368748
#> 537  -6.799895e+01    9.7895346  68.7000131
#> 538   1.402583e+02   68.8071976 156.2268267
#> 539   7.860683e+01   66.4823532 102.9511423
#> 540  -1.312786e+01  -50.0219078  51.7158771
#> 541  -8.192179e+01  -97.3053055 127.1986727
#> 542  -6.860419e+01  -24.3675537  72.8032427
#> 543  -2.966751e+01   76.2588959  81.8265256
#> 544   2.980941e+01    5.0154891  30.2283983
#> 545   4.746659e+01  -82.0091171  94.7553299
#> 546   4.848965e+01  -54.2462616  72.7592133
#> 547   3.166397e+01   34.8598862  47.0937202
#> 548   2.984564e+01   66.1224899  72.5461624
#> 549   6.707475e+01  -16.8912125  69.1688839
#> 550   3.889807e+01  -52.8495216  65.6211215
#> 551  -1.126065e+01   75.1134949  75.9528762
#> 552  -6.171885e+01  159.1925049 170.7380156
#> 553   2.299836e+01   26.4605980  35.0583484
#> 554   1.686762e+01 -146.5230713 147.4907698
#> 555  -2.069419e+01 -154.9083557 156.2845107
#> 556  -4.078989e+01  -20.9005184  45.8328138
#> 557  -5.814500e+00   65.2088470  65.4675656
#> 558   7.763519e+01    3.9490986  77.7355605
#> 559   2.454659e+01  -32.8495483  41.0076583
#> 560   1.270677e+00   -3.9421382   4.1418684
#> 561   6.254091e+00   49.0644836  49.4614720
#> 562   6.868610e+01  111.1369324 130.6491431
#> 563   8.341347e+01   34.5378075  90.2810428
#> 564  -6.915763e+01  -78.4370117 104.5712328
#> 565  -1.444362e+02  -85.2846909 167.7357768
#> 566  -1.426266e+02  -17.6417904 143.7135208
#> 567  -3.438126e+01   49.7927551  60.5094166
#> 568   4.384656e+01   36.9337692  57.3290874
#> 569   6.401946e+01   29.5607471  70.5147457
#> 570   9.104091e+00   51.7473450  52.5420991
#> 571  -3.267664e+01   -2.0641108  32.7417711
#> 572   8.988897e+01  -17.1746216  91.5149958
#> 573   1.033442e+02   16.8073673 104.7020420
#> 574  -1.886405e+01    6.6167774  19.9908474
#> 575  -3.042385e+01   -3.7484653  30.6539038
#> 576  -3.264149e+01  -13.9478149  35.4965980
#> 577   6.575686e+00   17.2110977  18.4244817
#> 578  -6.340285e+01  -13.2127838  64.7649496
#> 579  -5.993774e+01  -75.1112289  96.0948957
#> 580   3.461686e+01   -7.8593431  35.4978379
#> 581  -4.638787e+01   14.3376884  48.5531036
#> 582  -1.047128e+01   35.3221130  36.8415436
#> 583  -2.910263e+01   47.8324928  55.9902713
#> 584  -2.388370e+01   68.1848831  72.2468642
#> 585  -1.294989e+00    0.4628580   1.3752213
#> 586   5.711263e-01  -65.7226105  65.7250920
#> 587   3.411004e+01  -73.3443222  80.8880977
#> 588   9.949944e+00  -51.6268005  52.5768762
#> 589  -3.537946e+01  -43.4744263  56.0511542
#> 590  -3.488393e+01   51.7970009  62.4485240
#> 591  -1.179531e+01  114.5444717 115.1501857
#> 592   1.016453e+02   59.0087204 117.5320884
#> 593   1.079033e+02   18.6133595 109.4969800
#> 594   7.121037e+00  -31.5843372  32.3771452
#> 595  -8.450308e+01  -15.7555552  85.9593417
#> 596  -9.094563e+01  -63.9707603 111.1906755
#> 597   4.556324e+01   23.6404057  51.3310624
#> 598   5.491090e+01   41.3811111  68.7575661
#> 599   4.376693e+01   43.4220810  61.6524256
#> 600  -3.200974e+01   15.5010519  35.5655170
#> 601  -5.603194e+01  -25.9230537  61.7380196
#> 602  -8.740801e+01   36.5448608  94.7401049
#> 603  -6.735458e+01   28.0163555  72.9489897
#> 604   9.820505e+01   50.0164986 110.2083550
#> 605   9.169609e+01  -36.5956306  98.7289888
#> 606   3.254277e+01 -133.3450470 137.2586380
#> 607  -8.405245e+01 -133.9535065 158.1403067
#> 608  -2.889292e+01  -80.5910263  85.6137508
#> 609   6.127345e+01   12.2633591  62.4885994
#> 610   3.151767e+00  110.4942245 110.5391663
#> 611   5.916673e+01  142.4194794 154.2206519
#> 612  -2.750146e+00  100.0376129 100.0754081
#> 613   4.298892e+01  -38.3203812  57.5890531
#> 614   1.630000e+01 -222.6869965 223.2827546
#> 615  -8.560349e+01 -129.8434448 155.5225969
#> 616  -4.948450e+01   93.6394653 105.9106459
#> 617  -9.544683e+00  136.4806824 136.8140258
#> 618   3.788385e+01   27.4592800  46.7888679
#> 619  -3.993140e+01  -38.1985970  55.2598366
#> 620  -4.582877e+01    8.3756332  46.5878419
#> 621   9.860680e+01  -60.7729797 115.8302933
#> 622   9.904352e+01 -134.1219788 166.7282929
#> 623  -4.182967e+01   32.8440552  53.1832076
#> 624  -2.872639e+01  141.1029358 143.9973758
#> 625   1.810426e+01   17.9570599  25.4994172
#> 626   9.795666e+01 -169.7810516 196.0130410
#> 627   3.553581e+01 -125.8611679 130.7816014
#> 628  -4.765443e+01   10.5335875  48.8047215
#> 629  -1.026541e+02   17.1890087 104.0832570
#> 630  -1.494016e+02   12.3861361 149.9141530
#> 631  -7.356981e+00  109.8744431 110.1204723
#> 632   4.149800e+00  101.9563828 102.0407998
#> 633   9.026670e+01   24.0675392  93.4201462
#> 634   1.814588e+02   -2.5099914 181.4761294
#> 635   1.059227e+02   45.0669060 115.1114143
#> 636  -4.053369e+01   47.2548218  62.2575137
#> 637  -2.343241e+02 -122.4744186 264.4007668
#> 638  -1.284765e+02 -108.5122986 168.1699099
#> 639   7.405032e+01   57.0842133  93.4989722
#> 640   9.502089e+01   71.3628311 118.8344355
#> 641   9.184440e+00    5.7607350  10.8415866
#> 642  -1.016583e+02   21.0092640 103.8065116
#> 643  -4.848836e+01   87.5472946 100.0782175
#> 644   7.475072e+01  135.2507782 154.5329863
#> 645   9.379368e+01  -45.1826630 104.1092077
#> 646   3.181708e+01 -175.8494110 178.7046226
#> 647  -6.372445e+01 -115.6273651 132.0245923
#> 648  -1.232752e+02    2.4979043 123.3005351
#> 649  -3.465167e+01   66.3283310  74.8343898
#> 650   1.165258e+02   12.8898430 117.2365361
#> 651   1.406645e+02  -55.0839806 151.0653478
#> 652  -5.976678e+01   66.1270370  89.1339081
#> 653  -1.529025e+02  101.1049728 183.3068418
#> 654  -4.971039e+01   61.8303032  79.3354214
#> 655   3.647405e+01  -67.7834702  76.9737314
#> 656   3.933222e+01 -142.3607788 147.6943297
#> 657  -1.248575e+00  -64.6215897  64.6336506
#> 658   1.092138e+02  -32.9967270 114.0895969
#> 659   8.778655e+01  -29.4701595  92.6011290
#> 660  -1.495868e+01   -6.2304649  16.2043497
#> 661  -1.101411e+02   68.3714523 129.6368463
#> 662  -8.790970e+01   74.6572723 115.3335311
#> 663   1.002108e+02   -9.5852289 100.6681656
#> 664   1.611426e+02  -44.4466171 167.1599297
#> 665   7.600684e+01  -51.7966652  91.9779038
#> 666  -9.255997e+01   49.3898048 104.9128292
#> 667  -1.261977e+02  133.3557281 183.6017645
#> 668  -1.731146e+01   51.2872276  54.1300870
#> 669  -1.094839e+01   23.5045204  25.9293201
#> 670   2.167040e+01  -38.8131447  44.4529691
#> 671   8.470773e+01   18.1357975  86.6274045
#> 672   5.396082e+01   28.9755135  61.2482685
#> 673  -3.454506e+01   59.0963745  68.4524861
#> 674  -1.247278e+02   46.2526741 133.0275455
#> 675  -1.421430e+01 -124.8267670 125.6334670
#> 676   4.542916e+01 -274.3610535 278.0967392
#> 677  -1.282784e+01 -189.5617065 189.9952478
#> 678   3.678633e+01   58.3421478  68.9713031
#> 679   6.026334e+01  233.3134918 240.9706521
#> 680   2.701047e+01  168.8750916 171.0215256
#> 681  -1.026017e+02   44.6407509 111.8923414
#> 682  -6.551468e+01  -40.0391121  76.7808808
#> 683   9.304421e+01  -32.9403992  98.7030666
#> 684   5.908152e+01   23.2498817  63.4916015
#> 685  -9.777988e+01    3.1124177  97.8293996
#> 686  -1.844987e+02  -65.6250916 195.8224299
#> 687  -4.058287e+01 -130.4585114 136.6250060
#> 688   5.321480e+01  -59.8838921  80.1117698
#> 689   5.041680e+01    4.1756492  50.5894249
#> 690   2.658733e+01   29.1895866  39.4831379
#> 691  -2.487465e+01  -51.7872353  57.4514238
#> 692  -3.525638e+01  -25.3592358  43.4292909
#> 693  -5.833571e+01   14.3324680  60.0705834
#> 694  -2.116192e+01   36.4913101  42.1834379
#> 695   1.912698e+01   13.1065445  23.1866975
#> 696   4.681190e+01   34.5415955  58.1762455
#> 697   1.277638e+02   99.3756409 161.8613760
#> 698   1.066978e+02   13.9288378 107.6030818
#> 699   1.295600e+01  -64.9066162  66.1870590
#> 700  -6.088345e+01 -102.3649139 119.1023491
#> 701  -4.891651e+01  -36.1700974  60.8366751
#> 702   2.625893e+01  -14.0165596  29.7656758
#> 703  -3.376272e+01   21.6462708  40.1058903
#> 704  -2.596938e+01   62.0040779  67.2228692
#> 705  -2.001198e+01  -22.7559376  30.3036626
#> 706   1.599721e+01  -97.7182693  99.0190424
#> 707   1.282155e+01  -61.4068336  62.7311029
#> 708  -1.026168e+02   20.8456821 104.7126563
#> 709  -1.710504e+02   65.7529602 183.2531077
#> 710  -1.725572e+02   88.6114120 193.9792659
#> 711   9.749714e+00   79.1597672  79.7579191
#> 712   2.573846e+02    0.9854867 257.3864386
#> 713   2.423485e+02 -102.0979080 262.9767877
#> 714   4.298662e+01  -68.4364853  80.8170909
#> 715  -1.004639e+02   25.4440002 103.6358694
#> 716  -4.838734e+01  106.9792099 117.4133146
#> 717   1.023770e+02   72.0856018 125.2093727
#> 718   6.615705e+01    6.2812653  66.4545687
#> 719  -2.698997e+01  -39.0217667  47.4463551
#> 720  -5.495272e+01  -20.1211948  58.5206323
#> 721   1.777327e+01  106.5390701 108.0113999
#> 722  -8.718177e+00  -28.6575260  29.9543053
#> 723  -2.167167e+01 -132.4067841 134.1686161
#> 724   1.833296e+01 -128.9812164 130.2775947
#> 725   9.615687e+01    1.1973627  96.1643226
#> 726  -1.637521e+01  134.2756042 135.2704156
#> 727  -1.344866e+02   45.0684509 141.8373005
#> 728  -2.864678e+01   71.9647598  77.4568554
#> 729   1.054872e+02   53.5185051 118.2868246
#> 730   1.070919e+02    3.4945390 107.1488964
#> 731  -1.384777e+02  -76.8069687 158.3520676
#> 732  -1.593375e+02  -82.6768341 179.5101959
#> 733  -2.316972e+00   47.2197723  47.2765826
#> 734   1.469819e+02   78.3861465 166.5774725
#> 735   1.027154e+02   -3.9295897 102.7905481
#> 736  -6.271564e+01 -128.1423035 142.6664012
#> 737  -5.934589e+01  -99.7183838 116.0417648
#> 738   1.929299e+01   59.4447250  62.4971594
#> 739   5.844822e+01  147.3354645 158.5053104
#> 740  -1.129546e+01   70.3895950  71.2901287
#> 741  -6.021533e+01 -110.3082886 125.6734032
#> 742   6.829753e+00 -164.0882568 164.2303308
#> 743   5.453083e+01 -122.4661179 134.0580542
#> 744  -6.214607e+00   50.2904854  50.6730131
#> 745  -4.504614e+01  178.8046112 184.3915497
#> 746   6.440118e+01  158.9217224 171.4748535
#> 747   7.704779e+01   30.5189838  82.8720122
#> 748   3.427051e+01 -104.6553726 110.1236350
#> 749  -7.256131e+01  -12.9232740  73.7031525
#> 750  -1.294204e+02   -0.1261217 129.4204716
#> 751  -1.079229e+02   -1.2877878 107.9306261
#> 752  -3.954381e+01 -123.3432999 129.5271503
#> 753   1.001267e+02 -110.2303391 148.9163452
#> 754   5.724467e+01   40.8532944  70.3274027
#> 755   5.613612e+00  119.8962936 120.0276379
#> 756  -6.963959e+01  158.0447235 172.7072864
#> 757  -2.689244e+01  -54.1005669  60.4158473
#> 758  -2.359284e+01  -90.6779556  93.6969250
#> 759   5.684020e+00 -141.4450226 141.5591837
#> 760   1.348386e+02  -57.9072227 146.7470777
#> 761   4.747453e+01   80.0349197  93.0560009
#> 762  -4.846629e+01   60.5019073  77.5207175
#> 763  -5.905684e+01   83.0208740 101.8831476
#> 764   4.791584e+01   64.9044876  80.6754028
#> 765   9.200202e+01   84.9606018 125.2304910
#> 766   6.198813e+00  -36.0888100  36.6173114
#> 767  -1.040091e+01 -127.7381973 128.1609375
#> 768  -3.191744e+01  -85.2403641  91.0200138
#> 769  -1.039134e+02  -81.9442596 132.3361303
#> 770  -6.680409e+01  -34.0183525  74.9668932
#> 771  -2.687966e+01    4.4353929  27.2431425
#> 772   5.107715e+01   54.7263451  74.8588560
#> 773   1.192196e+02   68.2329330 137.3646838
#> 774   6.014511e+01   43.7798042  74.3915728
#> 775  -2.826647e+01   68.2945557  73.9130562
#> 776  -1.456126e+02   76.5409088 164.5039026
#> 777  -1.035049e+02   -2.7767007 103.5421744
#> 778  -4.489375e+01 -103.7858810 113.0794300
#> 779  -4.228652e+01 -117.0357132 124.4407802
#> 780   4.754877e+01   23.6654224  53.1125017
#> 781   8.412463e+01   88.8809280 122.3796281
#> 782   3.786615e+01   49.8193741  62.5764736
#> 783   3.756891e+01   26.2435837  45.8273812
#> 784   1.094379e+02  -29.1976757 113.2658635
#> 785   1.210827e+02  -40.5833511 127.7029379
#> 786  -1.576806e+01  -20.5872650  25.9319729
#> 787  -1.363840e+02   62.8442650 150.1665531
#> 788  -5.488202e+01   59.5457497  80.9798266
#> 789   5.489466e+01   -7.7224703  55.4351943
#> 790   2.622567e+01  -46.1202126  53.0552501
#> 791  -5.105489e+01  -91.0832214 104.4162564
#> 792  -8.085963e+01 -193.0033417 209.2571871
#> 793  -7.558076e+01  -77.5567856 108.2936147
#> 794  -5.756229e+01   49.3860207  75.8445535
#> 795   9.848696e+00  100.1692886 100.6522886
#> 796   8.175284e+01  116.5301590 142.3474780
#> 797   1.238933e+02   73.6402512 144.1264491
#> 798   6.041591e+01   50.5537491  78.7766690
#> 799   5.599481e+01  -68.3927002  88.3910652
#> 800  -1.627033e+01  -35.6092453  39.1502483
#> 801  -8.392488e+01   20.4462795  86.3796040
#> 802  -1.174508e+02   15.7700920 118.5047846
#> 803  -1.168080e+02   54.3052101 128.8144434
#> 804   3.984123e+01   18.6228580  43.9787975
#> 805   3.054002e+01  -19.7088413  36.3473724
#> 806   7.406695e+01  -91.3097076 117.5728518
#> 807   9.606433e+01  -18.1125336  97.7569413
#> 808   4.661284e+01   45.6199760  65.2222274
#> 809   1.841958e+01  -13.6299562  22.9141147
#> 810  -1.659230e+01    3.0894105  16.8774688
#> 811  -1.632183e+01   19.5429935  25.4623387
#> 812  -9.510595e+00   77.5268173  78.1079946
#> 813  -3.953249e+00   75.5374908  75.6408666
#> 814   4.214491e+01    0.5692993  42.1487500
#> 815  -1.158941e+00 -140.6060638 140.6108400
#> 816  -1.656708e+01 -109.2864990 110.5350937
#> 817   1.153393e+02   17.2865219 116.6275654
#> 818   1.558649e+02   79.4040375 174.9252935
#> 819   5.431948e+01   57.1679153  78.8592172
#> 820  -1.215607e+02   48.1150627 130.7366450
#> 821  -1.267703e+02   77.8226624 148.7516942
#> 822  -1.368297e+01  -67.9120483  69.2767642
#> 823   6.468980e+01 -138.4520874 152.8193419
#> 824  -2.318135e+01  -78.1649857  81.5299934
#> 825  -8.499044e+01  -13.1104956  85.9956978
#> 826  -1.285745e+02   36.3056870 133.6020185
#> 827  -5.934536e+00   -1.3927904   6.0957845
#> 828   7.304630e+01    1.4239771  73.0601811
#> 829   7.418579e+01   27.4381714  79.0973125
#> 830   4.698165e+01   50.1171608  68.6950144
#> 831  -4.945235e+01  -66.5593109  82.9197012
#> 832  -1.392833e+00 -189.7850952 189.7902061
#> 833  -2.789060e+01  -77.8570786  82.7019369
#> 834  -4.213647e+01  161.1579895 166.5754467
#> 835  -1.210911e+02  245.4225616 273.6700399
#> 836  -3.382426e+01  108.9653854 114.0944163
#> 837   9.678599e+01    0.7797339  96.7891287
#> 838   9.195636e+01   -0.7332906  91.9592836
#> 839  -2.868655e+01    2.7865016  28.8215637
#> 840  -8.014622e+01  -37.7647743  88.5979433
#> 841   6.983988e+01  -15.3945332  71.5164370
#> 842   1.937825e+01  -10.4982414  22.0392718
#> 843  -6.322104e+01  -39.7165642  74.6612731
#> 844  -2.598433e+01  -21.7133045  33.8622631
#> 845   1.160671e+02  -30.2309608 119.9394891
#> 846   1.475430e+02  -43.6938400 153.8768754
#> 847  -3.808787e+01  -45.2444458  59.1417411
#> 848  -1.014868e+02   21.1708202 103.6714736
#> 849  -1.568174e+01   46.7137299  49.2756468
#> 850   2.906262e+01  -90.5888367  95.1366018
#> 851   8.269756e+01 -102.6562653 131.8225872
#> 852   4.006224e+01   71.8999329  82.3078560
#> 853  -6.482711e+01  157.6731720 170.4798621
#> 854  -1.738654e+02   92.8093109 197.0856588
#> 855  -1.542944e+02  -56.5563774 164.3331145
#> 856   2.662366e+01  -88.8910751  92.7924699
#> 857   7.058276e+01  -96.2862396 119.3857839
#> 858  -6.854318e+01  -71.1085739  98.7653640
#> 859  -5.953473e+01   32.8181076  67.9809723
#> 860   4.886933e+01   18.9233589  52.4051971
#> 861   1.217132e+02  -49.2755013 131.3094918
#> 862   1.130682e+01  -41.0272484  42.5567777
#> 863  -1.317249e+02  118.1637344 176.9579533
#> 864   2.368376e+01  172.5401611 174.1580532
#> 865   1.394864e+02   56.7215729 150.5781560
#> 866   1.497979e+02  -18.1085548 150.8885135
#> 867  -5.733168e+01  -37.6903839  68.6111291
#> 868  -1.740164e+02 -102.2749252 201.8461911
#> 869   1.332911e+00 -132.3563538 132.3630652
#> 870   8.623992e+01    0.3326399  86.2405631
#> 871   9.085352e+01  151.0075684 176.2318030
#> 872  -3.709542e+01   83.5602493  91.4242048
#> 873  -5.824137e+01  -32.2046394  66.5522093
#> 874   2.311628e+01  -69.8596344  73.5848551
#> 875   3.675448e+01   29.7865658  47.3088912
#> 876  -3.783742e+01   71.3323746  80.7463815
#> 877  -1.190206e+02  103.0768738 157.4507873
#> 878  -3.761079e+01   71.0028229  80.3490660
#> 879   5.030610e+01  -36.5307083  62.1707055
#> 880   5.535581e+01 -210.0056763 217.1788424
#> 881  -8.277071e+01 -264.6042175 277.2478728
#> 882  -1.198040e+02  -90.7955246 150.3224248
#> 883  -3.467633e+01  111.5523834 116.8177316
#> 884   1.323672e+02  108.0207901 170.8495699
#> 885   2.023983e+02   38.4904480 206.0256938
#> 886   8.057294e+01  -27.6697903  85.1916397
#> 887   8.808168e-01  -24.8304176  24.8460355
#> 888  -1.018391e+02  -50.1496353 113.5173304
#> 889  -1.372692e+02  -10.5180531 137.6715407
#> 890  -4.441487e+01   92.0833130 102.2351080
#> 891   7.761693e+01   26.1471348  81.9027483
#> 892   2.078028e+02  -46.3294144 212.9047121
#> 893   1.500597e+02  -38.6678314 154.9616631
#> 894   1.982712e+01   88.5130692  90.7065493
#> 895  -9.147890e+01   43.3526726 101.2316365
#> 896  -1.778048e+02   -5.0155034 177.8754882
#> 897  -7.143619e+01  -52.4051132  88.5969797
#> 898   2.861308e+01   -1.0392104  28.6319484
#> 899   6.537162e+01   48.0756950  81.1462949
#> 900  -5.445294e+01    4.5432024  54.6421410
#> 901  -1.615296e+02   51.8921165 169.6602608
#> 902  -5.784155e+01  105.0228729 119.8976589
#> 903   1.138315e+01  158.9937134 159.4006809
#> 904   2.558883e+01   74.2837601  78.5675842
#> 905   5.504929e+01  -99.6291733 113.8261678
#> 906   9.273621e+01 -195.5504150 216.4254346
#> 907   1.469631e+02 -109.0334778 182.9930416
#> 908   1.238300e+02  -34.3669167 128.5105001
#> 909   4.259396e+01  -44.1169548  61.3233325
#> 910   5.288409e+00  -28.9907494  29.4691503
#> 911  -1.124559e+02    5.3702011 112.5840529
#> 912  -8.398303e+01  -38.0881348  92.2163527
#> 913  -5.587234e+01  -99.2400131 113.8872174
#> 914  -8.161038e+01  -54.9295883  98.3743571
#> 915   3.437092e+01   84.8621979  91.5584658
#> 916   1.273340e+02  149.8633118 196.6544218
#> 917   1.623760e+02   63.4948006 174.3489387
#> 918  -4.111063e+01   71.6812897  82.6334750
#> 919  -1.233928e+02   39.6262398 129.5994306
#> 920  -5.401559e+01   36.1229362  64.9811586
#> 921   5.296302e+00  -36.7393532  37.1191445
#> 922  -5.090133e+01  -35.6651802  62.1526377
#> 923  -1.178604e+02   -3.6584949 117.9171271
#> 924  -3.793375e+01  -14.2989082  40.5392178
#> 925   8.297607e+01  -11.7110023  83.7984276
#> 926   1.555932e+02  -82.9224167 176.3104401
#> 927   1.227894e+02   73.9888306 143.3582736
#> 928   3.783664e+01  108.5789108 114.9825689
#> 929  -1.150208e+02   -7.3745761 115.2569663
#> 930  -1.632225e+02 -147.1144562 219.7367946
#> 931  -1.398659e+02 -158.8772278 211.6705749
#> 932  -2.707574e+01   23.8830452  36.1039510
#> 933   3.655760e+01   90.6500626  97.7440131
#> 934   5.601328e+01   89.0501556 105.2017928
#> 935   1.841393e+01  -57.8352165  60.6958405
#> 936  -5.233871e+01  -98.4288635 111.4790663
#> 937  -5.893279e+01  -74.8113098  95.2355250
#> 938  -9.352628e+00  -13.1363354  16.1255993
#> 939   4.610781e+01  118.2794037 126.9486007
#> 940   6.588637e+01  176.7297974 188.6118627
#> 941   5.495459e+01  179.3807983 187.6099096
#> 942   1.261099e+01   55.8918228  57.2968832
#> 943  -9.305572e+01 -135.7531586 164.5851956
#> 944  -7.321916e+01 -170.5330353 185.5870734
#> 945   3.131024e+01  -17.0953579  35.6732726
#> 946   7.122800e+01  130.0129089 148.2456917
#> 947   6.814613e+01   29.0697708  74.0874286
#> 948   1.889498e+01 -126.9069138 128.3058267
#> 949  -6.794708e+00 -110.7868805 110.9950492
#> 950   6.643616e+00   62.9150581  63.2648574
#> 951   8.790650e+00  161.2648315 161.5042458
#> 952   5.087629e+01   55.4403229  75.2464366
#> 953  -5.863740e+01   16.9555130  61.0396124
#> 954  -1.160621e+02  -70.4690857 135.7803515
#> 955  -3.505691e+01  -40.6890945  53.7083717
#> 956  -5.351089e+01  -73.8870926  91.2289290
#> 957   5.041214e+01  -79.4816284  94.1207389
#> 958   1.722302e+01  -14.8697176  22.7539214
#> 959   1.047869e+02  101.5717773 145.9353318
#> 960   1.597631e+02  157.6964111 224.4825350
#> 961   4.881795e+01   41.2205925  63.8931108
#> 962   9.200782e+00  -28.4269047  29.8788101
#> 963  -1.341050e+02  -60.0319862 146.9285593
#> 964  -1.281217e+02   18.2126770 129.4097237
#> 965  -1.091212e+02   36.3018723 115.0010972
#> 966  -3.837355e+01  -43.9142876  58.3180420
#> 967   1.009265e+02  -79.6086426 128.5445157
#> 968   5.734667e+01  -98.0852737 113.6193721
#> 969  -4.096324e+01 -112.0193787 119.2741730
#> 970  -2.078053e+01  -83.7990723  86.3372153
#> 971   5.851606e+01   12.8391733  59.9080394
#> 972   1.144921e+02  125.6363754 169.9792255
#> 973   1.124487e+02  118.0821381 163.0585881
#> 974   6.198242e+01   51.6819534  80.7021961
#> 975   8.519756e+01  -18.7691422  87.2404961
#> 976  -3.639196e+00   57.3999367  57.5151848
#> 977  -7.619775e+01   85.5800705 114.5864135
#> 978  -8.668913e+01   50.9399757 100.5479265
#> 979  -8.938636e+01  -68.4790344 112.6023958
#> 980  -1.536304e+01 -158.7408295 159.4825195
#> 981   8.143954e-01  -47.2294693  47.2364902
#> 982   2.589763e+01  -37.1643944  45.2976753
#> 983   4.498835e+01  -17.4803638  48.2650469
#> 984  -3.595272e+01   36.8654938  51.4942956
#> 985  -9.152583e+01   99.3424835 135.0774066
#> 986  -8.965225e+01   42.6746750  99.2907559
#> 987   5.283197e+01 -152.7166290 161.5969872
#> 988   1.468150e+02 -154.7167358 213.2883337
#> 989   7.638568e+01   23.9586906  80.0549258
#> 990  -5.707289e+00  120.7034454 120.8383006
#> 991  -4.911657e+01   75.2897797  89.8943154
#> 992  -3.156867e+00    5.4971714   6.3391407
#> 993  -7.672978e+00   24.8984051  26.0538898
#> 994  -1.396763e+01   46.0932808  48.1631107
#> 995   3.718538e+01   21.5900288  42.9986283
#> 996   8.118732e+01   89.2686462 120.6659506
#> 997   1.926018e+01   59.0212402  62.0843069
#> 998  -1.343357e+02   11.8797941 134.8599571
#> 999  -3.440642e+02 -171.3951111 384.3910301
#> 1000            NA           NA          NA