Differentiate x and y and compute change amplitude via an algorithm proposed by Engbert & Kliegl (2003)

Differentiate x and y and compute change amplitude via an algorithm proposed by Engbert & Kliegl (2003)

diff_ek(x, y, trial, sample_rate, options = NULL)

Arguments

x

vector with x coordinates in degrees of visual angle

y

vector with y coordinates in degrees of visual angle

trial

vector with trial index

sample_rate

sample rate in Hz

options

List with method specific options, see Details.

Value

data.frame with columns x, y, and amp

Details

Method options, please refer to Engbert & Kliegl (2003) for details on parameters and the rationale for default values.

ek_velocity_time_window Time window for velocity computation in milliseconds. Defaults to 20 ms.

See also

Examples

diff_ek(rnorm(1000), rnorm(1000), rep(1, 1000), 250, list("ek_velocity_time_window" = 20))
#>                x             y        amp
#> 1             NA            NA         NA
#> 2     -7.7807264  148.18380737 148.387939
#> 3    141.3859253  -55.02769852 151.716932
#> 4    231.8797150  -67.57732391 241.526183
#> 5    163.7541809   -6.16642523 163.870243
#> 6     -4.9933333   33.77098083  34.138139
#> 7    -80.3289642  -87.88302612 119.063717
#> 8      5.7436266 -126.52433777 126.654638
#> 9     41.3500061  -95.12149048 103.720398
#> 10    20.5216351   21.97062683  30.064031
#> 11   -58.8488274   88.80052948 106.530364
#> 12   -55.1574326  108.64787292 121.847046
#> 13   -62.0384560   50.30990982  79.874007
#> 14   -64.7680283  -33.89635468  73.101712
#> 15   -12.1104469  -34.60469437  36.662621
#> 16    66.9018250   74.84572601 100.387932
#> 17   124.5554581  117.23715210 171.051489
#> 18    -1.1378177   45.01224899  45.026628
#> 19  -112.4374237  -87.39395905 142.407438
#> 20  -125.9852448 -105.95006561 164.613785
#> 21    54.6503143  -99.72917938 113.721441
#> 22   128.2355042  -53.32675171 138.881557
#> 23     0.5746619   79.69675446  79.698826
#> 24   -80.6697617   82.20320129 115.173681
#> 25   -62.5721817  -17.12911987  64.874376
#> 26    39.7912521 -100.95245361 108.511482
#> 27    -2.4415927  -81.11961365  81.156350
#> 28    57.5021248   47.86743546  74.818352
#> 29   147.9156647   98.68138885 177.811868
#> 30   112.4016113  120.38909912 164.704758
#> 31   -75.2783737  122.34976959 143.653401
#> 32  -111.5014725    3.74343085 111.564294
#> 33   -12.1067095  -74.17742157  75.158914
#> 34    11.4978800  -61.12017822  62.192262
#> 35   -29.8417549   16.04201126  33.880326
#> 36   -54.9254379   12.92437458  56.425554
#> 37    89.9275208  -66.34801483 111.754275
#> 38   171.7811584  -54.46542358 180.208903
#> 39    56.2550278  -10.81610966  57.285394
#> 40   -31.5730171   42.42235565  52.882054
#> 41   -55.1526146   22.13203049  59.427583
#> 42  -119.0335617   53.61416245 130.550631
#> 43   -22.3122692  -25.28539085  33.722223
#> 44   -20.3664742 -114.60657501 116.402149
#> 45    -3.0573950 -114.86662292 114.907305
#> 46   -26.3481903   24.20363235  35.777688
#> 47   -18.0109482  138.24345398 139.411789
#> 48   101.6074600   33.05022430 106.847523
#> 49    50.6092148  -73.51136017  89.248040
#> 50   -52.2354546 -113.04928589 124.533866
#> 51   -69.4354706  107.60911560 128.066414
#> 52   -29.4360523  143.96307373 146.941648
#> 53    13.0663404   35.33223724  37.670894
#> 54   -61.8600922 -161.42048645 172.867708
#> 55  -174.9582062 -108.50541687 205.873261
#> 56  -102.3232117  107.64234161 148.515701
#> 57   126.2799301   94.83585358 157.925488
#> 58   224.7817841   63.10520172 233.471876
#> 59   101.5964737   51.38397598 113.851467
#> 60   -38.5204010   56.23832321  68.165756
#> 61   -56.1792068  -60.43857193  82.516206
#> 62    62.1970139 -123.20462036 138.013938
#> 63    99.1173477  -77.40417480 125.760307
#> 64   -14.5895615    2.55980134  14.812423
#> 65  -165.2942047  -62.21430588 176.614818
#> 66  -131.1618042   10.32040977 131.567206
#> 67    65.7598801   76.60555267 100.959262
#> 68    89.8960953   23.99170685  93.042517
#> 69     5.8164396   49.61992264  49.959661
#> 70  -105.6905365   22.80083275 108.122003
#> 71    -1.6790286   59.47078323  59.494480
#> 72   132.5810547  -73.54148102 151.611627
#> 73    81.6305618  -90.93424225 122.198957
#> 74     1.9677286   66.82689667  66.855860
#> 75   -85.1728516   36.47365952  92.653885
#> 76   -38.3438148 -100.16223907 107.250745
#> 77   -77.1740112 -203.58177185 217.718547
#> 78   -85.7664108   57.91009903 103.486505
#> 79    -0.1062004  150.26565552 150.265693
#> 80    59.4980354   27.29861450  65.461673
#> 81   188.7041321  -66.83692932 200.190970
#> 82   181.6061554  -22.12328911 182.948724
#> 83   105.8421249   61.65881729 122.492307
#> 84  -105.8132553  -38.79749680 112.701778
#> 85  -264.1453552  -83.99526215 277.178594
#> 86   -75.1617355  -63.34711075  98.296200
#> 87    96.4694138  -29.11188507 100.766312
#> 88   157.5830688   50.04438400 165.338634
#> 89    37.5404320  102.49867249 109.157052
#> 90    -9.5389576   97.26573181  97.732361
#> 91   -73.1778870   72.91675568 103.304678
#> 92  -108.2108002  -25.55991745 111.188519
#> 93   -82.1420517  -76.78540802 112.442499
#> 94    28.0346375 -107.87528992 111.458598
#> 95   130.2844696  -51.66857529 140.155930
#> 96   168.6222382   61.02213287 179.324175
#> 97    69.9804993   84.42430878 109.657349
#> 98   -86.6435852   72.07600403 112.703421
#> 99  -158.0213318   -9.21973324 158.290065
#> 100 -131.5890350 -141.64436340 193.335977
#> 101  -62.3062935 -130.95321655 145.020065
#> 102   -3.1349177   27.50640869  27.684476
#> 103   95.3000793  108.40498352 144.338995
#> 104   52.4139252   43.73036957  68.261005
#> 105   21.7514954  -37.56709671  43.409841
#> 106  -57.1836815   69.69844055  90.154568
#> 107   21.1138000  129.07261658 130.788122
#> 108  -24.4919930   23.65844154  34.052600
#> 109   14.9646254  -49.60564423  51.813704
#> 110   24.6133251  -38.76391983  45.917940
#> 111 -101.5032425  -16.48455811 102.833112
#> 112  -67.3662033    0.40133628  67.367399
#> 113  -10.1563692  -84.66468048  85.271683
#> 114   90.9316864 -127.87138367 156.906540
#> 115   87.5384598  -36.16957474  94.716525
#> 116   -6.5814571   32.22711182  32.892284
#> 117  -47.9273987  -10.39803314  49.042376
#> 118 -169.2823792  -18.68123245 170.310048
#> 119 -117.3540878   70.18990326 136.742841
#> 120  122.8687134  105.32736969 161.835026
#> 121  156.7883453   21.63671494 158.274232
#> 122   57.3051605 -104.89302063 119.525843
#> 123  -65.6591568  -86.07979584 108.262903
#> 124  -33.9830284  -45.72549438  56.970756
#> 125  -13.5699177   15.15062428  20.339225
#> 126   19.9458523   57.91993332  61.258107
#> 127   91.9235153  -11.47993279  92.637582
#> 128  103.8226089  -24.06698418 106.575578
#> 129  -56.9959068   83.78201294 101.330938
#> 130  -67.6752853   97.25795746 118.486516
#> 131  -12.0978470   31.70837975  33.937873
#> 132  -58.5304832  -72.53223419  93.202696
#> 133  -44.9086914  -79.27008057  91.107279
#> 134   -7.4072728   26.90681839  27.907787
#> 135  128.1665192   55.66021347 139.730870
#> 136   67.0956497   47.41166306  82.156509
#> 137  -17.6796932    5.88453293  18.633284
#> 138  -62.3230743  -33.24103165  70.633786
#> 139 -132.4909058  -71.81415558 150.702067
#> 140   76.6262741  -51.90783691  92.552739
#> 141  170.7310028   42.46802521 175.933534
#> 142   69.0220261   58.89309311  90.732775
#> 143  -73.1532059   20.09345245  75.862628
#> 144  -74.6923981  -43.73366165  86.553957
#> 145   55.5769196  114.89539337 127.631287
#> 146  -21.5108986  135.69282532 137.387269
#> 147 -118.6049042   25.22805023 121.258310
#> 148  -10.6536703 -143.02688599 143.423118
#> 149   80.6166992 -128.43836975 151.642563
#> 150  -21.6777592   30.66543770  37.553885
#> 151  -51.7505646   75.59968567  91.615683
#> 152  -76.3029938   88.19949341 116.624601
#> 153  -12.8474035   -3.17974663  13.235051
#> 154   19.5013885  -37.88739014  42.611718
#> 155  -10.6015339  -93.44827271  94.047712
#> 156  -15.1612024  -24.69034195  28.973696
#> 157  -76.7970886   30.01570511  82.454444
#> 158   -6.1436419    7.20260954   9.466885
#> 159  107.4689407    7.83377504 107.754078
#> 160   58.5842323  -16.32789040  60.817039
#> 161 -103.3832397   24.39004898 106.221320
#> 162  -89.6708145  -52.53548050 103.927050
#> 163  -19.0665684  -55.32215881  58.515599
#> 164   58.6544113   31.94513321  66.789457
#> 165   99.6895523   65.08676147 119.055841
#> 166  102.8900833   95.91749573 140.664620
#> 167  190.2398071   24.41347694 191.799901
#> 168  115.0016785  -27.12688065 118.157749
#> 169  -37.6206436  -43.88900375  57.806206
#> 170 -214.3445740  -36.09032822 217.361699
#> 171 -189.9938507  -47.41487885 195.820923
#> 172   12.5146770 -123.23081207 123.864645
#> 173   78.8404846  -90.31777191 119.887956
#> 174   31.6233044   88.81578064  94.277655
#> 175  -46.5845222  206.86853027 212.048831
#> 176 -104.1464691  123.89435577 161.852706
#> 177  -71.9929810  -51.17813873  88.330013
#> 178  -59.5497971 -100.38827515 116.721824
#> 179    2.8453808    0.01813568   2.845439
#> 180   34.6570511   38.03362656  51.455495
#> 181   -4.3145976  -45.80394745  46.006710
#> 182   -0.2931468  -34.59003067  34.591273
#> 183  -45.4036789   52.21525955  69.194851
#> 184   -6.2926502  140.02197266 140.163299
#> 185   72.7428436    9.44209957  73.353081
#> 186   99.9777145  -34.30938339 105.700885
#> 187   36.8711853  -65.96759796  75.572537
#> 188   27.7766304  -98.99563599 102.818661
#> 189   46.2144470   -0.26044229  46.215181
#> 190   27.8961163   10.39067173  29.768429
#> 191  -19.9005909   52.13185120  55.801106
#> 192  -54.9210052   -0.05214188  54.921030
#> 193   13.7985859    0.95721316  13.831747
#> 194   12.3526754   34.75115585  36.881315
#> 195   -6.0986490  -54.46621704  54.806590
#> 196   20.4872322  -25.56856537  32.763977
#> 197  -10.9710236  137.26403809 137.701777
#> 198  -61.1643181  121.61771393 136.132076
#> 199  -19.6619816   26.36090660  32.886029
#> 200   76.4632034  -68.71568298 102.803047
#> 201  130.2209778   21.00172615 131.903660
#> 202   77.2407379   68.31198120 103.114782
#> 203  -16.6380482 -127.91790771 128.995410
#> 204  -84.9272003 -187.27670288 205.633637
#> 205  -58.2531738 -103.15737915 118.468887
#> 206 -103.6245804   53.36646271 116.559140
#> 207 -139.9662476    4.55996704 140.040508
#> 208  -36.6117096  -49.45965195  61.535961
#> 209   28.3265018  -62.33488083  68.469176
#> 210   56.8227730  -41.16561508  70.167196
#> 211   11.1111326   51.05272293  52.247850
#> 212   95.8855820  161.70938110 187.999917
#> 213  131.3482666  241.32481384 274.754496
#> 214   68.0760956   92.33050537 114.713892
#> 215   10.2567120 -139.36497498 139.741892
#> 216  -66.8759766 -320.09075928 327.002279
#> 217  -64.2250061 -141.67504883 155.552791
#> 218   55.0542221   18.65238953  58.128126
#> 219  205.8044128   66.51713562 216.286813
#> 220   69.6318970   16.23121071  71.498624
#> 221 -142.7745209   20.65156746 144.260359
#> 222 -146.8431396  114.42655182 186.162143
#> 223  -88.1625824   12.16742802  88.998243
#> 224  -35.9672432  -34.76118851  50.019824
#> 225  -43.1101151   42.93539429  60.843489
#> 226  -50.1011009   84.09043121  97.884222
#> 227   30.5915928   53.89208221  61.969364
#> 228   64.0706329  -17.71008301  66.473251
#> 229   86.1371231  -14.38996220  87.330836
#> 230  -20.9974232   30.63659859  37.141526
#> 231 -115.1238251    4.78823853 115.223358
#> 232   17.4988899  -56.20755386  58.868500
#> 233  158.0083313  -91.08806610 182.383301
#> 234  116.5345612  -65.91115570 133.882726
#> 235   -3.2418728  -30.14122391  30.315064
#> 236  -53.9461670   34.69013977  64.137312
#> 237  -68.4184265   28.15351868  73.984469
#> 238  -42.6885719   10.05851746  43.857587
#> 239   38.5431366   18.55828094  42.778303
#> 240   86.0390549   45.32783127  97.248811
#> 241   -3.7688286   96.25428772  96.328044
#> 242  -65.2843475   47.69317245  80.849766
#> 243   20.5968609   -4.55651140  21.094845
#> 244   42.0541077  -61.43453217  74.449645
#> 245 -127.3400879 -123.81697845 177.612337
#> 246 -156.1679077 -119.02497101 196.355186
#> 247  -68.7745743  -52.33489990  86.422704
#> 248  112.8241119   69.72891235 132.632581
#> 249   89.3656998  115.80510712 146.277309
#> 250  -91.8667526   72.18183136 116.832003
#> 251  -92.9494705  -39.56701660 101.020557
#> 252  -24.7626781 -111.96282196 114.668495
#> 253  126.2675476 -162.55857849 205.836792
#> 254  150.9990845  -31.63839531 154.278033
#> 255   84.7415619   42.39373779  94.754215
#> 256  -34.7484360   77.10009766  84.568782
#> 257 -122.9718704   55.71531677 135.004731
#> 258  -86.7513199   94.77500916 128.483827
#> 259    7.4068184  184.71987915 184.868317
#> 260   57.1358299   70.63929749  90.853802
#> 261  -49.0911446  -80.27308655  94.094149
#> 262  -97.9625397 -238.00871277 257.380665
#> 263  -48.1304626 -119.12041473 128.476514
#> 264   16.3026867  118.07798004 119.198100
#> 265  142.3766479  195.23471069 241.635474
#> 266  167.6743011   53.10479736 175.882889
#> 267  114.5931244  -87.03873444 143.900401
#> 268  -10.6483727  -14.72822952  18.174394
#> 269  -44.0209503   35.28707504  56.418275
#> 270   32.4969559  -80.76611328  87.058700
#> 271   72.4414749 -119.12498474 139.422126
#> 272   -9.3374901   -0.63780695   9.359248
#> 273  -12.4853964  101.73185730 102.495151
#> 274  -12.6169214   25.76575089  28.689033
#> 275  -98.6115112  -75.25057220 124.043858
#> 276 -104.1315002   42.89272308 112.619514
#> 277  -86.5669556   78.53098297 116.880080
#> 278    7.4551368  -33.40767670  34.229401
#> 279   20.2868233  -89.03759003  91.319481
#> 280   31.6575489   26.01570702  40.975815
#> 281   87.8032455   87.95162964 124.277508
#> 282   13.6450548 -112.60279846 113.426530
#> 283  -58.4272537 -185.41976929 194.407394
#> 284  -16.6712151  -49.76807404  52.486099
#> 285  135.9526978   88.70188141 162.330403
#> 286  180.9846344   20.10962677 182.098421
#> 287    3.9526875  -36.17604065  36.391340
#> 288 -142.6085815   17.55375290 143.684870
#> 289 -112.0072556   53.23265839 124.013472
#> 290  -24.9668388   66.04371643  70.605351
#> 291   13.9365969    8.97211075  16.574906
#> 292    4.6859202  -19.36771774  19.926524
#> 293   -7.6343398  -48.53606033  49.132803
#> 294   19.7684822   37.44206238  42.340299
#> 295   38.6344261  114.21220398 120.569675
#> 296  -19.3021698   59.58039093  62.629041
#> 297  -98.3080521  -88.12310028 132.023308
#> 298  -96.0422821  -91.34265900 132.542828
#> 299   -5.6207299  127.91730499 128.040734
#> 300   56.7821617  153.18560791 163.370880
#> 301   38.2705307   40.13835526  55.459184
#> 302   17.9716759  -91.18412781  92.938293
#> 303   51.9310226   -6.79997683  52.374333
#> 304  -17.1759911   61.28419495  63.645638
#> 305  -88.9138489 -127.95138550 155.811519
#> 306    2.7824285 -234.52476501 234.541270
#> 307   24.5610847 -185.98892212 187.603641
#> 308    6.2263694   30.39042282  31.021694
#> 309  -30.7821884  175.79554749 178.470215
#> 310  -12.2430010  111.87796783 112.545861
#> 311  -13.5147896    0.33964491  13.519057
#> 312  -59.5747032  -78.87183380  98.842862
#> 313    1.1391920   18.10074997  18.136563
#> 314   48.1341324   34.72183990  59.350660
#> 315  -14.5526934  -50.63608170  52.685801
#> 316   10.0076208  -16.92543221  19.662724
#> 317   69.4810867   75.76557159 102.800989
#> 318   52.5817833   76.13485718  92.527620
#> 319  -24.7712822  -74.51299286  78.522624
#> 320  -15.3835602 -117.18181610 118.187275
#> 321   32.8183479   22.18377495  39.612673
#> 322   48.0788651  111.07836151 121.037100
#> 323   18.8269329   25.84134293  31.972307
#> 324  -90.2710953  -59.95889282 108.369458
#> 325 -105.6383820  -53.38583374 118.361797
#> 326  -15.4323263   75.10372925  76.672856
#> 327   63.0534172  161.14173889 173.038705
#> 328   45.7172928  107.92945099 117.212786
#> 329   20.0188694   98.58966064 100.601572
#> 330   62.8045845   23.13451958  66.929977
#> 331   50.8212891  -47.80884171  69.774557
#> 332  -41.5171242  -95.95525360 104.551816
#> 333  -99.9167786  -68.52480316 121.156970
#>  [ reached 'max' / getOption("max.print") -- omitted 667 rows ]