Differentiate x and y and compute change amplitude via an algorithm proposed by Engbert & Kliegl (2003)

Differentiate x and y and compute change amplitude via an algorithm proposed by Engbert & Kliegl (2003)

Usage

diff_ek(x, y, trial, sample_rate, options = NULL)

Arguments

x

vector with x coordinates in degrees of visual angle

y

vector with y coordinates in degrees of visual angle

trial

vector with trial index

sample_rate

sample rate in Hz

options

List with method specific options, see Details.

Value

data.frame with columns x, y, and amp

Details

Method options, please refer to Engbert & Kliegl (2003) for details on parameters and the rationale for default values.

ek_velocity_time_window Time window for velocity computation in milliseconds. Defaults to 20 ms.

See also

Examples

diff_ek(rnorm(1000), rnorm(1000), rep(1, 1000), 250, list("ek_velocity_time_window" = 20))
#>                 x            y       amp
#> 1              NA           NA        NA
#> 2   -148.03674316  205.8607635 253.56169
#> 3    -43.62339783  147.0704803 153.40380
#> 4    -64.86721039   94.4495621 114.57956
#> 5    -37.83142471  -40.8235321  55.65768
#> 6     -9.75629139  -94.7735291  95.27438
#> 7      8.33357811 -144.9267120 145.16611
#> 8      3.52453279  -11.1638708  11.70702
#> 9    -70.07298279   12.2402086  71.13400
#> 10     3.29045081  -95.7246552  95.78119
#> 11   104.67865753 -128.6553650 165.86086
#> 12    73.81729889   -0.1278290  73.81741
#> 13    -3.01073122  182.6752777 182.70009
#> 14    -8.73173809  102.8936920 103.26352
#> 15   -40.41247559  -63.1066551  74.93743
#> 16   -18.98878670 -130.2952118 131.67162
#> 17  -150.32612610 -104.2653580 182.94592
#> 18   -74.34495544    0.8225567  74.34951
#> 19    82.50994110   -4.1057997  82.61203
#> 20   105.39205933  -16.9328728 106.74366
#> 21    39.61030960   56.8209190  69.26466
#> 22   -59.43163681   -6.6561537  59.80321
#> 23    52.33439255   -8.6864204  53.05038
#> 24    32.38088226    4.5036464  32.69257
#> 25   -84.60981750   64.4005051 106.33083
#> 26   -56.39690781   -1.9710143  56.43134
#> 27    72.86018372  -83.1968918 110.59082
#> 28    72.15229797   14.8947239  73.67365
#> 29    -9.23385239  150.0558777 150.33972
#> 30   -72.46547699   88.6975327 114.53601
#> 31   -57.14479065   -9.9432745  58.00341
#> 32   -72.64087677   -1.2711165  72.65200
#> 33   -91.27661133   32.7907715  96.98791
#> 34   -26.15897751   47.8615265  54.54372
#> 35    70.73267365    3.3871288  70.81373
#> 36   166.02352905   55.3834877 175.01755
#> 37   180.67044067   23.9470730 182.25057
#> 38     5.41535139  -32.1853485  32.63775
#> 39    18.79760933 -143.6670074 144.89154
#> 40    83.77397919  -74.8149796 112.31812
#> 41    42.95528793   88.5932159  98.45768
#> 42  -136.36177063   77.8040466 156.99682
#> 43  -255.24172974    0.3229343 255.24193
#> 44   -57.68195724 -191.2373047 199.74713
#> 45   120.86548615 -143.4396820 187.57241
#> 46   186.60177612  -11.2051792 186.93790
#> 47    46.25021744   95.0695572 105.72277
#> 48   -62.49583054  175.5604706 186.35237
#> 49     6.86803007   62.6260834  63.00156
#> 50    60.71588898  -21.5961018  64.44231
#> 51    -8.49944592   -5.6953506  10.23121
#> 52  -127.41846466   29.1015224 130.69952
#> 53   -84.82845306   -4.7207184  84.95971
#> 54    52.60114288  -75.2939453  91.84802
#> 55    70.71397400  -49.3054886  86.20613
#> 56    -9.22798634   69.6630783  70.27162
#> 57   -15.43244839   86.8759766  88.23602
#> 58   -33.59408569  -68.4999847  76.29424
#> 59     7.95479345 -152.0214539 152.22944
#> 60   -26.36241722 -180.2116547 182.12967
#> 61   -88.14541626  -68.4011307 111.57208
#> 62   -97.74520874   75.1834641 123.31536
#> 63   -26.95492935  124.6794128 127.55988
#> 64   152.74018860  130.1135101 200.64668
#> 65   154.43434143  -15.3084545 155.19122
#> 66    96.04911041  -49.7277374 108.15858
#> 67    13.98439789  -10.7810993  17.65773
#> 68    60.27598190  -43.8289185  74.52629
#> 69   -32.53069687  -43.3630638  54.20887
#> 70   -66.53470612   57.3164482  87.81823
#> 71   -76.10534668   89.8275070 117.73277
#> 72  -143.79721069  -57.7001114 154.94173
#> 73   -19.47476578 -229.2860718 230.11165
#> 74     8.36611938 -160.8332367 161.05068
#> 75    28.58185005  134.1823578 137.19266
#> 76    21.91023064  287.6563110 288.48953
#> 77    -0.53538620  162.5952454 162.59613
#> 78    91.04749298  -48.9479980 103.37095
#> 79    65.12182617  -93.1260757 113.63678
#> 80    50.64701462  -64.6145172  82.09845
#> 81     7.38728142  -23.6034145  24.73243
#> 82  -138.61523438   -5.5148220 138.72489
#> 83  -140.14746094   -6.3250470 140.29012
#> 84  -133.22825623  -13.7572165 133.93666
#> 85   -15.92920876  -39.1176529  42.23660
#> 86   115.93529510   -1.8246479 115.94965
#> 87   113.77887726   37.5821075 119.82507
#> 88     9.98859501   85.4520111  86.03382
#> 89   -10.45479107  116.4233475 116.89182
#> 90    89.40964508   17.6511250  91.13532
#> 91     9.95859337 -111.4110107 111.85521
#> 92  -149.96971130 -140.3367004 205.39061
#> 93   -59.51852036  -74.4323654  95.30284
#> 94   125.18811035   31.3456726 129.05276
#> 95    95.76403046  123.0338898 155.91051
#> 96    -4.19098806  140.2291870 140.29180
#> 97    41.29553986   -2.9092257  41.39789
#> 98    50.71747208  -31.7880478  59.85601
#> 99   -39.23697281   77.3535767  86.73590
#> 100 -124.35186005   64.1038361 139.90242
#> 101 -150.92073059  -80.0047607 170.81519
#> 102  -87.52608490 -192.3189392 211.29929
#> 103    3.45507479  -73.6919327  73.77288
#> 104   74.37302399   91.5776672 117.97379
#> 105   86.16448212  206.0668488 223.35591
#> 106   27.65031433  209.6203003 211.43607
#> 107  -15.78999805   33.4586945  36.99741
#> 108  -56.18442154 -205.3606567 212.90770
#> 109  -72.15676117 -330.4657898 338.25174
#> 110   48.35994339 -165.2770996 172.20686
#> 111   77.29987335  102.1169281 128.07473
#> 112   10.21462154  216.8745117 217.11493
#> 113  -74.65219116  160.8080902 177.29126
#> 114  -69.37572479 -104.5002899 125.43246
#> 115   14.15878296 -163.7595825 164.37053
#> 116   40.84291458 -126.3335953 132.77169
#> 117   76.47266388    5.8585310  76.69675
#> 118   62.59990692  107.9322968 124.77231
#> 119   15.43378162   68.6025696  70.31724
#> 120 -102.04600525   90.7251511 136.54465
#> 121 -144.73269653   65.7522430 158.96827
#> 122   20.04824829   46.2254524  50.38576
#> 123  180.32641602  -87.9232407 200.61932
#> 124  133.05580139  -95.0763092 163.53394
#> 125  -80.37734985   59.2743263  99.86974
#> 126 -165.54512024  113.6589966 200.80726
#> 127  -15.77695656   -7.4405975  17.44348
#> 128   96.43311310 -122.5386658 155.93290
#> 129   38.09603882  -83.1287918  91.44235
#> 130  -18.07139969   55.3462563  58.22185
#> 131    9.13555431   78.8989868  79.42612
#> 132  121.81773376   63.0725136 137.17763
#> 133  116.50546265  -60.6758995 131.35862
#> 134   10.02440739 -177.1731110 177.45647
#> 135  -47.93382263  -71.8314133  86.35626
#> 136  -92.60289001   32.9403610  98.28714
#> 137  -72.67231750   99.8202667 123.47207
#> 138  -45.81496429   39.0871086  60.22303
#> 139   41.68318176   -2.5588899  41.76165
#> 140   14.90186405  -25.9928207  29.96151
#> 141  -87.43692017    2.5035608  87.47275
#> 142  -14.05987263   32.5304108  35.43879
#> 143  145.17193604   16.9666119 146.16004
#> 144  131.41140747  -64.2341080 146.27022
#> 145  -48.24366760  -61.3139420  78.01827
#> 146 -186.74919128   53.1354294 194.16136
#> 147 -109.97779083   89.1336288 141.56242
#> 148   38.85809708  131.9927673 137.59376
#> 149   88.72646332   60.0708923 107.14895
#> 150  -60.37260437   39.2881050  72.03059
#> 151 -146.77854919  -40.4151497 152.24102
#> 152  -11.68393326  -60.7394981  61.85306
#> 153  164.17610168   17.3730545 165.09275
#> 154  167.65518188  -18.9679775 168.72476
#> 155  -14.35221481  -37.8448601  40.47492
#> 156  -36.85848999  -72.1638565  81.03191
#> 157  -32.88826752  -46.8192444  57.21608
#> 158    1.27756178  -44.9501724  44.96832
#> 159   -4.85225677  -82.1293106  82.27252
#> 160  -61.21609879    4.1867051  61.35910
#> 161  -14.67985058   13.6440592  20.04142
#> 162   -0.07113546  -66.4673996  66.46744
#> 163    3.39124966  -13.2271862  13.65500
#> 164    9.55591011   37.7551689  38.94571
#> 165  -10.56526470   82.5812531  83.25436
#> 166    2.57104731   94.9249573  94.95977
#> 167  -55.05803680   64.8408737  85.06307
#> 168  -34.10989380  -36.5627747  50.00321
#> 169   99.38402557 -163.3031006 191.16769
#> 170   76.23155975  -36.8705292  84.67991
#> 171   52.94832230  102.0497208 114.96813
#> 172  -49.08829498   52.2805481  71.71413
#> 173  -21.60476875    2.1929758  21.71578
#> 174   29.31210899   -1.7062540  29.36173
#> 175   24.80570030   14.5640087  28.76514
#> 176  104.30760193   46.4076729 114.16544
#> 177   42.45343399   67.0967407  79.39941
#> 178  -47.55519867  -17.6165714  50.71332
#> 179 -133.23637390 -191.9448547 233.65521
#> 180  -85.02392578 -146.2907410 169.20416
#> 181  -61.84287262   42.1337929  74.83179
#> 182    3.26778793  137.2701874 137.30908
#> 183  118.54824066   51.6454506 129.30947
#> 184  176.46203613   -4.8815808 176.52954
#> 185  147.87774658   20.5431728 149.29786
#> 186  -18.50273323   54.4847260  57.54074
#> 187 -114.54690552  -36.1132736 120.10480
#> 188 -114.82644653  -61.4828453 130.25073
#> 189  -40.99238205    1.5089043  41.02014
#> 190   21.00359917    1.6937838  21.07178
#> 191  -96.68663788  -28.2132225 100.71888
#> 192  -67.16267395  -86.2114105 109.28510
#> 193   61.35814667   -6.6666074  61.71925
#> 194   44.28789139   76.0419235  87.99882
#> 195   -2.89125490   93.1280441  93.17291
#> 196  -32.10513687    4.6218319  32.43611
#> 197  102.25756073 -133.8630524 168.45155
#> 198  117.72186279  -45.1985359 126.10053
#> 199   -0.76699018  130.2427521 130.24501
#> 200  -28.73829651  159.1613312 161.73503
#> 201  -87.95711517   65.4249725 109.62154
#> 202 -136.13844299  -67.3483505 151.88639
#> 203  -94.29679871  -32.2464371  99.65801
#> 204   25.84502792  -10.6269217  27.94453
#> 205  107.08683014   -7.8236341 107.37224
#> 206    2.60590267  -69.7422333  69.79090
#> 207  -48.41976166 -110.5166245 120.65818
#> 208   -6.08187675  -24.8282108  25.56226
#> 209   98.59023285   55.2998123 113.04027
#> 210   55.30258942  111.5712891 124.52521
#> 211  -32.27755737   41.1301384  52.28316
#> 212  -22.41018867   31.9280720  39.00793
#> 213  -46.78384018   42.9498940  63.50922
#> 214   58.95096588   68.3031540  90.22493
#> 215   65.82638550   66.8814545  93.84158
#> 216   30.67578506  -37.1055336  48.14379
#> 217  -35.26658249  -42.8276749  55.47920
#> 218 -116.93520355    2.7774346 116.96818
#> 219  -50.87775803   10.6725168  51.98508
#> 220    7.65480900 -103.9913254 104.27268
#> 221   14.80478764 -221.9575500 222.45075
#> 222   48.80106735 -181.6756287 188.11586
#> 223  -27.42279243    8.0380125  28.57655
#> 224 -100.94853973  175.9125214 202.81968
#> 225 -121.10878754   68.1517105 138.96760
#> 226  -32.88032913  -48.0308762  58.20723
#> 227  115.76257324  -41.0759277 122.83406
#> 228   73.23170471   43.3890419  85.12045
#> 229   39.66710663   88.1779861  96.68938
#> 230   15.10904980   18.3043270  23.73461
#> 231   28.88078690   66.8704453  72.84062
#> 232   13.97468281  126.9234238 127.69044
#> 233   13.96195698   94.1263962  95.15626
#> 234   32.24586105  -79.6806107  85.95810
#> 235   12.21419907 -124.7481079 125.34463
#> 236    9.39024639  -66.1533737  66.81651
#> 237  -29.69057655  -11.6762877  31.90401
#> 238  -73.59584808  -28.0564899  78.76240
#> 239 -103.48979187 -146.7783203 179.59402
#> 240  -17.36432076  -85.6737213  87.41571
#> 241  116.04266357  -13.1855173 116.78937
#> 242   94.32033539  109.7422256 144.70550
#> 243  -28.44374847  141.1335449 143.97126
#> 244  -79.91445160  -12.4062128  80.87171
#> 245  -20.27728653  -34.7206039  40.20807
#> 246  -11.92429543  -45.8565788  47.38159
#> 247  -21.73946953   60.2109985  64.01538
#> 248  -19.18189430   20.2472763  27.89081
#> 249   14.54584408   13.1885490  19.63465
#> 250  110.69893646  116.5846558 160.76765
#> 251   98.79455566    7.9057140  99.11037
#> 252   48.62723160  -84.6431274  97.61694
#> 253 -132.37806702 -170.7220917 216.03237
#> 254 -139.65530396  -10.5880957 140.05610
#> 255   49.82602692   79.1669159  93.54161
#> 256  146.55226135    0.9289264 146.55521
#> 257  139.97636414  -32.7747116 143.76218
#> 258   32.38032913  -41.3965912  52.55629
#> 259  -85.07211304   54.7277145 101.15526
#> 260 -178.68116760    3.0044918 178.70643
#> 261 -187.00323486   20.2711067 188.09872
#> 262  -18.70156860  -19.2878647  26.86578
#> 263  152.02728271   11.1990995 152.43922
#> 264  164.38891602   94.6615829 189.69589
#> 265   69.04249573    5.4777966  69.25946
#> 266 -104.48538971  -34.5783615 110.05844
#> 267 -163.10279846 -139.6372070 214.71160
#> 268  -92.88237762  -47.4990692 104.32304
#> 269   63.94679260   14.9626770  65.67400
#> 270  125.64134216   72.7757111 145.19659
#> 271   94.42511749   24.0251999  97.43363
#> 272  -29.01020813 -120.6402969 124.07930
#> 273 -128.66865540  -66.8676071 145.00655
#> 274  -97.73609161   51.7653313 110.59834
#> 275   74.46704102  228.6198730 240.44206
#> 276  208.13771057  161.2250519 263.27709
#> 277  178.37174988  -29.1446552 180.73708
#> 278   53.90274048  -50.9316177  74.15885
#> 279 -114.24414062  -46.2527275 123.25193
#> 280 -152.39982605  -43.4119263 158.46231
#> 281 -108.90306854  -54.1923447 121.64164
#> 282   21.17967796  -41.3089447  46.42206
#> 283  148.13948059   -2.0738921 148.15400
#> 284   77.40267181   -5.8250837  77.62155
#> 285  -85.57126617  -12.3027067  86.45113
#> 286 -238.04814148  -37.4441643 240.97507
#> 287 -181.89389038 -131.7091217 224.57222
#> 288   30.74142265 -141.7378235 145.03326
#> 289   74.82930756  100.9310074 125.64431
#> 290   36.00285339  235.6084137 238.34330
#> 291    1.26426589  137.6322174 137.63802
#> 292    0.41907090  -76.9786758  76.97982
#> 293   12.07301903 -190.4088593 190.79122
#> 294  -22.60583878  -66.5535889  70.28801
#> 295  -41.81657791   56.7652626  70.50476
#> 296   27.86227036  152.0686035 154.60002
#> 297   76.01909637  117.9763260 140.34713
#> 298  105.69941711   -9.8265648 106.15521
#> 299   34.84309769  -77.0919342  84.60028
#> 300  -52.41058731  -63.3733482  82.23777
#> 301  -41.36642456   39.4015427  57.12847
#> 302  -10.19580269   63.3544998  64.16967
#> 303   49.44040298   35.0360336  60.59602
#> 304   73.82360840  -76.7922287 106.52216
#> 305   65.44185638 -174.0915375 185.98521
#> 306   44.23889923  -61.9644623  76.13590
#> 307  -45.23918152   24.1129436  51.26419
#> 308  -27.15165901   96.4336777 100.18317
#> 309   -9.47002506  -15.4516811  18.12280
#> 310  -28.18081093   18.0835705  33.48393
#> 311  -90.22799683   42.5687675  99.76568
#> 312 -101.10459900  -64.4539337 119.90183
#> 313   -9.23685741 -138.5413818 138.84896
#> 314  -41.24295044    0.1222917  41.24313
#> 315  -21.29559135  223.0525055 224.06678
#> 316   -2.04106808  250.7608337 250.76914
#> 317   25.39597130   54.5278053  60.15178
#> 318   -6.07680607 -120.9612656 121.11381
#> 319   21.57141685    1.4620864  21.62091
#> 320   97.46353912  -17.5023575  99.02259
#> 321   45.99114609  -87.5279388  98.87530
#> 322   29.05008507 -121.6866379 125.10614
#> 323   46.83474731  -66.4184799  81.27059
#> 324   37.91622543  -42.7583313  57.14818
#> 325  -65.46844482 -108.2543259 126.51133
#> 326 -163.38850403   -8.1471176 163.59150
#> 327  -88.94184875  134.9095001 161.58968
#> 328   43.93496704  118.1760025 126.07874
#> 329  160.42935181   43.9187927 166.33231
#> 330  193.24638367  -15.3035069 193.85139
#> 331   37.82114792   13.6124315  40.19624
#> 332  -98.70739746   21.0903187 100.93538
#> 333 -138.16270447  -18.9135780 139.45127
#>  [ reached 'max' / getOption("max.print") -- omitted 667 rows ]