Differentiate x and y and compute change amplitude via an algorithm proposed by Engbert & Kliegl (2003)

Differentiate x and y and compute change amplitude via an algorithm proposed by Engbert & Kliegl (2003)

Usage

diff_ek(x, y, trial, sample_rate, options = NULL)

Arguments

x

vector with x coordinates in degrees of visual angle

y

vector with y coordinates in degrees of visual angle

trial

vector with trial index

sample_rate

sample rate in Hz

options

List with method specific options, see Details.

Value

data.frame with columns x, y, and amp

Details

Method options, please refer to Engbert & Kliegl (2003) for details on parameters and the rationale for default values.

ek_velocity_time_window Time window for velocity computation in milliseconds. Defaults to 20 ms.

See also

Examples

diff_ek(rnorm(1000), rnorm(1000), rep(1, 1000), 250, list("ek_velocity_time_window" = 20))
#>                x             y        amp
#> 1             NA            NA         NA
#> 2    106.7516174   91.94919586 140.892024
#> 3     53.3004799   18.61266136  56.456818
#> 4    109.1394119   10.67230988 109.659972
#> 5    105.9767227   51.76063156 117.941633
#> 6     25.6791229  129.72520447 132.242376
#> 7   -155.8343811    9.72913647 156.137793
#> 8   -180.6348724 -171.37490845 248.994611
#> 9    127.6012039 -126.55084229 179.714170
#> 10   236.2662201  -13.81827831 236.669963
#> 11   112.6465759   65.20629120 130.158025
#> 12   -69.2707825   36.97498703  78.521277
#> 13  -156.7835693   13.31303120 157.347782
#> 14   -79.8174973   20.88633347  82.504980
#> 15  -149.4924164   66.45052338 163.596010
#> 16   -58.9960403  115.30646515 129.522638
#> 17    62.6837006   46.31139374  77.935817
#> 18   154.9259644  -26.97948647 157.257582
#> 19   142.0158691  -57.86450577 153.351909
#> 20    21.1152859   61.13862610  64.682199
#> 21   -47.7940292   43.76888275  64.807286
#> 22   -57.7906609 -102.71272278 117.854418
#> 23    39.0038986 -133.93426514 139.497998
#> 24   -27.6583176 -180.80284119 182.906123
#> 25   -45.4522514 -105.55448151 114.924565
#> 26    -9.3410063 -103.32279205 103.744175
#> 27    35.7686234   66.39331818  75.415298
#> 28    29.5135136  319.32540894 320.686396
#> 29   -64.7138977  279.02752686 286.433673
#> 30   -25.1183453  112.30078125 115.075613
#> 31   -12.6419039  -46.95597839  48.627992
#> 32    43.6694870  -87.32222748  97.632963
#> 33    92.0733337 -128.67953491 158.227436
#> 34   101.2466507 -132.97612000 167.133278
#> 35    40.4890480  -17.28809547  44.025461
#> 36   -88.3164825   65.64364624 110.040399
#> 37   -74.6542435    6.82351875  74.965435
#> 38   -94.7940903 -134.09165955 164.214776
#> 39   -73.7760468 -172.19802856 187.336772
#> 40  -105.8735428  -75.57508087 130.079975
#> 41   -24.7069416   40.70814514  47.619177
#> 42    98.8846283  150.81417847 180.341582
#> 43   113.0360260  182.19302368 214.409517
#> 44    76.7205811  148.93452454 167.533699
#> 45    95.4782562   61.16926575 113.392136
#> 46   190.5465698  -67.82389832 202.257451
#> 47    94.4112701 -142.22612000 170.709570
#> 48   -70.3674545  -78.98133087 105.781044
#> 49  -107.6598129  -19.99911690 109.501598
#> 50   -61.6820183  -25.83128166  66.872464
#> 51   -89.1706314   87.66160583 125.043827
#> 52  -138.7463989   49.24409103 147.226165
#> 53   -75.9893570   53.39793777  92.874766
#> 54   124.3157806  -27.46423531 127.313383
#> 55   158.0284882  -78.59442902 176.493873
#> 56    67.0959625   45.36847305  80.994855
#> 57   -33.6729393   85.06634521  91.488523
#> 58   -30.6555386   98.64482880 103.298423
#> 59     3.6978617   -5.80703354   6.884462
#> 60   -73.5140762 -122.16608429 142.579352
#> 61  -161.3286896  -71.56325531 176.488656
#> 62  -105.9239426   54.28927612 119.026077
#> 63   116.3796997    1.17513013 116.385632
#> 64   196.0454712 -102.82321930 221.373985
#> 65   146.9974213 -181.63169861 233.662825
#> 66    20.5162907 -122.11919403 123.830593
#> 67   -36.0090141  -10.74420071  37.577745
#> 68    48.2878418   47.21875000  67.537590
#> 69    -8.4584723   86.60034180  87.012441
#> 70   -45.1687622  111.47686005 120.280121
#> 71   -98.9624786   95.52130890 137.542330
#> 72  -101.0283432   20.03461647 102.995689
#> 73   -53.9543839   26.55123329  60.133547
#> 74   -83.4858627   44.85530090  94.772819
#> 75     0.9832157   22.42127037  22.442818
#> 76    53.1224785   -8.63454533  53.819635
#> 77    31.5943565  -17.58770943  36.159797
#> 78    32.9135666  -16.19085884  36.680332
#> 79    37.7306747 -140.49708557 145.475204
#> 80    39.4178963 -130.96089172 136.764490
#> 81   -68.6016541   -8.74560738  69.156869
#> 82   -74.4640808   94.90097809 120.627920
#> 83    35.5141220  130.24406433 134.999145
#> 84    66.1794510   42.97882462  78.910703
#> 85    11.8221369    4.19083786  12.542968
#> 86  -156.1737213  -71.94383240 171.948092
#> 87  -102.5512619  -51.14920044 114.599311
#> 88    26.7806892  -68.04792786  73.128146
#> 89   146.4012909  -17.88956261 147.490252
#> 90    76.1436615  128.16345215 149.076248
#> 91    41.8920555  192.85852051 197.355905
#> 92    39.7094727   91.04534149  99.328226
#> 93    35.3583183  -82.79096222  90.025297
#> 94   -55.8011360  -68.87567902  88.643251
#> 95   -98.4705963   28.85358810 102.610857
#> 96   -33.4966049  -13.77598286  36.218783
#> 97   -23.4417591  -49.73769760  54.985040
#> 98   129.6645355    4.75125742 129.751556
#> 99   105.1044617   35.82201004 111.041273
#> 100   44.5983391   44.66936874  63.121822
#> 101 -137.2423096  -42.21774673 143.588961
#> 102 -223.5696869  -73.24647522 235.262515
#> 103 -209.4367371  -54.67131042 216.454843
#> 104 -165.8323059  -42.08898544 171.090141
#> 105   52.7644348   -8.59825325  53.460411
#> 106  158.8261871   29.68890190 161.577191
#> 107  112.0836334  116.56409454 161.709397
#> 108  -25.3546429   60.12777328  65.254939
#> 109  -12.9484959  -31.63137436  34.179049
#> 110  164.6648865 -108.79605865 197.360349
#> 111  301.7207642   -3.39564705 301.739871
#> 112  185.5673370   25.58793831 187.323194
#> 113  -86.7533798  -25.94744682  90.550643
#> 114 -206.0993500   49.65737534 211.997163
#> 115 -167.8008575   26.91980362 169.946473
#> 116  -47.7500916   18.35102272  51.154973
#> 117   30.0256863 -145.08134460 148.155791
#> 118  101.6472473 -120.00548553 157.268813
#> 119  115.1806641   93.83470917 148.564929
#> 120   13.9916840  200.33673096 200.824732
#> 121   24.1163883   61.71409225  66.258806
#> 122  -16.0101891 -167.71490479 168.477344
#> 123  -48.4412651 -201.04394531 206.797544
#> 124  -45.2652855  -39.92045975  60.353866
#> 125  -52.8275032   12.86367607  54.371125
#> 126    3.9201360   -6.64876890   7.718393
#> 127  -74.2814255   17.09677696  76.223552
#> 128  -86.4177170   62.47554779 106.635903
#> 129   -3.8830004   66.67569733  66.788669
#> 130  130.6273193   33.52458954 134.860649
#> 131  131.4575806  -15.49845219 132.368038
#> 132  -78.7472992   17.33547974  80.632847
#> 133 -181.1015625   89.95553589 202.212201
#> 134  -51.7878227  114.31264496 125.496452
#> 135   63.1445732   50.26559448  80.708532
#> 136  -15.4595175  -60.56329727  62.505277
#> 137  -62.4559746  -50.91063309  80.576928
#> 138   40.5134392  -31.90482712  51.567982
#> 139   53.0605698   14.96632481  55.130889
#> 140  -45.1185150  -56.71774673  72.474707
#> 141 -121.9559860  -49.81000137 131.735716
#> 142  -34.4179573   57.51778030  67.029030
#> 143  121.5945206   84.48143768 148.061949
#> 144   83.1749268   87.52192688 120.740035
#> 145   19.3532238  -53.07408142  56.492525
#> 146    6.7371631 -118.86729431 119.058066
#> 147   80.9525452  -38.55754089  89.666039
#> 148   65.4746628   71.70404816  97.099959
#> 149  -14.3394022  120.24279022 121.094785
#> 150 -157.8955841   42.57195282 163.534053
#> 151  -81.2623978  -93.15425110 123.617522
#> 152   86.1249390 -125.33434296 152.073017
#> 153  214.2683868  -46.78882599 219.317432
#> 154   83.5888062  -24.61828232  87.138673
#> 155 -144.9865265   -3.49539137 145.028654
#> 156  -51.2336769  -31.73513222  60.266145
#> 157   49.3070984  -57.27791595  75.577441
#> 158  -48.8910637  -63.19224167  79.897406
#> 159 -167.7622986  -23.18695831 169.357090
#> 160  -94.1147461  106.24855804 141.937809
#> 161  102.4964371  179.87170410 207.024997
#> 162  167.2020874   81.01274109 185.794516
#> 163   42.0058479   10.41043186  43.276649
#> 164   -3.0531235   51.18743515  51.278408
#> 165 -132.3994293   67.61914062 148.667270
#> 166 -104.8819504   -5.31814051 105.016695
#> 167   37.6519699  -43.16189194  57.276695
#> 168  159.2505493   59.60020447 170.038001
#> 169  152.6464233   85.41143799 174.917250
#> 170   58.6354866   21.99556923  62.625277
#> 171   -8.3047686  -87.58186340  87.974724
#> 172    6.1410241 -155.87091064 155.991836
#> 173   36.4698524 -195.43540955 198.809078
#> 174   25.7579613 -177.51187134 179.370948
#> 175   29.8700066  -35.80469513  46.628248
#> 176  -37.0937347   80.55327606  88.683569
#> 177  -79.9486465   80.31143951 113.321284
#> 178  -61.1877937    7.24638557  61.615389
#> 179  -15.6868563   -8.56524658  17.872910
#> 180  -44.7199974   33.42424774  55.830623
#> 181 -141.3657837   36.83636093 146.086284
#> 182  -93.0175552   29.29394913  97.521285
#> 183  107.7353668    3.44865584 107.790549
#> 184  169.3327789   51.35825729 176.949882
#> 185  -63.2848320   34.33726120  72.000121
#> 186 -141.2564240  -32.82141876 145.019388
#> 187   15.2150640    3.68464994  15.654866
#> 188  149.1471558  -54.71942902 158.868153
#> 189   49.6551399  -34.56941223  60.503530
#> 190  -57.6083298   17.16276169  60.110565
#> 191  -55.4483147   51.74891281  75.845010
#> 192  -30.0340176   30.52762794  42.824973
#> 193  -50.4067688 -103.92554474 115.504810
#> 194   -3.5281148  -62.78689194  62.885940
#> 195  107.1139374   65.72154236 125.669076
#> 196   61.8185501   53.32327271  81.638867
#> 197   -1.5123576   10.41020489  10.519486
#> 198 -106.7941742  -70.98193359 128.231940
#> 199 -108.2932663  -37.56259918 114.622774
#> 200  -73.1513672   20.73993874  76.034647
#> 201  -29.5685349  -13.54574299  32.523613
#> 202   23.1793118    2.10376334  23.274585
#> 203   47.8577652   50.36844254  69.479103
#> 204  134.3498993  180.24588013 224.807635
#> 205   64.8174515   64.93652344  91.749954
#> 206    6.1067653  -93.81306458  94.011615
#> 207  -35.8661880  -81.43636322  88.984632
#> 208  -42.1999397   -7.52376986  42.865394
#> 209  -51.7209167   -3.79400086  51.859885
#> 210   15.2724133 -149.78131104 150.557922
#> 211  123.9509888 -100.41047668 159.518373
#> 212  127.4488373   68.68818665 144.780085
#> 213   40.0869446   90.72274017  99.184569
#> 214  -87.2974167   48.17602921  99.708419
#> 215 -131.4947052   -5.56670952 131.612483
#> 216 -108.1788712  -12.58722687 108.908707
#> 217   29.5152836   53.85984039  61.416890
#> 218  108.6425171   68.34385681 128.351390
#> 219   97.4128799   14.64998341  98.508331
#> 220  -67.5999298  -33.28385544  75.349622
#> 221  -76.8290710  -44.32260895  88.697237
#> 222  -36.3591423   56.90711594  67.530786
#> 223   61.5438766   22.12606049  65.400392
#> 224   18.6369762 -125.19663239 126.576197
#> 225  -50.7111626  -65.21076202  82.607902
#> 226   26.0977859   38.00604630  46.103731
#> 227  106.3341370  105.02397919 149.455628
#> 228  140.2843628   68.76107788 156.229921
#> 229  -52.8139572   31.25828934  61.370960
#> 230 -109.0739670   55.87451172 122.552402
#> 231  -50.4531860  -84.79103088  98.666321
#> 232   40.6078987 -157.45422363 162.606377
#> 233   44.0878448 -172.10604858 177.663249
#> 234 -129.4271698  -89.37783813 157.288875
#> 235 -123.3105087   42.40784454 130.399029
#> 236   -6.0282540   81.43805695  81.660866
#> 237  105.7836227   59.89778900 121.564468
#> 238   56.2041283   13.57196999  57.819568
#> 239  -13.1302080   40.08443069  42.180137
#> 240  -50.7697906   31.41136169  59.701301
#> 241  -23.5421410  -51.46393204  56.593009
#> 242   64.2287292  -90.14810944 110.688804
#> 243   50.6879845   24.98197556  56.509918
#> 244   12.5499878   20.37618637  23.930967
#> 245   -9.2229757  -31.64303780  32.959750
#> 246   16.4689522    9.15142536  18.840779
#> 247  -20.7124691   73.95075989  76.796623
#> 248  -79.9448547   55.46124268  97.299174
#> 249  -72.9103317    0.02038919  72.910335
#> 250   11.3771648   70.79301453  71.701400
#> 251   -1.5317307  163.82690430 163.834065
#> 252   19.5276566   22.79233360  30.013661
#> 253   44.0620766  -65.06263733  78.578708
#> 254  -28.0377731  -84.20306396  88.748367
#> 255  -72.8151703 -100.03266907 123.727862
#> 256  -28.5313873 -115.86928558 119.330346
#> 257   49.1834679   -3.19596195  49.287196
#> 258   30.8072720  156.39817810 159.403507
#> 259   -3.9405327   83.84839630  83.940940
#> 260  -30.1099224  -47.92000198  56.594470
#> 261  -41.7633781 -158.44851685 163.860039
#> 262  -16.2757378   17.81166840  24.127892
#> 263   29.6458149  116.85495758 120.556856
#> 264   21.5802784   96.29483032  98.683346
#> 265  -60.1206703  -52.08094406  79.541937
#> 266  -99.0720673 -169.41026306 196.252673
#> 267  -57.7657051 -105.27636719 120.083264
#> 268  -21.3772774   13.50858688  25.287742
#> 269   70.9168701   53.71784592  88.965215
#> 270  227.3916016   19.38703537 228.216559
#> 271  271.5729675   70.46093750 280.564824
#> 272   92.6880417   11.81169128  93.437622
#> 273  -97.1816788   60.04135132 114.233281
#> 274  -82.3292313   35.57110596  89.685037
#> 275   14.4201603  102.24243927 103.254334
#> 276   -5.2363939   49.12544250  49.403734
#> 277 -125.2439804  -76.19229889 146.599185
#> 278 -191.9046783  -67.21833038 203.336444
#> 279  -68.2540894  -40.18964005  79.207499
#> 280   20.3852844  -23.58815002  31.176283
#> 281   83.1912994  -65.74082184 106.031354
#> 282  107.2947769  -61.45327377 123.647378
#> 283   71.1989594   47.19351196  85.419666
#> 284   85.1205215  124.71154022 150.991627
#> 285   10.5131760   10.78720760  15.062892
#> 286    5.5697865  -88.25298309  88.428567
#> 287  -59.0374718  -82.21895599 101.219464
#> 288  -64.3030472   68.09674835  93.659217
#> 289  -46.4359550  158.32952881 164.998599
#> 290   64.0682983   39.01101303  75.010706
#> 291   15.4505482 -102.13213348 103.294202
#> 292  -74.4308701  -82.02197266 110.759011
#> 293  -74.0864944  -67.60864258 100.298241
#> 294  -32.3114471   46.16956711  56.352982
#> 295  106.4545670   15.99134254 107.648957
#> 296  116.4890900  -26.23766136 119.407382
#> 297  136.7146301   32.00760651 140.411456
#> 298   56.0265732   10.15633011  56.939687
#> 299 -111.5168915   -4.07916451 111.591472
#> 300 -139.3864594  -81.04300690 161.234469
#> 301  -88.5384140  -91.01500702 126.975518
#> 302   47.0530891   29.72163773  55.654011
#> 303   76.5540924   15.25142288  78.058535
#> 304   12.7688046  -65.19910431  66.437682
#> 305   22.0615406  -34.15643311  40.661696
#> 306  -55.9365692   82.57173157  99.734601
#> 307    5.6483936  173.99195862 174.083618
#> 308   -8.0674305   42.79470825  43.548484
#> 309  -38.2705078  -63.51187897  74.151133
#> 310  -39.6686974  -45.07374573  60.043718
#> 311   25.1066780   -4.17211676  25.450969
#> 312   97.4872513   57.04888535 112.952820
#> 313   34.5778999   -4.81068468  34.910942
#> 314   -0.1299119  -44.65761566  44.657805
#> 315   22.4965534  -74.58203125  77.901055
#> 316  -21.6775208  -30.13898468  37.125103
#> 317 -113.2345963  114.61383057 161.116119
#> 318 -146.7039795   84.71233368 169.405540
#> 319  -59.5236320   18.01263237  62.189370
#> 320   28.0007896  -37.66842270  46.935640
#> 321  117.2066650 -114.91901398 164.145613
#> 322  123.3919144  -99.19403839 158.319366
#> 323   62.3501472   -9.38067055  63.051866
#> 324   11.6823502   51.86967468  53.168980
#> 325  -57.5499077  -13.08530712  59.018786
#> 326  -42.4747162  -96.03043365 105.004503
#> 327    0.1532933   -1.02693331   1.038312
#> 328   89.2078476   86.66732788 124.375503
#> 329   50.6332779   25.40080261  56.647415
#> 330  -29.0256863  -58.07488251  64.924436
#> 331 -185.7177277 -104.04988861 212.878965
#> 332 -114.7435913   94.52302551 148.663022
#> 333   30.0153484  225.01011658 227.003246
#>  [ reached 'max' / getOption("max.print") -- omitted 667 rows ]