Differentiate x and y and compute change amplitude via an algorithm proposed by Engbert & Kliegl (2003)

Differentiate x and y and compute change amplitude via an algorithm proposed by Engbert & Kliegl (2003)

Usage

diff_ek(x, y, trial, sample_rate, options = NULL)

Arguments

x

vector with x coordinates in degrees of visual angle

y

vector with y coordinates in degrees of visual angle

trial

vector with trial index

sample_rate

sample rate in Hz

options

List with method specific options, see Details.

Value

data.frame with columns x, y, and amp

Details

Method options, please refer to Engbert & Kliegl (2003) for details on parameters and the rationale for default values.

ek_velocity_time_window Time window for velocity computation in milliseconds. Defaults to 20 ms.

See also

Examples

diff_ek(rnorm(1000), rnorm(1000), rep(1, 1000), 250, list("ek_velocity_time_window" = 20))
#>                 x             y        amp
#> 1              NA            NA         NA
#> 2      83.5964584  353.71929932 363.463493
#> 3      36.2506027    7.41033697  37.000261
#> 4      -1.9237646  -85.10935974  85.131099
#> 5     -51.0079002   47.53446960  69.723251
#> 6      15.9906721  118.30906677 119.384827
#> 7     -17.9623680   27.57589722  32.910132
#> 8     -18.1186352  -37.53885269  41.682735
#> 9     -34.0242195  -70.46408081  78.248541
#> 10      3.8607659 -102.30443573 102.377259
#> 11     39.4519424  -98.58553314 106.186454
#> 12    -49.2111359  -30.40192413  57.844731
#> 13     -4.3369188  128.18641663 128.259761
#> 14    140.7018127  117.41658020 183.258433
#> 15    156.8242340   47.53774261 163.870917
#> 16    -49.7212753  -43.63943100  66.155916
#> 17   -179.4933014  -94.01418304 202.624065
#> 18    -37.7476273  -60.60478973  71.399047
#> 19    -19.0446453  -13.77004623  23.501334
#> 20    -50.1590385   60.04493713  78.238888
#> 21    -84.1467438  120.39293671 146.884763
#> 22     56.5592842  105.32484436 119.550305
#> 23    100.9957504  -17.21533775 102.452474
#> 24    -15.6510620 -189.18746948 189.833755
#> 25    -72.2022781 -127.20296478 146.266070
#> 26     31.2211132  -11.67975807  33.334287
#> 27    265.7444763   49.24003601 270.267845
#> 28    221.2923279   12.76550102 221.660218
#> 29     -5.6680579   32.79402924  33.280253
#> 30   -151.5105286  113.18076324 189.117227
#> 31   -137.0838318   79.85826111 158.648413
#> 32   -151.2914429   74.79066467 168.768315
#> 33   -187.0836182  -16.23209381 187.786477
#> 34     17.0406609 -165.74623108 166.619918
#> 35    156.5747375 -179.70677185 238.348846
#> 36    140.4532623    7.01097870 140.628136
#> 37    -27.1780853  169.37025452 171.536968
#> 38    -40.7624283  101.68067169 109.546951
#> 39     88.5297318  -92.22016907 127.836118
#> 40     48.7814827  -64.69036102  81.021453
#> 41     -8.8640146   30.37243271  31.639460
#> 42    -62.0974045   64.47207642  89.513889
#> 43     64.0016937   68.18375397  93.515994
#> 44     82.8871384    0.77158552  82.890730
#> 45    -41.9130974  -39.16460800  57.363527
#> 46   -121.6546097  -53.16825485 132.765611
#> 47    -94.7093887  -63.87655258 114.236957
#> 48    -26.8081303   44.97620392  52.359667
#> 49    -42.4895668   46.66805267  63.113156
#> 50    -80.3070755   44.10061264  91.619269
#> 51     62.6301842   61.31841278  87.649802
#> 52    181.1142578  -36.29592896 184.715373
#> 53    159.2825165 -115.93933105 197.009768
#> 54    -22.1916294 -119.32911682 121.375066
#> 55   -114.7984390  -49.22388840 124.906656
#> 56    -70.9710007   -8.94394493  71.532350
#> 57     33.2290421  -45.72446060  56.523407
#> 58     49.4227371   16.67633057  52.160396
#> 59      3.4006853  128.05355835 128.098706
#> 60   -100.3439713  -53.24666214 113.596301
#> 61   -116.6013184 -158.78785706 197.001145
#> 62     -2.6708071   -6.49963808   7.026984
#> 63    128.8948975  203.88111877 241.208220
#> 64    163.2560883  179.27153015 242.468208
#> 65     84.1502151  -23.41862679  87.348101
#> 66     43.0553589  -21.36544800  48.065022
#> 67    -54.7816086    0.15178701  54.781819
#> 68   -112.0419769  -63.11496353 128.595891
#> 69   -126.7290421  -35.75094223 131.675282
#> 70    -43.8736267   90.23932648 100.339579
#> 71     39.8857574  200.98425293 204.903742
#> 72     52.7658920  136.94410706 146.758059
#> 73    -18.8162727  -24.19443703  30.650007
#> 74    -83.0406723  -94.57434845 125.857303
#> 75    -57.0097542 -138.57025146 149.839336
#> 76    101.7038040  -76.19267273 127.078665
#> 77    142.4222717   -3.39869571 142.462818
#> 78    109.0482635  -34.44440079 114.358824
#> 79     -1.2944989   45.45014572  45.468577
#> 80    -41.1319962   16.59608650  44.353931
#> 81      6.0705128  -90.73576355  90.938605
#> 82    -23.4316483 -118.72833252 121.018425
#> 83    -69.2164383   13.22089386  70.467775
#> 84   -183.9283752  126.79405212 223.397356
#> 85    -77.0902710   16.29553032  78.793745
#> 86     41.3868370 -146.63650513 152.365137
#> 87     55.2916069 -138.93598938 149.533845
#> 88     32.6053658  -20.12469673  38.315967
#> 89     41.1934395  -24.68583298  48.023846
#> 90     91.3922195   -4.52782631  91.504311
#> 91     96.9611664   -3.40769768  97.021030
#> 92     40.2726593   63.99882126  75.615714
#> 93     -7.0726838   83.68404388  83.982391
#> 94    -68.8004303   16.93660927  70.854414
#> 95    -97.0167999  -74.67823029 122.429970
#> 96    -47.0187912  -51.90851593  70.037567
#> 97    -12.9483662   66.30415344  67.556650
#> 98     35.6335182  129.56027222 134.371172
#> 99    -17.3966465  141.38569641 142.451951
#> 100    50.4268303  -46.24490356  68.421169
#> 101    82.6846771 -101.37976074 130.822826
#> 102    32.1437492  -74.53298950  81.168880
#> 103   -21.1888981   23.83857155  31.894308
#> 104   -45.8895302   91.95376587 102.768400
#> 105   -10.4918242  -23.32550240  25.576502
#> 106   -69.0387497  -13.85454655  70.415179
#> 107  -110.7522125   54.41586685 123.398295
#> 108   -69.6824417  115.49557495 134.888363
#> 109    67.0328445   62.59561539  91.714848
#> 110   120.9660950 -146.95712280 190.339676
#> 111   141.5367432 -147.61238098 204.504437
#> 112    43.0280571   16.60894012  46.122344
#> 113   -99.9705887  142.53858948 174.101603
#> 114   -99.2521362   99.05429840 140.223894
#> 115    28.0268269  -30.54415703  41.454174
#> 116   131.2266998  -49.68000793 140.315893
#> 117    89.0407562  -23.03337288  91.971694
#> 118   -20.8925362  -23.46214294  31.416082
#> 119   -87.1069794  -43.57197189  97.396830
#> 120   -39.7460747 -123.47339630 129.712875
#> 121   -65.9408340  -51.87054825  83.897243
#> 122   -69.4172592   73.83473206 101.342605
#> 123   -34.2696686  140.04449463 144.176526
#> 124   -62.1524734  102.28570557 119.688327
#> 125   -42.5893555   50.75358963  66.255415
#> 126   -57.4504051  -39.37132263  69.646609
#> 127    79.1130905 -122.52713776 145.848485
#> 128   206.9855499  -95.28136444 227.863021
#> 129    94.0290222   21.01173973  96.348068
#> 130   -49.7541237  100.40355682 112.055107
#> 131  -121.6758270   55.83302689 133.874321
#> 132   -97.2508850   28.28610992 101.280989
#> 133    42.0805321   -8.81852150  42.994622
#> 134   117.7131042  -22.32388115 119.811229
#> 135     9.1105795  -45.27533340  46.182881
#> 136   -14.0547094  -47.29813766  49.342159
#> 137    -2.1738048  -49.33021545  49.378088
#> 138    16.7503681   -5.41187811  17.602933
#> 139   -55.9012337   -9.83218575  56.759315
#> 140   -80.0961838   19.91369057  82.534561
#> 141   -48.6606064   58.13911057  75.815637
#> 142   -85.5807037  -10.57353210  86.231412
#> 143   -19.6909447  -12.33423710  23.235032
#> 144   159.3507843    2.22299767 159.366289
#> 145   255.2388153   -0.57013750 255.239452
#> 146    91.3393097   34.59766769  97.672248
#> 147  -111.7990189   -5.36026239 111.927445
#> 148  -141.2005463  -13.02712727 141.800213
#> 149     3.7758646 -104.41368866 104.481939
#> 150   124.2413788  -36.88462448 129.600909
#> 151    91.8211594  117.24235535 148.919089
#> 152   -70.7185822   -4.84669352  70.884471
#> 153   -82.6106873 -158.77243042 178.978240
#> 154   -53.1674156 -135.55189514 145.605942
#> 155   -36.6575775   92.76426697  99.744610
#> 156    20.2388325  194.10791016 195.160168
#> 157   -42.2785339   83.76420593  93.829189
#> 158   -62.7411652  -28.57018471  68.939896
#> 159   -68.1157379 -101.02299500 121.841698
#> 160    55.4656868  -88.36479187 104.330144
#> 161    92.8313904   21.38428688  95.262557
#> 162    57.2518196   36.54690170  67.922359
#> 163    27.8955269   87.82576752  92.149476
#> 164    50.1905785   17.40625191  53.123175
#> 165    51.1857452  -44.90263367  68.089845
#> 166   -40.3380852  -54.90361023  68.129051
#> 167  -112.4156723  -12.23504925 113.079529
#> 168  -122.9173355   42.81399536 130.160323
#> 169   -11.9670296  -42.72460938  44.368931
#> 170    56.5762177  -38.76630783  68.583489
#> 171    37.3342361  147.04826355 151.713668
#> 172   -32.6348228  150.90957642 154.397966
#> 173   -43.7315216  -83.47206879  94.233923
#> 174   -38.9181557 -147.06231689 152.124777
#> 175    59.4101944   18.70984840  62.286673
#> 176    95.7946930  108.88218689 145.023977
#> 177   111.0327454   16.05690765 112.187766
#> 178    66.4907608  -66.08896637  93.748455
#> 179    -5.3833032  -92.89695740  93.052806
#> 180   -81.4554367  -27.76130104  86.056249
#> 181  -143.2841644  -24.88187981 145.428538
#> 182  -164.6732330  130.85159302 210.331674
#> 183  -124.4492722   91.21720886 154.299062
#> 184   -36.9026604  -33.76340485  50.017735
#> 185   102.2727203  -17.10977936 103.694040
#> 186   186.8810883  -52.18580627 194.030666
#> 187   105.5477142   42.98226166 113.964007
#> 188     9.5038176    4.66323090  10.586230
#> 189   -74.9679337  -22.36788940  78.233711
#> 190   -86.0345764  -13.23497295  87.046613
#> 191    -8.9459982   33.19957733  34.383758
#> 192    14.1716557   58.80925369  60.492678
#> 193    32.0004730   20.03813744  37.756552
#> 194    38.0872498  -29.58199883  48.225857
#> 195    46.0789108  -56.21105957  72.683899
#> 196    46.1166992  -32.42123413  56.372745
#> 197   -21.9444962 -102.38980865 104.715012
#> 198    13.6995945  -49.47499466  51.336673
#> 199    17.8661613    7.11783838  19.231832
#> 200    -1.3027704   68.79072571  68.803061
#> 201    83.9183884   46.59579468  95.986791
#> 202    21.3223743  -81.00074768  83.760162
#> 203   -14.2809944 -103.11479950 104.099033
#> 204   -53.0888786   11.96696568  54.420927
#> 205   -54.8915558  116.71440887 128.978045
#> 206    71.4269333   37.38277435  80.618104
#> 207    90.7818832    5.78904533  90.966276
#> 208    42.2976379   57.64610672  71.499397
#> 209   -84.8083725  117.38185883 144.813538
#> 210  -118.6647034   87.14585114 147.226734
#> 211   -18.0314598   34.22853851  38.687548
#> 212   -28.7923164   28.36307907  40.416107
#> 213   -66.2289581 -130.55703735 146.394723
#> 214   -40.6946449 -190.04518127 194.353351
#> 215   -19.6708374 -148.45184326 149.749429
#> 216   -18.1602211    9.38801956  20.443301
#> 217   -83.2091980   59.61377716 102.360017
#> 218   -21.9610424  117.37867737 119.415415
#> 219    86.5766373  142.57514954 166.802840
#> 220   147.8436584   65.13146210 161.554495
#> 221    33.9878693   18.80188942  38.841811
#> 222   -96.9891434 -113.23935699 149.097438
#> 223   -13.6476269 -128.45288086 129.175850
#> 224    94.9342728 -210.20164490 230.645285
#> 225   163.8195038  -80.85556030 182.686758
#> 226    16.7928543  228.14105225 228.758256
#> 227   -56.7815895  245.11437988 251.605263
#> 228   -24.3652306   79.95312500  83.583292
#> 229    96.1388931 -138.04924011 168.226869
#> 230   140.3195953  -73.23486328 158.281187
#> 231   -35.5582314   56.29784393  66.587049
#> 232  -130.9377594   54.10880661 141.677309
#> 233   -50.8165512   59.40374374  78.173695
#> 234    71.0454407  -54.54254532  89.567538
#> 235     9.4007597  -95.94954681  96.408972
#> 236  -132.5621643  -79.37492371 154.509242
#> 237  -148.4096985  -48.70329285 156.196829
#> 238   -23.3553200   51.26379013  56.333357
#> 239    59.0662994   87.67605591 105.716217
#> 240    67.7534714   58.45033646  89.481700
#> 241     5.9081607  -81.86647034  82.079384
#> 242    -5.4913239  -96.60652161  96.762465
#> 243   -15.2783537   45.48671341  47.984051
#> 244    41.0019417   93.26740265 101.882126
#> 245    58.1964302  -76.71002197  96.287341
#> 246    -2.4931648 -158.67016602 158.689752
#> 247   -44.8687134  -32.06776047  55.150183
#> 248     2.1397357   56.37292862  56.413523
#> 249   -18.7389259   64.41403961  67.084393
#> 250    26.9712315  102.31047821 105.805866
#> 251   141.2466431  136.94068909 196.731712
#> 252    97.9345856   79.38700104 126.069342
#> 253   -20.9999752  -20.54061699  29.375430
#> 254  -166.7921295  -84.06656647 186.780090
#> 255   -68.4442444 -100.19620514 121.342054
#> 256    -3.1101296 -123.01626587 123.055575
#> 257  -136.7077179   45.43517303 144.060248
#> 258   -34.6769714   77.65416718  85.045059
#> 259   112.3130035   85.74031067 141.299723
#> 260   224.8787537   -4.17773104 224.917557
#> 261    88.6502151 -111.17916870 142.195880
#> 262  -126.7739029 -155.04444885 200.275819
#> 263  -123.9640579  -56.24002075 136.125044
#> 264   -50.4061279  170.37187195 177.672036
#> 265    45.9022865  148.46655273 155.400570
#> 266    14.4426622   72.97444153  74.389916
#> 267   -39.5041389   12.57084942  41.456040
#> 268    33.5073433   35.37828064  48.727454
#> 269    59.6390152  -18.54594994  62.456100
#> 270    22.0536766 -148.80691528 150.432253
#> 271    19.9764786 -123.53957367 125.144261
#> 272    56.6563950  -62.98833084  84.719991
#> 273    84.1867905   62.99137115 105.144322
#> 274   -12.1377859  117.97622681 118.598971
#> 275   -38.1852913   52.16507721  64.647597
#> 276   -43.6848335   10.81493378  45.003638
#> 277   -81.9223328  -77.29937744 112.634197
#> 278   -93.0114365  -26.13617134  96.613802
#> 279   -64.7565994    7.76248217  65.220191
#> 280    48.4983711   56.86454391  74.737329
#> 281    56.5669670   70.50070190  90.388997
#> 282    49.4861641   -5.75842094  49.820075
#> 283    -4.9494057  -44.45799637  44.732651
#> 284   -39.1371536 -119.52637482 125.770708
#> 285   -55.5324249  -28.50234032  62.419818
#> 286     2.0592864   33.90407562  33.966557
#> 287    54.1134109   19.72943306  57.597845
#> 288    12.3774672  -51.72109985  53.181518
#> 289   -32.9690704 -160.86483765 164.208573
#> 290  -128.4083862  -95.12213898 159.802800
#> 291   -22.7235298   16.28308678  27.955281
#> 292    68.1771622   79.37742615 104.636997
#> 293   153.1009979   89.08130646 177.131010
#> 294    45.6674309   34.86347580  57.454122
#> 295  -125.2268143  104.69649506 163.227176
#> 296   -27.9611073   68.30666351  73.808020
#> 297   106.1894989  -14.03986073 107.113619
#> 298   105.9476624  -65.41093445 124.513041
#> 299   -90.9123611  -76.48400116 118.805976
#> 300  -225.7930450    0.32669726 225.793281
#> 301  -225.7936554   25.43455887 227.221679
#> 302   -68.0328979   10.78770065  68.882869
#> 303   114.7939224  -49.89190674 125.167276
#> 304   114.1936264  -55.64629364 127.030289
#> 305    57.1300735  -24.34141350  62.099515
#> 306    38.6665382  144.20930481 149.303130
#> 307    72.6796265  213.09367371 225.147156
#> 308    79.3488617  177.41348267 194.349648
#> 309    77.5934906   34.77040100  85.027822
#> 310    49.8036995 -114.00360870 124.407521
#> 311   -19.2590084 -156.51304626 157.693510
#> 312  -135.8672943 -189.30311584 233.014144
#> 313  -148.8911438 -137.12570190 202.415491
#> 314   -43.6580467   -0.71714067  43.663936
#> 315   101.9239197   89.25527954 135.480590
#> 316   235.2789154   48.17904663 240.161172
#> 317   124.6417465    4.07052517 124.708196
#> 318   -37.5765495   30.05844498  48.119717
#> 319   -94.9339371   52.63769150 108.550352
#> 320   -27.0823307  -23.79408264  36.050118
#> 321    30.3802032  -69.45016479  75.804236
#> 322   -21.1398373    2.77409410  21.321077
#> 323  -112.6311722   87.74143219 142.773737
#> 324  -132.8941345   13.18703365 133.546804
#> 325     9.8286209 -119.49548340 119.899009
#> 326    80.8523178  -33.02212906  87.335894
#> 327   153.0110168   30.38905525 155.999570
#> 328   113.7554321   -8.53473854 114.075151
#> 329    86.0363617 -118.64096832 146.553522
#> 330   -69.8014221  -42.37222290  81.655642
#> 331  -179.4603424  135.38674927 224.801215
#> 332  -140.3579254  107.02261353 176.505487
#> 333   -60.5411987   54.85614014  81.697202
#> 334   107.5112228  -61.23849869 123.728803
#> 335   100.7055893  -94.23670959 137.920895
#> 336    31.4961166  -21.16209030  37.945216
#> 337  -100.3967056  133.80575562 167.282631
#> 338  -125.1553574  197.96586609 234.210050
#> 339   -61.0122643   58.20656204  84.323782
#> 340    23.7614403 -138.04557800 140.075650
#> 341   -46.2971573 -194.99154663 200.412400
#> 342   -47.0850296  -49.29986572  68.172405
#> 343    30.1624260   84.13680267  89.379939
#> 344   159.2349243   91.58595276 183.694714
#> 345   116.5656509   37.50329590 122.450187
#> 346   -18.5230465  -21.25127602  28.190778
#> 347    22.3406525    6.88158512  23.376505
#> 348    89.1506271    4.47213984  89.262727
#> 349    11.5853319  -52.71152496  53.969665
#> 350  -139.7593994  -23.38891792 141.702968
#> 351  -119.0064240   66.38847351 136.271634
#> 352   -38.4153175   96.44178772 103.811151
#> 353   -37.6732254   38.04404449  53.540837
#> 354   -60.8524094 -110.70909882 126.330995
#> 355    10.7191858 -104.82679749 105.373424
#> 356   133.9055939  -80.71343231 156.350140
#> 357   171.4027557  -96.16698456 196.537512
#> 358     3.5379148   33.00851440  33.197573
#> 359   -65.3148956  111.41201019 129.145931
#> 360   -89.9586334  179.96701050 201.198113
#> 361   -16.7003250   32.00393295  36.099205
#> 362   -48.0035324   12.92300606  49.712606
#> 363  -118.9486160   36.63335800 124.461947
#> 364   -57.3005371  -39.00902176  69.318506
#> 365   101.6471176  -78.04384613 128.152169
#> 366   186.8071594 -167.73942566 251.064593
#> 367   114.2499466  -11.41457844 114.818739
#> 368   -17.6588554   94.73165131  96.363483
#> 369   -50.9501457  189.07359314 195.818132
#> 370   -49.0838280  110.91436768 121.289815
#> 371  -110.5472031  -81.09359741 137.101625
#> 372  -111.6426086 -131.72462463 172.671505
#> 373   -61.0903778 -145.82363892 158.103029
#> 374    62.1180344    2.39283299  62.164104
#> 375   104.9382477    6.56835032 105.143612
#> 376   137.3597412   34.82942200 141.706694
#> 377   145.9811249   45.28936768 152.845071
#> 378    78.1532211   62.73230362 100.216106
#> 379   -68.4336472   50.32675934  84.946729
#> 380  -120.2359009   14.96143723 121.163181
#> 381   -84.1738739    5.53101921  84.355398
#> 382   -49.6635780  -10.57871437  50.777753
#> 383   -53.8854752  -30.53139877  61.933922
#> 384    29.0648918  -98.36572266 102.569895
#> 385   134.5310059  -43.27827072 141.320912
#> 386    95.6856003  -69.92468262 118.512427
#> 387   -50.7136688   -7.60549545  51.280793
#> 388  -209.6713409   42.73087311 213.981305
#> 389  -149.4402313    9.11961651 149.718236
#> 390    25.4984264  -31.05534744  40.182140
#> 391   165.3285675 -111.26795959 199.283953
#> 392   221.0568542  -45.30927658 225.652528
#> 393    84.3420715    8.96253395  84.816933
#> 394   -36.7235985   45.06487274  58.133170
#> 395  -149.6265259   88.19984436 173.687391
#> 396  -170.9530182  107.13507843 201.749497
#> 397    -6.9377403   12.37346363  14.185727
#> 398    61.5700684 -104.07921600 120.927071
#> 399    29.2146702 -176.59748840 178.997681
#> 400   -56.7082558 -177.27491760 186.124213
#> 401   -33.0058670   25.81067657  41.899622
#> 402     7.8842397  110.31368256 110.595071
#> 403     6.0709472  131.03355408 131.174116
#> 404   -12.4504805   33.88129425  36.096490
#> 405   -94.9812622   31.65414429 100.117057
#> 406   -48.0206070   51.74245834  70.592214
#> 407    44.8474197  -27.22745323  52.465467
#> 408   192.5456696  -14.51291370 193.091842
#> 409   144.2350006   36.92821884 148.887302
#> 410   -42.7991753   71.38059998  83.228357
#> 411   -84.9586639   35.97161102  92.260129
#> 412   -11.4332485   27.03132248  29.349814
#> 413    34.5739784   71.24048615  79.186911
#> 414   -27.8708477    2.61035752  27.992823
#> 415   -35.8743706  -47.68965149  59.676405
#> 416    -6.8925009  -98.43029785  98.671324
#> 417   -36.2662773  -81.18564606  88.917670
#> 418   -92.8056107   -0.39394140  92.806447
#> 419   -52.2323265   27.24114037  58.909215
#> 420    17.9682007   29.19582558  34.281955
#> 421    36.9471016  -25.13275528  44.684938
#> 422    -2.6366920  -11.14871693  11.456266
#> 423    56.5762253   -3.21412635  56.667450
#> 424    56.7289124   26.80868912  62.744524
#> 425    71.7051544  -41.89389801  83.046540
#> 426    54.9393845  -65.37674713  85.395872
#> 427    31.8091602   15.94537640  35.581986
#> 428    16.1197205   73.17456818  74.929052
#> 429   -64.3943939  137.73326111 152.043050
#> 430   -46.4279251   34.35950851  57.759225
#> 431   -27.0355015  -63.89004517  69.374752
#> 432    21.1195889 -165.06848145 166.414064
#> 433   -11.6708202 -137.98939514 138.482061
#> 434   -10.9844284   32.08378983  33.912052
#> 435    47.7716331   99.38629150 110.271319
#> 436    46.1059265  108.92401886 118.280169
#> 437   -18.9777145   39.39282990  43.725836
#> 438   -56.6279716   65.92487335  86.906939
#> 439   -57.7246704   25.41861343  63.073318
#> 440    42.6936111  -14.20941353  44.996132
#> 441    88.9130859   -7.58073473  89.235668
#> 442    -7.8545365 -115.41759491 115.684549
#> 443   -49.6012001 -142.42697144 150.816847
#> 444   -60.1692543  -43.46732712  74.227675
#> 445    -2.0122550  105.92121887 105.940331
#> 446    49.5195923   81.65150452  95.494284
#> 447    87.7206497  -68.46677399 111.277183
#> 448   156.4388733 -117.10975647 195.417031
#> 449    64.0165634  -10.09620762  64.807822
#> 450   -58.2434883   43.89039230  72.929215
#> 451   -69.6927261   33.09745026  77.152559
#> 452  -155.2043304    1.19356704 155.208920
#> 453  -176.0703278  -24.22784424 177.729426
#> 454   -30.3613148   50.08656311  58.570242
#> 455   142.2644196   64.43812561 156.177582
#> 456   177.1892090   34.32496262 180.483292
#> 457    -6.2568121  -70.14714050  70.425628
#> 458   -29.8326340  -79.19102478  84.623900
#> 459    59.5998001   61.69883728  85.783930
#> 460    74.8785019  163.66917419 179.984412
#> 461    35.3056374   40.47869492  53.712315
#> 462    -3.0417588  -62.97676849  63.050184
#> 463   -27.8608036  -66.73352051  72.315884
#> 464   -46.4419403   40.81893539  61.830731
#> 465   -66.6277618  -46.17556381  81.064427
#> 466    -1.3270563 -128.34994507 128.356805
#> 467   -26.6817932  -23.67130470  35.668596
#> 468  -145.1918182   92.72094727 172.272569
#> 469  -136.7307587  217.31539917 256.751403
#> 470   -37.7828712   81.10695648  89.475604
#> 471   155.5397644  -15.52028942 156.312180
#> 472   124.4590149  -61.31369400 138.742263
#> 473    47.3268318 -122.14013672 130.988709
#> 474   -12.5603676 -123.22706604 123.865543
#> 475   -58.4159050  -80.71937561  99.639528
#> 476   -14.4633217  -22.53883362  26.780342
#> 477    60.4352112   15.39232063  62.364560
#> 478    39.6909637   -0.20260152  39.691481
#> 479   -69.4351120  104.88651276 125.787183
#> 480  -158.4856110  182.44665527 241.670170
#> 481   -10.4412708   -1.93092012  10.618314
#> 482   147.2464447 -182.46691895 234.468957
#> 483    97.6405411 -185.63845825 209.750596
#> 484    -4.5013924    2.25362134   5.034019
#> 485   -72.0336838  160.60171509 176.016370
#> 486   -30.5319977  178.75192261 181.340709
#> 487    27.6846313   13.37137699  30.744634
#> 488    61.6097679 -121.53536224 136.259340
#> 489    19.1033115  -91.62683105  93.597076
#> 490  -116.2347717   53.50143814 127.956735
#> 491  -179.0079041   91.16770935 200.886488
#> 492   -69.8987198  -57.40713120  90.451146
#> 493    93.6198502  -15.20528984  94.846598
#> 494    74.6580505   37.04909515  83.345426
#> 495    16.7354736   46.29816055  49.230029
#> 496    -7.1853790  -27.03193474  27.970613
#> 497    13.0348206  -63.15201569  64.483204
#> 498    37.5728340   29.76360130  47.933181
#> 499    13.7159300   10.30802441  17.157567
#> 500    35.6383781    9.56806278  36.900431
#> 501     5.4869413   -7.09735394   8.971007
#> 502   -24.7587204   53.96280670  59.371531
#> 503   -46.3495941  132.35081482 140.232033
#> 504   -27.4754162   80.70433044  85.253079
#> 505   -37.6956329   35.24022675  51.602658
#> 506    -1.6862965  -30.77552986  30.821694
#> 507    36.1173019  -35.51512527  50.653565
#> 508    30.2395744   -8.87391376  31.514730
#> 509    -3.5222266  -34.42670059  34.606413
#> 510    -7.7864885  -50.53136063  51.127760
#> 511    15.2044544 -123.15292358 124.087945
#> 512   -53.9195747  -53.17428589  75.728629
#> 513   -74.1284409   72.67083740 103.807882
#> 514   -74.9365692   82.93401337 111.774505
#> 515    -3.5397546    8.89166355   9.570347
#> 516   -22.0965748  -76.99301910  80.101084
#> 517   -59.7010880  -20.25344467  63.043016
#> 518    30.0396385  -46.01666641  54.953739
#> 519    84.3136368  -85.82469940 120.310716
#> 520    70.0436249  -99.35766602 121.565025
#> 521    33.9923782  -23.72604752  41.453674
#> 522    63.8838463  -11.78269482  64.961356
#> 523   119.8960114  -79.46016693 143.836614
#> 524    18.1740093  -27.54918671  33.003823
#> 525   -91.1484222   62.20498657 110.351689
#> 526   -14.8548098  192.13780212 192.711184
#> 527    12.6472893  142.58732605 143.147125
#> 528   -71.6823883   31.80784225  78.422596
#> 529  -135.9187164   -1.23919380 135.924365
#> 530   -39.2975082   21.32375717  44.710142
#> 531   125.2388840   39.89365768 131.439271
#> 532    50.4841614   -6.58682585  50.912050
#> 533   -18.8516312   16.62634850  25.136019
#> 534   -31.3603153   73.74674225  80.137703
#> 535     7.5080624   59.44783783  59.920084
#> 536    36.3577271  -77.54874420  85.648655
#> 537   -31.6356640 -189.62280273 192.243654
#> 538     1.9634423 -168.73693848 168.748362
#> 539   109.4104919    4.17459965 109.490105
#> 540   136.2031555   49.50543213 144.920969
#> 541    28.8234806  -58.93613434  65.606867
#> 542   -97.4447021   36.94104004 104.211854
#> 543  -156.1456299   68.50159454 170.510780
#> 544   -16.8315067   26.81923676  31.663403
#> 545    -2.2511320  -57.73617935  57.780048
#> 546    31.2469845   42.27008820  52.565525
#> 547   -28.5091171  205.75358582 207.719301
#> 548  -123.1428680   52.02584839 133.681917
#> 549    15.7306128  -75.01670837  76.648279
#> 550   184.8143768 -136.21923828 229.591016
#> 551   182.2845001 -156.05897522 239.962586
#> 552     9.3592215  -71.42420959  72.034802
#> 553   -40.3714523   58.08672714  70.738406
#> 554    -6.0150137   52.22120285  52.566476
#> 555  -101.0722809   18.42146111 102.737316
#> 556   -99.8079147   52.90150070 112.961005
#> 557   -17.5215015   45.53818130  48.792714
#> 558     9.0755358    7.80321169  11.968937
#> 559    18.0194855  -52.92538071  55.908835
#> 560   122.8254852  -30.02327728 126.441674
#> 561   104.1867599  -51.23449326 116.102775
#> 562  -138.1550446  -26.24502945 140.625808
#> 563  -200.1587982  112.59050751 229.652274
#> 564    -1.9338949  120.08949280 120.105063
#> 565   173.7665863   69.39836121 187.112156
#> 566    64.7601013  -36.94757843  74.558663
#> 567    -0.8270794   51.70781708  51.714431
#> 568    27.6821270  -30.84523010  41.445487
#> 569    80.9578857 -177.69155884 195.265126
#> 570   -58.1492920  -63.65600586  86.217326
#> 571  -191.2136078   85.87374115 209.611410
#> 572   -88.9607162  148.75463867 173.326142
#> 573    96.5626907  -39.70607376 104.407498
#> 574   222.8290863 -136.36541748 261.243811
#> 575    75.5989532  -55.00630951  93.492758
#> 576    -7.9803700  -31.07775688  32.086029
#> 577   -43.0117378  -22.37997437  48.485800
#> 578    -8.7867832   31.39543915  32.601858
#> 579    65.3967209   36.71134186  74.996358
#> 580     5.2977371   -1.70826626   5.566345
#> 581   -54.6311035  -51.95933151  75.394493
#> 582   -75.3463821  -79.06707764 109.218497
#> 583   -42.0752602 -105.46029663 113.543832
#> 584   -30.0780888   -0.30986598  30.079685
#> 585     5.0201411  109.44892120 109.563991
#> 586    48.3419456  116.17124939 125.828069
#> 587    -2.0780525   78.12123108  78.148865
#> 588  -125.3302307   46.18814468 133.570249
#> 589   -49.1492348  105.03423309 115.964811
#> 590    63.5846214   -5.42267179  63.815433
#> 591   145.2703552 -106.36695099 180.048339
#> 592     7.6480598  -47.06723022  47.684557
#> 593   -87.4070206   53.71175766 102.591131
#> 594  -100.9039078  122.79978180 158.938306
#> 595   -88.2923279   49.46884155 101.206232
#> 596    14.0797634  -99.11278534 100.107862
#> 597   -57.9644470 -190.69317627 199.308215
#> 598   -32.6528435  -43.64490891  54.507672
#> 599   -42.2003822   98.84909058 107.480300
#> 600    97.5170441    2.91791654  97.560689
#> 601   131.5012512 -113.26596832 173.556212
#> 602    69.2873535   -5.96918344  69.544004
#> 603    -8.4007692  151.29113770 151.524194
#> 604    27.5223675  113.12616730 116.425987
#> 605   130.8254547  -34.34069061 135.257468
#> 606    14.6574154  -62.26845169  63.970305
#> 607   -82.9580002    9.81087399  83.536118
#> 608  -140.5053558  -16.05609894 141.419777
#> 609   -20.2583256  -97.19918060  99.287867
#> 610    40.4895325 -107.90291595 115.249475
#> 611    15.6169319  -38.16882706  41.240125
#> 612   -54.1206131   49.84659576  73.578012
#> 613  -126.5006714   36.11052322 131.553752
#> 614    22.2479649   76.53102112  79.699242
#> 615   104.7810211  107.56975555 150.167622
#> 616    19.1071930   59.38010025  62.378531
#> 617  -134.3301392  -16.52739143 135.343049
#> 618  -191.9993591  -66.28918457 203.120678
#> 619    50.0943642  -80.61018372  94.907571
#> 620   229.7772369   20.81455421 230.718062
#> 621   188.5512695   71.17938995 201.539293
#> 622    11.2592573   72.39346313  73.263800
#> 623  -161.8825226   13.76240063 162.466473
#> 624   -93.1781311  -49.95124054 105.722706
#> 625    37.0654106   35.01506042  50.989206
#> 626   140.6601715 -102.22858429 173.884925
#> 627   195.2699738  -70.09934998 207.471158
#> 628    34.8568306    8.14429951  35.795646
#> 629  -183.9811554   -1.05141044 183.984160
#> 630  -182.6408691   68.27725220 194.985821
#> 631   -31.8032627   46.25936127  56.137118
#> 632    96.4436951   76.65398407 123.195859
#> 633    89.0545425  -89.78849792 126.462192
#> 634    -5.7383189 -167.02134705 167.119893
#> 635   -57.5423622  -70.69983673  91.156955
#> 636   -55.7481728   48.28030396  73.748536
#> 637    15.6619272   88.00556946  89.388345
#> 638    12.5065355   21.25176048  24.658685
#> 639   -54.2361450   14.60653782  56.168589
#> 640   -71.7220078   -1.26922834  71.733237
#> 641    45.4569168   21.97994041  50.492069
#> 642    38.6457405   14.69994831  41.347089
#> 643   -48.1601295    5.62311316  48.487292
#> 644     1.6138332   89.96566010  89.980134
#> 645   109.4103470   93.51651001 143.930406
#> 646   134.6110382   -2.16930079 134.628517
#> 647     0.8480313 -117.94136810 117.944417
#> 648   -76.0415649  -69.91884613 103.300361
#> 649   -61.2592735  114.40133667 129.770430
#> 650   -28.3255978  100.27233887 104.196360
#> 651   -29.0958424  -30.45660019  42.120928
#> 652   -84.2030945 -158.91867065 179.848005
#> 653   -75.3944244 -112.62754822 135.533331
#> 654    31.3891163   18.34716415  36.357875
#> 655    68.8789215   78.30856323 104.290637
#> 656    88.8375397   33.25576782  94.858076
#> 657    84.3188248  -73.97174072 112.167208
#> 658   108.1730423  -81.24784851 135.287176
#> 659    14.2068090   30.37677956  33.534790
#> 660   -85.4733200  104.84680176 135.272097
#> 661   -83.7314911   55.93468857 100.695839
#> 662  -132.2292938  -88.50804901 159.117129
#> 663  -174.2733154 -102.97983551 202.425381
#> 664  -138.7221375  -34.35727692 142.913449
#> 665    95.5983810    9.31388855  96.051023
#> 666   237.7726746   27.03472710 239.304662
#> 667   183.2294006  -17.88891029 184.100588
#> 668    68.1651917   18.00237846  70.502333
#> 669   -23.9374256   38.87934875  45.657465
#> 670   -35.9954376   11.39296722  37.755413
#> 671  -116.9543076    5.56992197 117.086866
#> 672  -104.5876541  -46.31287766 114.382953
#> 673    -1.8271295  -13.21302223  13.338754
#> 674    51.3592796    0.61474198  51.362959
#> 675    65.2660828   77.59511566 101.393607
#> 676    49.2013893  112.89172363 123.147546
#> 677    20.0566769   80.92660522  83.374971
#> 678   -43.6643867  -25.04704666  50.338188
#> 679  -137.7418365  -98.69691467 169.451747
#> 680  -103.4494095   85.34954834 134.113108
#> 681    -7.0282640  119.71266937 119.918805
#> 682    63.1692200   27.58241844  68.928515
#> 683    59.4594612  -82.95883942 102.066628
#> 684   -49.1958389  -86.04809570  99.118643
#> 685    -4.8708482   39.44057083  39.740204
#> 686    52.2637024  -25.79071426  58.280833
#> 687   169.0625153 -100.51262665 196.684830
#> 688   132.1769562  -70.58987427 149.845514
#> 689   -39.8517075   78.43416595  87.977707
#> 690  -117.3011627  142.32049561 184.430709
#> 691   -70.4129257  -37.23768234  79.653155
#> 692     5.0576138 -130.50541687 130.603382
#> 693    62.7414665 -103.39685059 120.943790
#> 694    70.8412933   -5.88927794  71.085670
#> 695   177.0006866   55.45821381 185.485462
#> 696   111.9951706   66.29211426 130.144392
#> 697  -126.8248825   35.33552933 131.655423
#> 698  -214.7178192  -69.86581421 225.798525
#> 699  -183.3981628   -0.77238756 183.399789
#> 700    23.7276459   76.34977722  79.951796
#> 701    93.0550003   80.88027954 123.291738
#> 702   127.6712036   44.01086044 135.044038
#> 703    96.7854843  -51.80478668 109.777802
#> 704    38.5571136  -57.65282059  69.357759
#> 705   -57.0709419  -54.40851593  78.850358
#> 706  -112.0690002   51.22482300 123.221116
#> 707   -98.7415390   46.48808289 109.137681
#> 708    20.5829544  -72.76154327  75.616798
#> 709   109.4668732  -80.23478699 135.722575
#> 710    83.3930969  -28.12186813  88.007091
#> 711    41.6402054  -76.65395355  87.233797
#> 712   -77.4824295  -77.27952576 109.433322
#> 713  -170.3559723   75.01540375 186.140990
#> 714  -186.0869751  214.92301941 284.289055
#> 715   -72.6301041  131.42756653 150.161038
#> 716   105.8966293  -88.64528656 138.101712
#> 717   134.6324310  -86.95889282 160.273955
#> 718    44.2380486  -19.12644196  48.195702
#> 719    51.4249611  -30.87010002  59.979077
#> 720    11.7458181  -64.97853851  66.031619
#> 721    26.2503071  -31.27182007  40.828977
#> 722    52.9106712   69.32434082  87.208964
#> 723    15.9843712  172.44462585 173.183859
#> 724     1.6271659  173.26663208 173.274272
#> 725   -82.8598557   18.50565147  84.901206
#> 726   -77.0725861  -32.19221878  83.525580
#> 727   -30.4186687 -127.83921814 131.408375
#> 728   -23.0215454 -174.67655945 176.187094
#> 729   -16.6603642  -67.66413879  69.685030
#> 730   -15.7666225   30.22868347  34.093397
#> 731   152.4528961  184.13098145 239.052513
#> 732   173.1341858   79.09725952 190.346586
#> 733    19.9368629   67.93158722  70.796745
#> 734  -186.0033875  -17.79656029 186.852824
#> 735  -175.5900574 -107.18144226 205.717597
#> 736    83.4858780 -114.65119934 141.826617
#> 737   124.6319275  -96.55952454 157.660582
#> 738   111.7669220  -57.97257614 125.907365
#> 739   -41.2646866  -93.99811554 102.656807
#> 740  -100.3314819   24.20195961 103.209210
#> 741   -96.9480591  107.51243591 144.768263
#> 742   -19.8751202  130.78652954 132.288082
#> 743    78.9090500   -2.37575412  78.944806
#> 744   -27.3217449  -67.71886444  73.022752
#> 745   -61.3839836   -6.91287947  61.772011
#> 746   -59.5965080   29.36269569  66.437276
#> 747    45.2097588  -30.71806526  54.658227
#> 748   149.8528595  -97.11428070 178.569491
#> 749   112.0874863  -55.30201721 124.987670
#> 750    75.5009995   -7.23812675  75.847158
#> 751     5.1797924    9.83347893  11.114295
#> 752   -32.0203552   68.68015289  75.777744
#> 753   -10.7113819   83.78404236  84.465966
#> 754  -112.9874344   19.92040062 114.730043
#> 755  -123.2190933  -45.76586533 131.443750
#> 756   -10.4778271  -25.24215889  27.330412
#> 757    36.5383301   10.13768196  37.918626
#> 758    36.6772232  -22.40138054  42.977210
#> 759   -50.7141991  -40.39194489  64.833936
#> 760   -55.5917168   32.05152893  64.169615
#> 761   -34.3332405  111.97346497 117.118864
#> 762   -46.1355705    4.15422916  46.322225
#> 763   -13.1604919  -35.79003906  38.132997
#> 764    16.2084312   32.15732956  36.011208
#> 765   119.0135803   81.37802887 144.175642
#> 766   128.1477966    7.11335802 128.345073
#> 767    36.1984062 -127.20988464 132.259893
#> 768   -89.3464050   30.12089920  94.287055
#> 769  -103.5685120   98.09017944 142.646837
#> 770   -49.8138924   79.13164520  93.505300
#> 771   -25.9502106  -91.63839722  95.241846
#> 772    18.0621185 -160.21633911 161.231248
#> 773    23.1025772  -70.90977478  74.578316
#> 774    14.3914194  -17.55210876  22.697786
#> 775   -72.3206482   71.29862213 101.556731
#> 776    -3.7400007  -54.57344055  54.701444
#> 777    50.9085655  -61.72308350  80.008881
#> 778    31.4610500   29.09338951  42.851172
#> 779    13.2849798  135.32417297 135.974713
#> 780    62.4088058   90.71240997 110.107222
#> 781    86.7586517  -40.20917511  95.623435
#> 782   -52.0069313  -32.96717453  61.575608
#> 783  -112.8553009   67.98697662 131.751842
#> 784   -29.2935448   72.48961639  78.184757
#> 785    13.7422915    2.58801699  13.983862
#> 786     9.2948647 -122.79952240 123.150791
#> 787   -20.4895744  -59.44213104  62.874395
#> 788    25.5938339   26.37437439  36.751217
#> 789    58.9082031   45.86746216  74.659229
#> 790    -3.7834582   57.43982697  57.564297
#> 791    29.6942425   -0.99607199  29.710944
#> 792   -17.7926102   -1.49390256  17.855216
#> 793     6.7915139 -176.64994812 176.780454
#> 794   -61.0584221 -164.53779602 175.501616
#> 795   -72.1943741   -3.74798155  72.291597
#> 796   -12.9082460  108.16497803 108.932480
#> 797    40.9936104  192.95417786 197.260718
#> 798     7.8897495   57.76594925  58.302256
#> 799   -66.4138184 -139.17840576 154.212269
#> 800    70.0606995 -154.23263550 169.399550
#> 801    62.5212898  -32.47725296  70.453415
#> 802    25.7911510  154.54878235 156.686022
#> 803   -70.5499268   78.93276978 105.866304
#> 804   -68.5084381   -0.05196597  68.508458
#> 805    54.9011383   92.80934143 107.831854
#> 806    10.5692272   32.44820404  34.126156
#> 807    56.6812553  -84.85074615 102.041236
#> 808   -19.7271614 -184.33525085 185.387825
#> 809   -35.2816467  -55.47848892  65.746919
#> 810    88.6191940  121.11059570 150.070443
#> 811    16.2063923   44.49874496  47.358056
#> 812  -109.3505325  -31.98201752 113.931508
#> 813  -181.3531647  -33.86872101 184.488646
#> 814   -25.7871189   -8.16621971  27.049263
#> 815   114.7398453   66.94623566 132.842126
#> 816   109.0229263   46.97034836 118.710623
#> 817   128.2617798    5.02881718 128.360325
#> 818    25.7306328  -30.97612572  40.268919
#> 819   -68.5109787   54.67219925  87.651603
#> 820  -124.3453598  186.07548523 223.798692
#> 821     7.0576692   77.84130096  78.160596
#> 822   100.7067337  -24.03801918 103.535852
#> 823   -10.1668682  -52.92540741  53.893079
#> 824   -24.2491989  -79.49601746  83.112216
#> 825   -38.6844139 -152.94720459 157.763530
#> 826    67.2051392  -88.20837402 110.892957
#> 827    14.4319057   15.46834660  21.155369
#> 828   -29.4180737   -8.61698437  30.654127
#> 829    45.1290092    0.21561970  45.129524
#> 830    56.4098206   60.02664948  82.372729
#> 831    27.4626827   61.59896088  67.443539
#> 832   -43.4580879    2.94052291  43.557457
#> 833  -130.8421326   54.81522369 141.860398
#> 834  -121.7227478   52.30109406 132.483326
#> 835    59.5882492  -18.32297516  62.341726
#> 836   157.4215393  -59.29666519 168.219011
#> 837    29.5864315   26.37176132  39.633656
#> 838  -229.6710968   43.61085892 233.774934
#> 839  -184.2365417  -49.30099487 190.718881
#> 840    58.8981438  -17.17185402  61.350337
#> 841   115.3360977    4.86092758 115.438486
#> 842    97.7816849   28.61313248 101.882134
#> 843    25.0093842  -94.18183136  97.445814
#> 844    -8.3288527 -148.76522827 148.998198
#> 845    26.0718136  -59.56177902  65.018036
#> 846    56.4626350   48.14164734  74.200050
#> 847    87.8214111   96.43609619 130.432055
#> 848    77.2607040   36.19853973  85.320283
#> 849   -82.3505707   53.92259979  98.434055
#> 850  -109.4471130  -20.09889412 111.277294
#> 851   -75.0184326  -92.44841003 119.056599
#> 852   -17.7257462   -9.56267643  20.140677
#> 853    39.0245323  101.01403809 108.290120
#> 854    92.3139954   20.04048157  94.464251
#> 855   155.8670959 -202.89953613 255.856939
#> 856    42.4992599 -173.00849915 178.151980
#> 857  -109.0843735   38.66167450 115.732993
#> 858  -128.3490906  157.48907471 203.165690
#> 859   -25.0663300   95.90406799  99.125734
#> 860   -16.7322750  -20.03409576  26.102376
#> 861    27.0403862  -29.83960533  40.268903
#> 862    41.7719879  -49.91159439  65.085069
#> 863    51.1649590   -4.14737129  51.332774
#> 864    44.3078270  -60.05405807  74.630245
#> 865   -32.2662506  -26.37051964  41.671516
#> 866    -4.5315733  105.65487671 105.752012
#> 867   107.0349121  141.67626953 177.563053
#> 868    52.7880974  150.97978210 159.942108
#> 869   -83.5145798  -57.89480591 101.619356
#> 870  -162.2339478 -150.99514771 221.628943
#> 871   -84.5955811 -117.62879181 144.889423
#> 872    39.6866760   35.42530441  53.197598
#> 873   -26.4979973  153.21624756 155.490715
#> 874   -24.7459602   77.67240906  81.519112
#> 875    -2.5882359  -73.03876495  73.084610
#> 876    35.5626984 -145.41162109 149.697178
#> 877     3.4250724  -64.92474365  65.015025
#> 878   -84.1108780   82.08507538 117.527016
#> 879   -62.6771278   74.24530029  97.163712
#> 880    66.8728561  -53.33461761  85.536895
#> 881   120.8453827  -91.92028809 151.831966
#> 882     0.9305125   20.39293671  20.414155
#> 883   -34.8540573   81.75837708  88.877655
#> 884   -71.0721283    0.87834841  71.077556
#> 885   -28.5247631   26.99128532  39.270747
#> 886    64.4483871   36.96887589  74.298670
#> 887   128.6447296   15.07261753 129.524709
#> 888   155.2311401  -50.22849274 163.155166
#> 889    62.5088272   -9.96369457  63.297936
#> 890     6.4210076    7.93792820  10.209801
#> 891  -122.4959564  -52.38946152 133.228807
#> 892  -202.8885803  -75.70763397 216.553508
#> 893   -79.5799484   26.22938919  83.791104
#> 894    83.2540207  169.53517151 188.874049
#> 895   209.5854340  126.49894714 244.802038
#> 896   101.3797607   82.44014740 130.668412
#> 897   -20.2794876  116.65853119 118.408068
#> 898  -162.2394867  -43.48349380 167.965667
#> 899  -190.3264160 -229.73602295 298.333345
#> 900  -174.0025482 -255.91816711 309.468892
#> 901  -183.4140167   -1.56081223 183.420658
#> 902    -2.4684658  144.33927917 144.360385
#> 903   155.4639282  -54.04090881 164.588738
#> 904   234.2266846  -44.76667023 238.466338
#> 905    84.5915985   74.56131744 112.761379
#> 906   -19.7551575   87.54898834  89.750162
#> 907    -4.1776524  -13.93958759  14.552144
#> 908    49.0055809 -207.88095093 213.579111
#> 909    27.4243679  -58.83164978  64.909622
#> 910     4.8655181  105.25066376 105.363065
#> 911    -9.0758572  177.69111633 177.922747
#> 912    -3.0122936  136.91114807 136.944282
#> 913     3.0923550  -67.87599945  67.946405
#> 914   -30.7906513  -76.06844330  82.063830
#> 915   -47.1990662  -13.34436989  49.049200
#> 916   -55.3784332    1.17240632  55.390842
#> 917   -11.8270197   -3.52888656  12.342262
#> 918    13.6950855   -1.42761648  13.769294
#> 919    17.8647881   42.60732651  46.201027
#> 920   -10.7895975   88.10796356  88.766146
#> 921    51.5941391  -46.45420074  69.425845
#> 922    62.9200897  -63.05637360  89.078864
#> 923    27.2374268  -95.17897797  98.999572
#> 924   -53.2209969   11.81117916  54.515855
#> 925   -86.9746628  113.73764801 143.181160
#> 926   -40.7972946  110.25904846 117.564778
#> 927   -48.0674858   -6.99186802  48.573341
#> 928     7.2399201 -106.66944122 106.914855
#> 929    -3.6681592   -0.86765486   3.769379
#> 930    28.7754478  -30.66808701  42.054226
#> 931   -51.3247986 -134.27729797 143.751966
#> 932   -84.3175430 -151.96133423 173.786349
#> 933    46.8767586  -62.37742615  78.028032
#> 934   125.3532867  129.13464355 179.970005
#> 935    78.6933975  137.49111938 158.418619
#> 936   -27.5179005   61.81945801  67.667424
#> 937   -13.2460756   37.36479568  39.643240
#> 938    -8.0209322  -31.53678513  32.540808
#> 939   -10.7191019  119.89277649 120.370997
#> 940    49.0057449   98.60384369 110.110313
#> 941   103.9762497   -6.41156769 104.173743
#> 942    59.3661156 -181.92922974 191.370270
#> 943   -16.6194019 -293.68783569 294.157695
#> 944   -50.3360710 -145.29930115 153.771281
#> 945   -22.2826500    5.27036762  22.897451
#> 946   -58.5756760  205.36494446 213.555310
#> 947    20.5405521  241.39302063 242.265360
#> 948    50.7427940   65.28002930  82.682001
#> 949    34.7519722  -73.74512482  81.523267
#> 950    -4.4286189  -95.58882141  95.691355
#> 951   -12.0851011   53.44244385  54.791829
#> 952    -4.6490116   84.57082367  84.698510
#> 953  -106.7987289  -16.49017334 108.064306
#> 954  -125.3063660 -111.84302521 167.959958
#> 955    20.8454800 -136.31622314 137.900858
#> 956   158.0633240  -19.60743713 159.274813
#> 957    36.3882637   77.25717926  85.397760
#> 958  -191.9616241   61.19828415 201.480756
#> 959  -195.2181854   -9.81122780 195.464575
#> 960    32.8515358  -22.53733063  39.839110
#> 961   113.3392334   20.89177704 115.248636
#> 962    79.5936279   38.11619568  88.249589
#> 963   -30.0145874   21.09443092  36.685835
#> 964   -35.2751389   75.46855164  83.305688
#> 965    39.7306519   66.99158478  77.887079
#> 966    12.4576435   50.03414917  51.561701
#> 967    20.1899147  -65.29892731  68.348976
#> 968    19.5047035 -219.10140991 219.967864
#> 969    88.5096817 -193.04284668 212.366439
#> 970     4.1786437 -107.88674164 107.967634
#> 971   -21.7836952   88.63711548  91.274682
#> 972    26.0683823  134.07048035 136.581310
#> 973    87.7466431   97.93984985 131.497861
#> 974   103.5846558   37.52047729 110.170627
#> 975   -40.7527962  -95.86643982 104.168924
#> 976  -135.5520477  -14.78865814 136.356379
#> 977   -88.5272751   76.09767914 116.738748
#> 978    61.4047356  108.29721832 124.494293
#> 979    87.0911102    8.66520882  87.521125
#> 980   -22.3174458  -59.57806015  63.620859
#> 981  -103.0204086   48.38182449 113.815665
#> 982   -79.6916046    4.06521606  79.795224
#> 983   -81.7762756  -72.67581177 109.403532
#> 984    26.3790627 -110.76658630 113.864356
#> 985   138.0275421   51.70655823 147.394608
#> 986   144.7605286   90.80200195 170.881872
#> 987    52.8921623   27.22468758  59.487515
#> 988    -3.5382910  -88.60686493  88.677483
#> 989   -30.8501015 -138.92897034 142.312992
#> 990  -141.4641113    1.57533455 141.472882
#> 991   -97.8177185   41.35221481 106.199396
#> 992     5.9488869  109.06201172 109.224135
#> 993   101.5705185  -33.44538498 106.935326
#> 994    44.9607124  -50.56818771  67.665407
#> 995   -87.8767090  -27.11765671  91.965664
#> 996  -112.6401901   80.47837067 138.436197
#> 997   -98.9477310   84.07678986 129.844369
#> 998    10.9167747   39.27291489  40.761965
#> 999   143.1657562  -87.79339600 167.940805
#> 1000           NA            NA         NA