Differentiate x and y and compute change amplitude via an algorithm proposed by Engbert & Kliegl (2003)

Differentiate x and y and compute change amplitude via an algorithm proposed by Engbert & Kliegl (2003)

diff_ek(x, y, trial, sample_rate, options = NULL)

Arguments

x

vector with x coordinates in degrees of visual angle

y

vector with y coordinates in degrees of visual angle

trial

vector with trial index

sample_rate

sample rate in Hz

options

List with method specific options, see Details.

Value

data.frame with columns x, y, and amp

Details

Method options, please refer to Engbert & Kliegl (2003) for details on parameters and the rationale for default values.

ek_velocity_time_window Time window for velocity computation in milliseconds. Defaults to 20 ms.

See also

Examples

diff_ek(rnorm(1000), rnorm(1000), rep(1, 1000), 250, list("ek_velocity_time_window" = 20))
#>                 x             y        amp
#> 1              NA            NA         NA
#> 2    160.95440674   52.81591034 169.398469
#> 3    -20.38085556   20.72726631  29.068864
#> 4    -61.74895096   -7.68057442  62.224787
#> 5    -44.17317200   79.48471832  90.934535
#> 6    -55.94441605    9.66238689  56.772699
#> 7    -18.68130684  -20.27139664  27.566660
#> 8     22.99825859 -144.44297791 146.262414
#> 9     29.71320152 -104.55315399 108.693313
#> 10  -111.79867554   31.98283577 116.283471
#> 11   -69.80689240   99.94558716 121.910306
#> 12   157.62495422   46.02266693 164.206309
#> 13   237.69689941 -110.86546326 262.280321
#> 14    58.81400681  -63.55234909  86.590926
#> 15  -135.48524475   17.03535652 136.552023
#> 16   -78.50932312   17.08882713  80.347631
#> 17    53.59314346  -66.69136047  85.556780
#> 18    91.25271606  -54.79923248 106.442539
#> 19    67.25933075   -0.80753297  67.264178
#> 20   -43.72927094   33.83095932  55.288181
#> 21  -127.12856293   71.87538910 146.040210
#> 22  -115.27228546   73.91484833 136.934673
#> 23    38.90488815   51.28178024  64.369335
#> 24   140.13330078  -24.85528374 142.320508
#> 25   125.41052246  -82.06851196 149.876749
#> 26    64.06622314  -83.21435547 105.019569
#> 27   -20.68382645  -69.68530273  72.690179
#> 28   -90.20465851  -66.70950317 112.191970
#> 29   -88.91088867 -116.95762634 146.915733
#> 30    31.82866096    2.53979492  31.929833
#> 31    55.55297852  113.57179260 126.430556
#> 32    -9.48290157  150.02749634 150.326894
#> 33  -116.34452820  -36.21122360 121.849505
#> 34   -92.88623047  -87.04077148 127.294728
#> 35   -11.96474934   29.79020691  32.103141
#> 36    16.04588318  109.42449951 110.594717
#> 37    96.53105927   87.66173553 130.394882
#> 38   103.19171906  -49.47248077 114.438006
#> 39    70.17716980 -147.49371338 163.337781
#> 40    -3.02835727  -73.56295013  73.625258
#> 41   -41.97515488  126.03106689 132.837282
#> 42   -24.23225403  141.42680359 143.487780
#> 43   -54.14571381   12.23180580  55.510138
#> 44   -67.06543732  -91.74974823 113.647654
#> 45   -22.59484673  -17.87645912  28.811367
#> 46    34.77915955   39.97171783  52.984226
#> 47    74.69052124    8.20530796  75.139877
#> 48   -16.68011665  -59.75230026  62.036793
#> 49   -90.05335236  -74.79496002 117.063625
#> 50   -76.77766418  -92.89910126 120.519927
#> 51    15.18720627  -94.93357849  96.140707
#> 52    96.06744385   74.79770660 121.752416
#> 53    70.63272858   66.83541870  97.241738
#> 54    40.92386627   10.35416508  42.213405
#> 55   -53.56586075  -95.35777283 109.372786
#> 56   -69.50486755  -37.92905045  79.180424
#> 57  -129.03668213  121.96494293 177.555379
#> 58   -93.98260498  152.17755127 178.859546
#> 59    30.17512894  131.83239746 135.241707
#> 60   120.02846527   -3.46675181 120.078519
#> 61    40.21212006 -101.10681915 108.809942
#> 62   -85.41580963 -144.80519104 168.120207
#> 63   -16.28437233   -3.49666524  16.655553
#> 64   136.07960510   34.42463303 140.366357
#> 65    11.55528927   42.91896820  44.447301
#> 66  -102.35359955  -13.41409111 103.228858
#> 67   -54.75131607   24.34463310  59.919678
#> 68    84.53427887   53.53371811 100.059499
#> 69   117.29569244  -53.80957413 129.049408
#> 70   -56.94894409  -28.26198959  63.576114
#> 71  -116.79505157   16.67353439 117.979197
#> 72   -47.92838669  120.44279480 129.628689
#> 73   126.79781342   26.94870949 129.629929
#> 74    99.09508514  -65.58447266 118.832483
#> 75  -116.80191803    2.52478695 116.829203
#> 76  -137.40708923   42.34251404 143.783158
#> 77    -8.42153168   71.97651672  72.467518
#> 78    84.93224335  -58.70531082 103.246305
#> 79    -8.40385532  -57.04838562  57.664054
#> 80   -25.86625671  -82.76878357  86.716404
#> 81    71.10575104 -128.54100037 146.897300
#> 82    79.10823059  -73.76971436 108.166921
#> 83    37.48500061   -2.38917971  37.561063
#> 84   -73.66664124   83.49587250 111.347810
#> 85   -71.05363464   66.39163208  97.244372
#> 86   -18.89093018   47.27371979  50.908465
#> 87   -48.88053131   35.98851395  60.699913
#> 88   -44.98052216   19.04681778  48.846992
#> 89    51.82801056  -29.81250000  59.790700
#> 90   148.08554077  -21.48562622 149.636090
#> 91    99.00511932   34.04209137 104.694210
#> 92   -30.63081169  134.37409973 137.821063
#> 93    49.51791382  125.56004333 134.971657
#> 94   136.44863892  -38.79249191 141.855872
#> 95    17.03706169 -181.66793823 182.465069
#> 96  -123.29528809 -191.51313782 227.769642
#> 97   -83.80097961  -96.09619904 127.503269
#> 98    60.46145630   13.98965931  62.058829
#> 99    45.78685760   73.58416748  86.666407
#> 100  -98.88829803   57.99286652 114.638859
#> 101 -129.87619019   18.23637199 131.150257
#> 102  -50.57397461  -11.07850075  51.773160
#> 103   89.87651062   25.90427017  93.535118
#> 104  101.40947723   -1.82649517 101.425924
#> 105   62.06584930  -40.03591156  73.858269
#> 106  -36.08080292  -24.85507011  43.813227
#> 107  -99.24241638   56.89841461 114.396183
#> 108   39.57685089  138.01231384 143.574809
#> 109   90.95823669  110.94507599 143.465016
#> 110  101.52866364   84.99912262 132.411934
#> 111  -29.25353432   35.30216599  45.847707
#> 112 -184.70722961  -73.74170685 198.883383
#> 113 -139.27862549 -161.64154053 213.369452
#> 114   40.02721405 -184.17227173 188.471758
#> 115  192.35974121  -64.55889130 202.904215
#> 116   76.82130432  133.40084839 153.939271
#> 117 -138.23652649  112.76264954 178.394934
#> 118  -96.67220306   39.11756516 104.286618
#> 119   34.50850677  -41.21197891  53.751877
#> 120   76.43877411   37.20776367  85.013551
#> 121   99.27052307   63.63646698 117.916227
#> 122   59.37224579  -40.38613892  71.806015
#> 123   -0.43293756  -40.55798721  40.560298
#> 124  -97.57418060  -72.76273346 121.717444
#> 125  -69.45354462   12.12254620  70.503553
#> 126  -42.09887314   17.62181091  45.638179
#> 127  -29.19197845  112.74947357 116.467229
#> 128    5.26084805   76.24212646  76.423415
#> 129   45.50227356  -57.70454788  73.486541
#> 130   44.40058899 -172.16139221 177.794705
#> 131  -65.57574463  -60.67929459  89.342907
#> 132  -19.84664726  115.18241882 116.879763
#> 133   43.92508316   85.86142731  96.444791
#> 134   74.70408630  -56.46582413  93.643418
#> 135   41.39479065 -235.06027222 238.677314
#> 136  -37.38652039 -179.28552246 183.142159
#> 137  -66.98860931  -58.28730774  88.796869
#> 138 -116.46329498  178.13711548 212.829817
#> 139  -24.57698250  195.07673645 196.618822
#> 140  123.56452179   95.68944550 156.283912
#> 141   33.37053680   47.25807190  57.852555
#> 142  -43.50907516   57.48181915  72.091603
#> 143  -66.41272736    8.92610073  67.009892
#> 144   65.34075165 -158.57865906 171.512696
#> 145   68.86150360 -101.67769623 122.801712
#> 146 -125.93002319   -9.71623421 126.304299
#> 147 -159.49771118   -4.96963787 159.575115
#> 148 -137.05686951  -46.63030624 144.772135
#> 149   56.97639084   -2.14760852  57.016851
#> 150  171.94882202   86.33657837 192.406866
#> 151   91.70223236   55.55863571 107.219688
#> 152   14.13138580   71.31010437  72.696816
#> 153   49.53326797   63.19907379  80.297370
#> 154  178.02647400  -32.41943359 180.954263
#> 155  106.56456757 -191.52734375 219.177395
#> 156  -38.45603561 -175.02651978 179.201421
#> 157  -85.62114716   48.70066452  98.502465
#> 158  -30.97741318  162.29530334 165.225197
#> 159   44.10549545  163.91137695 169.741669
#> 160   -3.06659222  -87.54616547  87.599858
#> 161  -32.95389175 -183.68750000 186.620086
#> 162 -183.09960938  -11.89025211 183.485272
#> 163 -195.21725464  123.40940094 230.953798
#> 164  -13.63065815  111.31678009 112.148207
#> 165  117.92003632  -25.05365562 120.552149
#> 166   98.38012695  -52.28072739 111.408814
#> 167   34.24045563   55.46933365  65.186316
#> 168   -9.13545513   74.80869293  75.364428
#> 169  -68.80738831   70.06351471  98.200574
#> 170  -51.12636566   36.59746933  62.875115
#> 171   39.83956146  -57.86931992  70.257020
#> 172  172.47396851   -4.29001188 172.527314
#> 173   86.04506683  -75.79003906 114.664221
#> 174  -53.31404114 -118.05915070 129.538990
#> 175  -75.67212677  -18.98548317  78.017430
#> 176  -91.07590485   51.89393997 104.822714
#> 177   -5.37751245   68.49532318  68.706091
#> 178   37.41598892  -42.36507416  56.522170
#> 179   36.59008408  -43.59940720  56.918736
#> 180   38.00950241  -20.94831276  43.399932
#> 181  -16.01297188   -2.66455674  16.233149
#> 182  -51.37955093   17.89604568  54.407046
#> 183  -54.39760590   10.03949165  55.316281
#> 184   44.33796692  -35.66500854  56.902093
#> 185   32.57141113  -10.87067413  34.337565
#> 186  -33.80479050   56.13455200  65.527489
#> 187  -79.86755371   21.47190285  82.703499
#> 188  -39.60745621  -22.02240181  45.318173
#> 189    8.47511387  -48.88803101  49.617206
#> 190  -90.79559326    6.31873226  91.015197
#> 191  -64.82318878   44.67007065  78.723955
#> 192    0.18408394  -32.75666428  32.757182
#> 193  105.94131470   -9.70434952 106.384851
#> 194  144.28874207   19.20134735 145.560753
#> 195   62.80364227   22.42794418  66.688156
#> 196   60.30104828   39.16162872  71.901666
#> 197    9.89735603  -66.46468353  67.197558
#> 198 -106.93024445   10.95139503 107.489582
#> 199  -71.93504333   11.54382896  72.855408
#> 200  -15.40420532   59.05524826  61.031237
#> 201   38.58776474   96.13886261 103.593902
#> 202   14.13342857  -12.32586670  18.753154
#> 203  -23.82175446  -99.67160034 102.478797
#> 204   78.48921967 -137.60769653 158.418546
#> 205   84.97846222  -41.36739731  94.512436
#> 206   92.27015686   43.18185043 101.874698
#> 207  115.59316254   -1.24245787 115.599840
#> 208  -56.35721970  -37.03569031  67.437219
#> 209 -180.55995178   62.22626495 190.981686
#> 210 -140.70202637   46.97417831 148.336218
#> 211   -7.59937429    3.77132797   8.483714
#> 212   67.00070190   -6.63315582  67.328247
#> 213  -62.06425858    0.66017485  62.067770
#> 214 -116.04092407   27.89300156 119.346201
#> 215  -59.52561188  -41.72414398  72.692521
#> 216   -0.77147400   21.10395050  21.118047
#> 217  108.75315094   80.06433868 135.046459
#> 218   70.34349823  101.36681366 123.383300
#> 219   39.77587128   17.61881065  43.503361
#> 220  -60.37347031  -58.00954819  83.726123
#> 221 -114.05629730  -99.15499878 151.130913
#> 222  -66.80593872 -134.63844299 150.301510
#> 223   43.72975540  -73.24816895  85.308767
#> 224  164.66314697  -55.15236664 173.654069
#> 225  150.55615234   66.84265137 164.727335
#> 226   91.72106934   87.64593506 126.864355
#> 227  -14.94168758   47.10339355  49.416432
#> 228  -89.88838959  -18.98774910  91.871961
#> 229  -71.78220367  -33.85676575  79.366021
#> 230  -20.46035576   -6.62715006  21.506866
#> 231    4.00623178   41.87377167  42.064981
#> 232    5.95948648  107.77281952 107.937464
#> 233  -13.48343372   55.39666748  57.013979
#> 234  -37.61662674  -57.86634827  69.018294
#> 235 -120.37464142 -120.66319275 170.439609
#> 236  -47.53048706   -1.61178923  47.557808
#> 237  -29.76260757  102.87325287 107.092105
#> 238   22.85802269  -63.96831131  67.929626
#> 239  104.03731537 -221.82759094 245.012741
#> 240  211.69795227  -63.96039581 221.149170
#> 241  213.21084595  231.52061462 314.739034
#> 242   -9.13702965  290.56985474 290.713477
#> 243 -128.34117126   72.82192993 147.561817
#> 244 -159.51152039  -47.50910950 166.436296
#> 245  -77.42742920  -67.09934998 102.456476
#> 246  -26.39572144  -86.52680969  90.463379
#> 247   90.55123901 -108.89559174 141.625481
#> 248  186.29846191 -124.10591125 223.851277
#> 249  131.00503540  -67.98168182 147.593456
#> 250  -27.94762993   59.17645264  65.444041
#> 251 -139.96080017  133.09304810 193.139289
#> 252  -72.17671204  108.89688873 130.644595
#> 253  -33.51733398  -45.92101669  56.852014
#> 254  -62.54678345 -114.81504822 130.746302
#> 255 -103.03398895  -79.31188965 130.024531
#> 256  -25.91527557  -32.54300690  41.601067
#> 257   54.49884033   78.98854065  95.965166
#> 258   11.48447990   53.09904099  54.326802
#> 259  -30.19388962   20.10084534  36.272785
#> 260  -24.40831566  -13.20516205  27.751436
#> 261   -1.74765956    9.34365654   9.505695
#> 262   60.82427597   35.01378250  70.182316
#> 263   92.64584351   -7.70908213  92.966027
#> 264  139.90768433   11.08665848 140.346265
#> 265   33.30353546   -7.99925184  34.250745
#> 266  -81.47298431   80.70437622 114.677999
#> 267  -62.68546295   75.44666290  98.090092
#> 268  -19.16712570   34.85837555  39.780461
#> 269   -0.02145432   20.13529205  20.135303
#> 270  -38.08424759  -92.94854736 100.448207
#> 271  -48.96025848  -65.03098297  81.401079
#> 272   53.56797409  -79.54528046  95.900884
#> 273  118.42950439   -9.71140385 118.827012
#> 274   74.71733856   74.34899139 105.406135
#> 275   31.12773132   25.12059784  39.999751
#> 276  -22.45550346    5.16942930  23.042843
#> 277  -69.73574829  -94.28886414 117.275166
#> 278  -97.63967133 -126.90122986 160.116918
#> 279  -47.92731476    1.65779734  47.955978
#> 280    4.83535290  104.91230011 105.023670
#> 281   30.44248009  109.34921265 113.507687
#> 282  -45.99375534   34.57789993  57.541782
#> 283  -46.39493561   23.31446266  51.923542
#> 284   -8.71826839   56.22201920  56.893968
#> 285    0.33777514  -32.21507263  32.216843
#> 286  -21.14675903 -102.87728119 105.028189
#> 287  -43.55963516  -63.99013519  77.409168
#> 288  -45.27138519   11.11567783  46.616055
#> 289   -1.83521438  -66.90228271  66.927449
#> 290   29.31253433 -107.10344696 111.042213
#> 291   99.26145935   63.06958771 117.603615
#> 292  178.50572205  136.55706787 224.749028
#> 293   28.43584061  116.51150513 119.931346
#> 294 -115.17998505  -57.83029175 128.882782
#> 295  -91.74391174  -57.06054306 108.040969
#> 296   46.95495605   48.72112274  67.664730
#> 297  103.85042572   13.63900280 104.742223
#> 298   30.90923500    0.02168762  30.909243
#> 299  -26.60593796  -38.66022491  46.930682
#> 300  -36.84747696   55.52936935  66.642685
#> 301  -38.76455688   60.80058289  72.106877
#> 302  -60.10649490  -24.88136101  65.052847
#> 303  -12.17000961  -31.84296417  34.089346
#> 304    4.16158676   -0.68011874   4.216796
#> 305   37.05499268   24.82743263  44.603519
#> 306   90.56098938  -23.60462379  93.586703
#> 307   14.64395428  -96.34392548  97.450487
#> 308  -10.68675423  -97.93535614  98.516703
#> 309    1.29207075 -118.75139618 118.758425
#> 310   84.17320251    7.73522949  84.527876
#> 311   -1.38965809  132.28540039 132.292699
#> 312  -85.64037323  141.90518188 165.744847
#> 313  -64.65158081   88.06042480 109.244978
#> 314   10.09483719   10.89061069  14.849617
#> 315   18.66069031  -51.36598969  54.650583
#> 316  -12.14181137  -84.98880005  85.851731
#> 317  -16.69356537   10.31065750  19.621029
#> 318  -36.80729675  143.85017395 148.484510
#> 319   48.40723038   91.34089661 103.375139
#> 320   66.91756439  -61.61365891  90.962648
#> 321   49.35814285  -91.70236206 104.141968
#> 322  -37.74134445  -78.84551239  87.412950
#> 323   29.31048203 -103.26760864 107.346650
#> 324   -7.24728203  -74.59731293  74.948530
#> 325 -108.70255280    0.58245629 108.704113
#> 326  -58.07542801   37.55007553  69.157527
#> 327  -18.42039490   65.56549072  68.103924
#> 328   26.23460579   38.50801086  46.595294
#> 329  -33.21652985   46.17497253  56.881156
#> 330   64.14926147  -52.80126190  83.084902
#> 331  128.01683044   -1.33169365 128.023757
#> 332  -17.82753754  146.73297119 147.811995
#> 333  -67.49650574  202.28680420 213.250391
#>  [ reached 'max' / getOption("max.print") -- omitted 667 rows ]